refactor(textbox): Move input engine into TextBoxCore
Move the editing implementation out of TextBox into the internal
TextBoxCore engine, reached through ITextBoxHost. TextBox keeps its
public dependency properties, events and methods, plus the protected
override shells that forward to the core; everything else -- selection,
caret, undo history, IME, pointers, grippers, clipboard, the selection
flyout and the proofing menu -- now lives on the engine.

Retypes the internal consumers the move exposes: the TextBox-typed
notifications singleton, the IME extension, and the out-of-assembly
consumers in the Skia platform runtimes.

Audited for the two silent-failure modes this shape of refactor invites:
partial-method declarations left behind without their implementation
(calls are silently elided) and protected overrides converted to plain
internal methods without a forwarding override (the handler stops being
called). Both compile clean while dropping behaviour.

No behaviour change -- PasswordBox still derives from TextBox here.
