+++ src/Uno.UI.Runtime.Skia.Win32/UI/Xaml/Controls/TextBox/Win32ImeTextBoxExtension.cs
@@ -34,9 +34,9 @@ internal sealed class Win32ImeTextBoxExtension : IImeTextBoxExtension
 	public event EventHandler<ImeCompositionEventArgs>? CompositionCompleted;
 	public event EventHandler? CompositionEnded;
 
-	public void StartImeSession(TextBox textBox)
+	public void StartImeSession(TextBoxCore core)
 	{
-		var wrapper = (Win32WindowWrapper)XamlRootMap.GetHostForRoot(textBox.XamlRoot!)!;
+		var wrapper = (Win32WindowWrapper)XamlRootMap.GetHostForRoot(core.Owner.XamlRoot!)!;
 		_hwnd = (HWND)((Win32NativeWindow)wrapper.NativeWindow!).Hwnd;
 	}
