+++ src/Uno.UI.RuntimeTests/Tests/Windows_UI_Xaml_Controls/Given_TextBox.skia.cs
@@ -4821,21 +4821,21 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			finger.Press(SUT.GetAbsoluteBoundsRect().GetCenter());
 			finger.Release();
 			await WindowHelper.WaitForIdle();
-			Assert.AreEqual(TextBox.CaretDisplayMode.CaretWithThumbsBothEndsShowing, SUT.CaretMode);
+			Assert.AreEqual(TextBoxCore.CaretDisplayMode.CaretWithThumbsBothEndsShowing, SUT.CaretMode);
 			Assert.AreEqual("Text", SUT.SelectedText);
 
 			// clicking inside the selected area keeps the selection
 			finger.Press(SUT.GetAbsoluteBoundsRect().GetCenter());
 			finger.Release();
 			await WindowHelper.WaitForIdle();
-			Assert.AreEqual(TextBox.CaretDisplayMode.CaretWithThumbsBothEndsShowing, SUT.CaretMode);
+			Assert.AreEqual(TextBoxCore.CaretDisplayMode.CaretWithThumbsBothEndsShowing, SUT.CaretMode);
 			Assert.AreEqual("Text", SUT.SelectedText);
 
 			// clicking outside the selected area drops it
 			finger.Press(SUT.GetAbsoluteBoundsRect().GetCenter() + new Point(100, 0));
 			finger.Release();
 			await WindowHelper.WaitForIdle();
-			Assert.AreEqual(TextBox.CaretDisplayMode.CaretWithThumbsOnlyEndShowing, SUT.CaretMode);
+			Assert.AreEqual(TextBoxCore.CaretDisplayMode.CaretWithThumbsOnlyEndShowing, SUT.CaretMode);
 			Assert.AreEqual("", SUT.SelectedText);
 		}
 
@@ -4859,7 +4859,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			finger.Press(new Point(bounds.Left + 15, bounds.GetCenter().Y));
 			finger.Release();
 			await WindowHelper.WaitFor(
-				() => SUT.CaretMode == TextBox.CaretDisplayMode.CaretWithThumbsBothEndsShowing,
+				() => SUT.CaretMode == TextBoxCore.CaretDisplayMode.CaretWithThumbsBothEndsShowing,
 				message: "first tap should select the word and show both thumbs");
 			var lengthBeforeDrag = SUT.SelectionLength;
 
@@ -4891,24 +4891,24 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			// The drag readjusted (extended) the selection...
 			Assert.IsTrue(SUT.SelectionLength > lengthBeforeDrag, $"drag should extend the selection (was {lengthBeforeDrag}, now {SUT.SelectionLength})");
 			// ...and both thumbs must remain visible (the readjust must not be mistaken for a long-press that opens the context menu and hides the thumbs).
-			Assert.AreEqual(TextBox.CaretDisplayMode.CaretWithThumbsBothEndsShowing, SUT.CaretMode);
+			Assert.AreEqual(TextBoxCore.CaretDisplayMode.CaretWithThumbsBothEndsShowing, SUT.CaretMode);
 		}
 
 		[TestMethod]
 		[PlatformCondition(ConditionMode.Include, RuntimeTestPlatforms.SkiaDesktop | RuntimeTestPlatforms.SkiaAndroid)] // Android convention: run on Desktop (dev) + real Android only
 		public Task When_Touch_Gripper_Drag_Readjusts_Selection_Keeps_Thumbs_Android()
-			=> AssertGripperDragReadjustsWordSelection(TextBox.TouchTextSelectionConvention.Android);
+			=> AssertGripperDragReadjustsWordSelection(TextBoxCore.TouchTextSelectionConvention.Android);
 
 		[TestMethod]
 		[PlatformCondition(ConditionMode.Include, RuntimeTestPlatforms.SkiaDesktop | RuntimeTestPlatforms.SkiaUIKit)] // iOS convention: run on Desktop (dev) + real iOS only
 		public Task When_Touch_Gripper_Drag_Readjusts_Selection_Keeps_Thumbs_iOS()
-			=> AssertGripperDragReadjustsWordSelection(TextBox.TouchTextSelectionConvention.iOS);
+			=> AssertGripperDragReadjustsWordSelection(TextBoxCore.TouchTextSelectionConvention.iOS);
 
 		// Native iOS/Android sister of When_Touch_Gripper_Drag_Readjusts_Selection_Keeps_Thumbs: on mobile a
 		// single tap only places a caret, so the word selection comes from a double-tap. Dragging the end
 		// gripper then readjusts (extends) the selection, and spanning past the 800ms hold threshold must NOT
 		// be mistaken for a long-press (which word-selects on Android / caret-drags on iOS) and must keep both thumbs.
-		private static async Task AssertGripperDragReadjustsWordSelection(TextBox.TouchTextSelectionConvention convention)
+		private static async Task AssertGripperDragReadjustsWordSelection(TextBoxCore.TouchTextSelectionConvention convention)
 		{
 			var SUT = new TextBox
 			{
@@ -4930,7 +4930,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			finger.Press(wordPoint);
 			finger.Release();
 			await WindowHelper.WaitFor(
-				() => SUT.CaretMode == TextBox.CaretDisplayMode.CaretWithThumbsBothEndsShowing,
+				() => SUT.CaretMode == TextBoxCore.CaretDisplayMode.CaretWithThumbsBothEndsShowing,
 				message: "double-tap should select the word and show both thumbs");
 			var lengthBeforeDrag = SUT.SelectionLength;
 
@@ -4962,7 +4962,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			// The drag readjusted (extended) the selection...
 			Assert.IsTrue(SUT.SelectionLength > lengthBeforeDrag, $"drag should extend the selection (was {lengthBeforeDrag}, now {SUT.SelectionLength})");
 			// ...and both thumbs must remain visible (the readjust must not be mistaken for a long-press that changes the selection and hides thumbs).
-			Assert.AreEqual(TextBox.CaretDisplayMode.CaretWithThumbsBothEndsShowing, SUT.CaretMode);
+			Assert.AreEqual(TextBoxCore.CaretDisplayMode.CaretWithThumbsBothEndsShowing, SUT.CaretMode);
 		}
 
 		[TestMethod]
@@ -4972,7 +4972,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			{
 				Width = 300,
 				Text = "The quick brown fox jumps over",
-				TouchSelectionConvention = TextBox.TouchTextSelectionConvention.Android
+				TouchSelectionConvention = TextBoxCore.TouchTextSelectionConvention.Android
 			};
 
 			await UITestHelper.Load(SUT);
@@ -4985,7 +4985,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			finger.Press(new Point(bounds.Left + 90, bounds.GetCenter().Y));
 			finger.Release();
 			await WindowHelper.WaitFor(
-				() => SUT.CaretMode == TextBox.CaretDisplayMode.CaretWithThumbsOnlyEndShowing,
+				() => SUT.CaretMode == TextBoxCore.CaretDisplayMode.CaretWithThumbsOnlyEndShowing,
 				message: "tap should place the single insertion handle");
 
 			await WindowHelper.WaitFor(
@@ -5023,7 +5023,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			{
 				Width = 300,
 				Text = "The quick brown fox jumps over",
-				TouchSelectionConvention = TextBox.TouchTextSelectionConvention.Android
+				TouchSelectionConvention = TextBoxCore.TouchTextSelectionConvention.Android
 			};
 
 			await UITestHelper.Load(SUT);
@@ -5036,7 +5036,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			finger.Press(new Point(bounds.GetCenter().X, bounds.GetCenter().Y));
 			finger.Release();
 			await WindowHelper.WaitFor(
-				() => SUT.CaretMode == TextBox.CaretDisplayMode.CaretWithThumbsOnlyEndShowing,
+				() => SUT.CaretMode == TextBoxCore.CaretDisplayMode.CaretWithThumbsOnlyEndShowing,
 				message: "tap should place the single insertion handle");
 			await WindowHelper.WaitFor(
 				() => SUT.VisibleGrippersForTesting is { } vg && vg.end.GetAbsoluteBoundsRect().Width > 0,
@@ -5070,30 +5070,30 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 
 		[TestMethod]
 		public Task When_Touch_SingleTap_Places_Caret_Android()
-			=> AssertTouchSingleTapPlacesCaret(TextBox.TouchTextSelectionConvention.Android, TextBox.CaretDisplayMode.CaretWithThumbsOnlyEndShowing);
+			=> AssertTouchSingleTapPlacesCaret(TextBoxCore.TouchTextSelectionConvention.Android, TextBoxCore.CaretDisplayMode.CaretWithThumbsOnlyEndShowing);
 
 		[TestMethod]
 		public Task When_Touch_SingleTap_Places_Caret_iOS()
-			=> AssertTouchSingleTapPlacesCaret(TextBox.TouchTextSelectionConvention.iOS, TextBox.CaretDisplayMode.ThumblessCaretShowing);
+			=> AssertTouchSingleTapPlacesCaret(TextBoxCore.TouchTextSelectionConvention.iOS, TextBoxCore.CaretDisplayMode.ThumblessCaretShowing);
 
 		[TestMethod]
 		public Task When_Touch_DoubleTap_Selects_Word_Android()
-			=> AssertTouchDoubleTapSelectsWord(TextBox.TouchTextSelectionConvention.Android);
+			=> AssertTouchDoubleTapSelectsWord(TextBoxCore.TouchTextSelectionConvention.Android);
 
 		[TestMethod]
 		public Task When_Touch_DoubleTap_Selects_Word_iOS()
-			=> AssertTouchDoubleTapSelectsWord(TextBox.TouchTextSelectionConvention.iOS);
+			=> AssertTouchDoubleTapSelectsWord(TextBoxCore.TouchTextSelectionConvention.iOS);
 
 		[TestMethod]
 		public Task When_Touch_Tap_Collapses_Selection_Android()
-			=> AssertTouchTapCollapsesSelection(TextBox.TouchTextSelectionConvention.Android);
+			=> AssertTouchTapCollapsesSelection(TextBoxCore.TouchTextSelectionConvention.Android);
 
 		[TestMethod]
 		public Task When_Touch_Tap_Collapses_Selection_iOS()
-			=> AssertTouchTapCollapsesSelection(TextBox.TouchTextSelectionConvention.iOS);
+			=> AssertTouchTapCollapsesSelection(TextBoxCore.TouchTextSelectionConvention.iOS);
 
 		// Native iOS/Android: a single tap places a caret (it does NOT select a word like the Windows convention).
-		private static async Task AssertTouchSingleTapPlacesCaret(TextBox.TouchTextSelectionConvention convention, TextBox.CaretDisplayMode expectedCaret)
+		private static async Task AssertTouchSingleTapPlacesCaret(TextBoxCore.TouchTextSelectionConvention convention, TextBoxCore.CaretDisplayMode expectedCaret)
 		{
 			var SUT = new TextBox
 			{
@@ -5121,7 +5121,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 		}
 
 		// Native iOS/Android: a double-tap selects the word under the tap.
-		private static async Task AssertTouchDoubleTapSelectsWord(TextBox.TouchTextSelectionConvention convention)
+		private static async Task AssertTouchDoubleTapSelectsWord(TextBoxCore.TouchTextSelectionConvention convention)
 		{
 			var SUT = new TextBox
 			{
@@ -5144,7 +5144,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			await WindowHelper.WaitForIdle();
 
 			Assert.AreEqual("Text", SUT.SelectedText);
-			Assert.AreEqual(TextBox.CaretDisplayMode.CaretWithThumbsBothEndsShowing, SUT.CaretMode);
+			Assert.AreEqual(TextBoxCore.CaretDisplayMode.CaretWithThumbsBothEndsShowing, SUT.CaretMode);
 		}
 
 		[TestMethod]
@@ -5154,7 +5154,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			{
 				Width = 400,
 				Text = "Some Text",
-				TouchSelectionConvention = TextBox.TouchTextSelectionConvention.Android
+				TouchSelectionConvention = TextBoxCore.TouchTextSelectionConvention.Android
 			};
 
 			await UITestHelper.Load(SUT);
@@ -5170,7 +5170,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			finger.Press(point);
 			finger.Release();
 			await WindowHelper.WaitForIdle(); // let the insertion handle render + position (a real double-tap has this gap)
-			Assert.AreEqual(TextBox.CaretDisplayMode.CaretWithThumbsOnlyEndShowing, SUT.CaretMode);
+			Assert.AreEqual(TextBoxCore.CaretDisplayMode.CaretWithThumbsOnlyEndShowing, SUT.CaretMode);
 			Assert.IsNotNull(SUT.VisibleGrippersForTesting, "insertion handle should be visible after the first Android tap");
 
 			// Second tap a few px away: still lands on the insertion handle (within its hit-rect) yet within the
@@ -5180,7 +5180,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			await WindowHelper.WaitForIdle();
 
 			Assert.AreEqual("Some", SUT.SelectedText); // native Android selects just the word, not the trailing space
-			Assert.AreEqual(TextBox.CaretDisplayMode.CaretWithThumbsBothEndsShowing, SUT.CaretMode);
+			Assert.AreEqual(TextBoxCore.CaretDisplayMode.CaretWithThumbsBothEndsShowing, SUT.CaretMode);
 		}
 
 		// Regression: tapping the single Android insertion handle re-sampled the finger point, which sits on the
@@ -5194,7 +5194,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			{
 				Width = 300,
 				Text = "The quick brown fox jumps over",
-				TouchSelectionConvention = TextBox.TouchTextSelectionConvention.Android
+				TouchSelectionConvention = TextBoxCore.TouchTextSelectionConvention.Android
 			};
 
 			await UITestHelper.Load(SUT);
@@ -5207,7 +5207,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			finger.Press(new Point(bounds.Left + 90, bounds.GetCenter().Y));
 			finger.Release();
 			await WindowHelper.WaitFor(
-				() => SUT.CaretMode == TextBox.CaretDisplayMode.CaretWithThumbsOnlyEndShowing,
+				() => SUT.CaretMode == TextBoxCore.CaretDisplayMode.CaretWithThumbsOnlyEndShowing,
 				message: "tap should place the single insertion handle");
 
 			await WindowHelper.WaitFor(
@@ -5266,7 +5266,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			{
 				Width = 400,
 				Text = "Some Text",
-				TouchSelectionConvention = TextBox.TouchTextSelectionConvention.Android
+				TouchSelectionConvention = TextBoxCore.TouchTextSelectionConvention.Android
 			};
 
 			await UITestHelper.Load(SUT);
@@ -5279,7 +5279,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			finger.Press(new Point(bounds.Left + 20, bounds.GetCenter().Y));
 			finger.Release();
 			await WindowHelper.WaitFor(
-				() => SUT.CaretMode == TextBox.CaretDisplayMode.CaretWithThumbsOnlyEndShowing,
+				() => SUT.CaretMode == TextBoxCore.CaretDisplayMode.CaretWithThumbsOnlyEndShowing,
 				message: "the first Android tap should place the single insertion handle");
 			Assert.IsNotNull(SUT.VisibleGrippersForTesting, "the insertion handle should be visible after the first Android tap");
 
@@ -5435,18 +5435,18 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 		[TestMethod]
 		[PlatformCondition(ConditionMode.Include, RuntimeTestPlatforms.SkiaDesktop | RuntimeTestPlatforms.SkiaAndroid)] // mobile conventions: run on Desktop (dev) + real Android only
 		public Task When_Touch_Tap_Empty_Places_Caret_Android()
-			=> AssertTouchTapOnEmptyBoxPlacesCaret(TextBox.TouchTextSelectionConvention.Android, TextBox.CaretDisplayMode.CaretWithThumbsOnlyEndShowing);
+			=> AssertTouchTapOnEmptyBoxPlacesCaret(TextBoxCore.TouchTextSelectionConvention.Android, TextBoxCore.CaretDisplayMode.CaretWithThumbsOnlyEndShowing);
 
 		[TestMethod]
 		[PlatformCondition(ConditionMode.Include, RuntimeTestPlatforms.SkiaDesktop | RuntimeTestPlatforms.SkiaAndroid)]
 		public Task When_Touch_Tap_Empty_Places_Caret_iOS()
-			=> AssertTouchTapOnEmptyBoxPlacesCaret(TextBox.TouchTextSelectionConvention.iOS, TextBox.CaretDisplayMode.ThumblessCaretShowing);
+			=> AssertTouchTapOnEmptyBoxPlacesCaret(TextBoxCore.TouchTextSelectionConvention.iOS, TextBoxCore.CaretDisplayMode.ThumblessCaretShowing);
 
 		// Native iOS/Android: a single tap in an EMPTY field places the caret - Android with its insertion handle, iOS
 		// as a bare caret. The empty box used to swallow the tap entirely (the tap-to-caret path was gated on non-empty
 		// text), leaving no caret affordance and no handle to open the flyout from. A single tap must NOT pop the
 		// flyout though: that belongs to the double-tap / long-press / handle-tap.
-		private static async Task AssertTouchTapOnEmptyBoxPlacesCaret(TextBox.TouchTextSelectionConvention convention, TextBox.CaretDisplayMode expectedCaret)
+		private static async Task AssertTouchTapOnEmptyBoxPlacesCaret(TextBoxCore.TouchTextSelectionConvention convention, TextBoxCore.CaretDisplayMode expectedCaret)
 		{
 			using var _ = new TextBoxFeatureConfigDisposable();
 			using var __ = new DisposableAction(() =>
@@ -5471,7 +5471,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 
 			Assert.AreEqual(0, SUT.SelectionStart, "the caret should sit at the only available index");
 			Assert.AreEqual(0, SUT.SelectionLength, "there is nothing to select in an empty box");
-			if (convention == TextBox.TouchTextSelectionConvention.Android)
+			if (convention == TextBoxCore.TouchTextSelectionConvention.Android)
 			{
 				Assert.IsNotNull(SUT.VisibleGrippersForTesting, "Android should show the insertion handle so it can open the flyout");
 			}
@@ -5485,27 +5485,27 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 		[TestMethod]
 		[PlatformCondition(ConditionMode.Include, RuntimeTestPlatforms.SkiaDesktop | RuntimeTestPlatforms.SkiaAndroid)] // mobile conventions: run on Desktop (dev) + real Android only
 		public Task When_Touch_DoubleTap_Empty_Opens_Flyout_Android()
-			=> AssertTouchGestureOnEmptyBoxOpensFlyout(TextBox.TouchTextSelectionConvention.Android, longPress: false);
+			=> AssertTouchGestureOnEmptyBoxOpensFlyout(TextBoxCore.TouchTextSelectionConvention.Android, longPress: false);
 
 		[TestMethod]
 		[PlatformCondition(ConditionMode.Include, RuntimeTestPlatforms.SkiaDesktop | RuntimeTestPlatforms.SkiaAndroid)]
 		public Task When_Touch_DoubleTap_Empty_Opens_Flyout_iOS()
-			=> AssertTouchGestureOnEmptyBoxOpensFlyout(TextBox.TouchTextSelectionConvention.iOS, longPress: false);
+			=> AssertTouchGestureOnEmptyBoxOpensFlyout(TextBoxCore.TouchTextSelectionConvention.iOS, longPress: false);
 
 		[TestMethod]
 		[PlatformCondition(ConditionMode.Include, RuntimeTestPlatforms.SkiaDesktop | RuntimeTestPlatforms.SkiaAndroid)]
 		public Task When_Touch_LongPress_Empty_Opens_Flyout_Android()
-			=> AssertTouchGestureOnEmptyBoxOpensFlyout(TextBox.TouchTextSelectionConvention.Android, longPress: true);
+			=> AssertTouchGestureOnEmptyBoxOpensFlyout(TextBoxCore.TouchTextSelectionConvention.Android, longPress: true);
 
 		[TestMethod]
 		[PlatformCondition(ConditionMode.Include, RuntimeTestPlatforms.SkiaDesktop | RuntimeTestPlatforms.SkiaAndroid)]
 		public Task When_Touch_LongPress_Empty_Opens_Flyout_iOS()
-			=> AssertTouchGestureOnEmptyBoxOpensFlyout(TextBox.TouchTextSelectionConvention.iOS, longPress: true);
+			=> AssertTouchGestureOnEmptyBoxOpensFlyout(TextBoxCore.TouchTextSelectionConvention.iOS, longPress: true);
 
 		// Native iOS/Android pop the text flyout (Paste) over an EMPTY field on a double-tap or a long-press. Neither
 		// gesture has a word to select nor a caret to drag there, so the mobile conventions used to swallow it and show
 		// nothing at all. The clipboard is populated so Paste exists: with no command the flyout self-hides (correctly).
-		private async Task AssertTouchGestureOnEmptyBoxOpensFlyout(TextBox.TouchTextSelectionConvention convention, bool longPress)
+		private async Task AssertTouchGestureOnEmptyBoxOpensFlyout(TextBoxCore.TouchTextSelectionConvention convention, bool longPress)
 		{
 			if (!Uno.Foundation.Extensibility.ApiExtensibility.IsRegistered<Uno.ApplicationModel.DataTransfer.IClipboardExtension>())
 			{
@@ -5564,9 +5564,9 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 
 			Assert.AreEqual("", SUT.SelectedText, "there is nothing to select in an empty box");
 			// The thumbless iOS caret blinks on a 500 ms timer, so it can be mid-blink by the time the flyout opens.
-			var expectedCaret = convention == TextBox.TouchTextSelectionConvention.Android
-				? TextBox.CaretDisplayMode.CaretWithThumbsOnlyEndShowing
-				: TextBox.CaretDisplayMode.ThumblessCaretShowing;
+			var expectedCaret = convention == TextBoxCore.TouchTextSelectionConvention.Android
+				? TextBoxCore.CaretDisplayMode.CaretWithThumbsOnlyEndShowing
+				: TextBoxCore.CaretDisplayMode.ThumblessCaretShowing;
 			await WindowHelper.WaitFor(
 				() => SUT.CaretMode == expectedCaret,
 				timeoutMS: 3000,
@@ -5604,7 +5604,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			{
 				Width = 400,
 				Text = "",
-				TouchSelectionConvention = TextBox.TouchTextSelectionConvention.Android
+				TouchSelectionConvention = TextBoxCore.TouchTextSelectionConvention.Android
 			};
 
 			await UITestHelper.Load(SUT);
@@ -5628,7 +5628,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 				message: "Select All alone should keep the empty-box flyout open with an empty clipboard");
 			await WindowHelper.WaitForIdle();
 
-			Assert.AreEqual(TextBox.CaretDisplayMode.CaretWithThumbsOnlyEndShowing, SUT.CaretMode, "the gesture should still place the Android insertion caret");
+			Assert.AreEqual(TextBoxCore.CaretDisplayMode.CaretWithThumbsOnlyEndShowing, SUT.CaretMode, "the gesture should still place the Android insertion caret");
 
 			if (SUT.SelectionFlyout is not TextCommandBarFlyout flyout)
 			{
@@ -5664,7 +5664,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			var SUT = new PasswordBox
 			{
 				Width = 400,
-				TouchSelectionConvention = TextBox.TouchTextSelectionConvention.Android
+				TouchSelectionConvention = TextBoxCore.TouchTextSelectionConvention.Android
 			};
 
 			await UITestHelper.Load(SUT);
@@ -5719,7 +5719,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			var SUT = new PasswordBox
 			{
 				Width = 400,
-				TouchSelectionConvention = TextBox.TouchTextSelectionConvention.Android
+				TouchSelectionConvention = TextBoxCore.TouchTextSelectionConvention.Android
 			};
 
 			await UITestHelper.Load(SUT);
@@ -5762,17 +5762,17 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 		[TestMethod]
 		[PlatformCondition(ConditionMode.Include, RuntimeTestPlatforms.SkiaDesktop | RuntimeTestPlatforms.SkiaAndroid)] // mobile conventions: run on Desktop (dev) + real Android only
 		public Task When_Touch_Tap_Selection_Thumb_Keeps_Selection_Android()
-			=> AssertTouchTapSelectionThumbKeepsSelection(TextBox.TouchTextSelectionConvention.Android, tapStartThumb: false);
+			=> AssertTouchTapSelectionThumbKeepsSelection(TextBoxCore.TouchTextSelectionConvention.Android, tapStartThumb: false);
 
 		[TestMethod]
 		[PlatformCondition(ConditionMode.Include, RuntimeTestPlatforms.SkiaDesktop | RuntimeTestPlatforms.SkiaAndroid)]
 		public Task When_Touch_Tap_Start_Selection_Thumb_Keeps_Selection_Android()
-			=> AssertTouchTapSelectionThumbKeepsSelection(TextBox.TouchTextSelectionConvention.Android, tapStartThumb: true);
+			=> AssertTouchTapSelectionThumbKeepsSelection(TextBoxCore.TouchTextSelectionConvention.Android, tapStartThumb: true);
 
 		[TestMethod]
 		[PlatformCondition(ConditionMode.Include, RuntimeTestPlatforms.SkiaDesktop | RuntimeTestPlatforms.SkiaAndroid)]
 		public Task When_Touch_Tap_Selection_Thumb_Keeps_Selection_iOS()
-			=> AssertTouchTapSelectionThumbKeepsSelection(TextBox.TouchTextSelectionConvention.iOS, tapStartThumb: false);
+			=> AssertTouchTapSelectionThumbKeepsSelection(TextBoxCore.TouchTextSelectionConvention.iOS, tapStartThumb: false);
 
 		// Native iOS/Android: with a range selected (both thumbs showing), tapping either thumb must KEEP the
 		// selection - the thumb is a selection edge, not a caret. It used to route through the caret-placing tap path
@@ -5780,7 +5780,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 		// The tap is delivered through the gripper host seam instead of by injecting at the thumb's coordinates:
 		// grippers live in popups clipped to the TextBox, so in the (short) test host the thumb hangs outside the
 		// control and a coordinate-aimed tap silently lands on the text - or on nothing - rather than the gripper.
-		private static async Task AssertTouchTapSelectionThumbKeepsSelection(TextBox.TouchTextSelectionConvention convention, bool tapStartThumb)
+		private static async Task AssertTouchTapSelectionThumbKeepsSelection(TextBoxCore.TouchTextSelectionConvention convention, bool tapStartThumb)
 		{
 			using var _ = new TextBoxFeatureConfigDisposable();
 			using var __ = new DisposableAction(() =>
@@ -5806,7 +5806,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			finger.Press(wordPoint);
 			finger.Release();
 			await WindowHelper.WaitFor(
-				() => SUT.SelectedText == "Some" && SUT.CaretMode == TextBox.CaretDisplayMode.CaretWithThumbsBothEndsShowing,
+				() => SUT.SelectedText == "Some" && SUT.CaretMode == TextBoxCore.CaretDisplayMode.CaretWithThumbsBothEndsShowing,
 				message: "the double-tap should select the word and show both thumbs");
 
 			Assert.IsNotNull(SUT.VisibleGrippersForTesting, "both selection thumbs should be showing before tapping one");
@@ -5828,11 +5828,11 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			// The presenter pins a gripper tap to the selection edge the thumb points at, never to the finger's
 			// position on the thumb (see TextSelectionGripperPresenter.OnGripperPointerReleased).
 			var anchorIndex = tapStartThumb ? SUT.SelectionStart : SUT.SelectionStart + SUT.SelectionLength;
-			((ITextSelectionGripperHost)SUT).OnGripperTapped(press, anchorIndex);
+			((ITextSelectionGripperHost)SUT.Core).OnGripperTapped(press, anchorIndex);
 			await WindowHelper.WaitForIdle();
 
 			Assert.AreEqual("Some", SUT.SelectedText, $"tapping the {(tapStartThumb ? "start" : "end")} thumb (anchor {anchorIndex}) must keep the selection, not collapse it");
-			Assert.AreEqual(TextBox.CaretDisplayMode.CaretWithThumbsBothEndsShowing, SUT.CaretMode, "both thumbs must remain after tapping one of them");
+			Assert.AreEqual(TextBoxCore.CaretDisplayMode.CaretWithThumbsBothEndsShowing, SUT.CaretMode, "both thumbs must remain after tapping one of them");
 		}
 
 		// Repro: with a full selection's touch flyout open (both thumbs showing), Select All sits in the OVERFLOW
@@ -5862,7 +5862,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			{
 				Width = 400,
 				Text = "asd qwertyuiopasdfghjkl",
-				TouchSelectionConvention = TextBox.TouchTextSelectionConvention.Android
+				TouchSelectionConvention = TextBoxCore.TouchTextSelectionConvention.Android
 			};
 
 			await UITestHelper.Load(SUT);
@@ -5882,7 +5882,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			finger.Press(wordPoint);
 			finger.Release();
 			await WindowHelper.WaitFor(
-				() => SUT.SelectedText == "asd" && SUT.CaretMode == TextBox.CaretDisplayMode.CaretWithThumbsBothEndsShowing,
+				() => SUT.SelectedText == "asd" && SUT.CaretMode == TextBoxCore.CaretDisplayMode.CaretWithThumbsBothEndsShowing,
 				message: "the double-tap should select all the text and show both thumbs");
 			await WindowHelper.WaitFor(
 				() => (SUT.SelectionFlyout as TextCommandBarFlyout)?.IsOpen == true,
@@ -5930,7 +5930,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			finger.Press(farPoint);
 			finger.Release();
 			await WindowHelper.WaitFor(
-				() => SUT.SelectedText == "" && SUT.CaretMode == TextBox.CaretDisplayMode.CaretWithThumbsOnlyEndShowing,
+				() => SUT.SelectedText == "" && SUT.CaretMode == TextBoxCore.CaretDisplayMode.CaretWithThumbsOnlyEndShowing,
 				message: "a plain tap should collapse the selection to the single insertion handle");
 
 			// The gripper popups are (re)positioned on a later frame, so GetAbsoluteBoundsRect is stale right after the
@@ -6011,7 +6011,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 				Width = 400,
 				Text = "helllo", // a single misspelled word, so a touch double-tap selects it whole
 				IsSpellCheckEnabled = true,
-				TouchSelectionConvention = TextBox.TouchTextSelectionConvention.Android
+				TouchSelectionConvention = TextBoxCore.TouchTextSelectionConvention.Android
 			};
 
 			await UITestHelper.Load(SUT);
@@ -6083,7 +6083,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 		}
 
 		// Native iOS/Android: tapping collapses an existing selection to a caret (Windows keeps it).
-		private static async Task AssertTouchTapCollapsesSelection(TextBox.TouchTextSelectionConvention convention)
+		private static async Task AssertTouchTapCollapsesSelection(TextBoxCore.TouchTextSelectionConvention convention)
 		{
 			var SUT = new TextBox
 			{
@@ -6112,11 +6112,11 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 
 		[TestMethod]
 		public Task When_Touch_LongPress_Selects_Word_Android()
-			=> AssertTouchLongPress(TextBox.TouchTextSelectionConvention.Android, expectWordSelected: true);
+			=> AssertTouchLongPress(TextBoxCore.TouchTextSelectionConvention.Android, expectWordSelected: true);
 
 		[TestMethod]
 		public Task When_Touch_LongPress_Keeps_ContextMenu_Desktop()
-			=> AssertTouchLongPress(TextBox.TouchTextSelectionConvention.Desktop, expectWordSelected: false);
+			=> AssertTouchLongPress(TextBoxCore.TouchTextSelectionConvention.Desktop, expectWordSelected: false);
 
 		// A touch long-press on a mobile convention must select the word BEFORE the text control's flyout
 		// computes its commands. Regression: the inner DisplayBlock's ContextRequested class handler used to
@@ -6132,7 +6132,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			{
 				Width = 400,
 				Text = "Some Text",
-				TouchSelectionConvention = TextBox.TouchTextSelectionConvention.Android
+				TouchSelectionConvention = TextBoxCore.TouchTextSelectionConvention.Android
 			};
 
 			await UITestHelper.Load(SUT);
@@ -6181,7 +6181,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			{
 				Width = 400,
 				Text = "Some Text",
-				TouchSelectionConvention = TextBox.TouchTextSelectionConvention.Android
+				TouchSelectionConvention = TextBoxCore.TouchTextSelectionConvention.Android
 			};
 
 			await UITestHelper.Load(SUT);
@@ -6244,7 +6244,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			{
 				Width = 400,
 				Text = "Some Text",
-				TouchSelectionConvention = TextBox.TouchTextSelectionConvention.Android
+				TouchSelectionConvention = TextBoxCore.TouchTextSelectionConvention.Android
 			};
 
 			await UITestHelper.Load(SUT);
@@ -6268,7 +6268,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			// Steady state (Assert, not WaitFor): after the flyout opens, the word must stay selected AND both
 			// thumbs must remain — a bug that flips CaretMode back to thumbless leaves the highlight but drops the thumbs.
 			Assert.AreEqual("Text", SUT.SelectedText, "selection should persist while the flyout is open");
-			Assert.AreEqual(TextBox.CaretDisplayMode.CaretWithThumbsBothEndsShowing, SUT.CaretMode, "both thumbs must survive the flyout opening");
+			Assert.AreEqual(TextBoxCore.CaretDisplayMode.CaretWithThumbsBothEndsShowing, SUT.CaretMode, "both thumbs must survive the flyout opening");
 			Assert.IsTrue(
 				SUT.VisibleGrippersForTesting is { } vg
 					&& vg.start.GetAbsoluteBoundsRect().Width > 0
@@ -6379,7 +6379,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 
 		// Native Android: a touch long-press selects the word (and suppresses the context menu).
 		// The Desktop convention keeps the default context-menu behavior (no auto word selection).
-		private static async Task AssertTouchLongPress(TextBox.TouchTextSelectionConvention convention, bool expectWordSelected)
+		private static async Task AssertTouchLongPress(TextBoxCore.TouchTextSelectionConvention convention, bool expectWordSelected)
 		{
 			var SUT = new TextBox
 			{
@@ -6411,17 +6411,17 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 
 		[TestMethod]
 		public Task When_Touch_Hold_Does_Not_Flag_ContextMenu_Android()
-			=> AssertTouchHoldDoesNotFlagContextMenuOnHolding(TextBox.TouchTextSelectionConvention.Android);
+			=> AssertTouchHoldDoesNotFlagContextMenuOnHolding(TextBoxCore.TouchTextSelectionConvention.Android);
 
 		[TestMethod]
 		public Task When_Touch_Hold_Does_Not_Flag_ContextMenu_iOS()
-			=> AssertTouchHoldDoesNotFlagContextMenuOnHolding(TextBox.TouchTextSelectionConvention.iOS);
+			=> AssertTouchHoldDoesNotFlagContextMenuOnHolding(TextBoxCore.TouchTextSelectionConvention.iOS);
 
 		// Native iOS/Android handle a touch-and-hold without opening a context menu. Routing it through the
 		// ContextMenuProcessor (the path the Holding gesture uses) must NOT flag the hold as menu-showing;
 		// otherwise a later HoldingState.Canceled (the finger moving during the caret-drag / after word-select)
 		// would spuriously raise ContextCanceled or close a light-dismiss popup the TextBox lives in.
-		private static async Task AssertTouchHoldDoesNotFlagContextMenuOnHolding(TextBox.TouchTextSelectionConvention convention)
+		private static async Task AssertTouchHoldDoesNotFlagContextMenuOnHolding(TextBoxCore.TouchTextSelectionConvention convention)
 		{
 			var SUT = new TextBox
 			{
@@ -6452,7 +6452,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			{
 				Width = 400,
 				Text = "Some Text long enough",
-				TouchSelectionConvention = TextBox.TouchTextSelectionConvention.iOS
+				TouchSelectionConvention = TextBoxCore.TouchTextSelectionConvention.iOS
 			};
 
 			await UITestHelper.Load(SUT);
@@ -7414,7 +7414,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			public event EventHandler<ImeCompositionEventArgs> CompositionCompleted;
 			public event EventHandler CompositionEnded;
 
-			public void StartImeSession(TextBox textBox) { }
+			public void StartImeSession(TextBoxCore core) { }
 
 			public void EndImeSession()
 			{
