+++ src/Uno.UI/UI/Xaml/Controls/TextBox/TextBox.cs
@@ -1,4 +1,4 @@
-#if IS_UNIT_TESTS || UNO_REFERENCE_API
+﻿#if IS_UNIT_TESTS || UNO_REFERENCE_API
 #pragma warning disable CS0067, CS649
 #endif
 
@@ -32,25 +32,6 @@ using Uno.UI.Xaml.Controls;
 using System.Linq;
 #if __SKIA__
 using Microsoft.UI.Xaml.Internal;
-using System.Collections.Generic;
-using System.Windows.Input;
-using Windows.UI;
-using Microsoft.UI.Composition;
-using Microsoft.UI.Xaml.Controls.Primitives;
-using Microsoft.UI.Xaml.Documents;
-using Microsoft.UI.Xaml.Shapes;
-using SkiaSharp;
-using Uno.Foundation.Extensibility;
-using Uno.UI.Dispatching;
-using Uno.UI.Xaml;
-using Uno.UI.Xaml.Controls.Extensions;
-using System.Runtime.InteropServices.JavaScript;
-using Microsoft.UI.Xaml.Documents.TextFormatting;
-using Uno.Foundation;
-using DispatcherQueuePriority = Microsoft.UI.Dispatching.DispatcherQueuePriority;
-using Microsoft.UI.Xaml.Media.Media3D;
-using System.Numerics;
-using SelectionDetails = (int start, int length, bool selectionEndsAtTheStart);
 #endif
 
 namespace Microsoft.UI.Xaml.Controls
@@ -65,30 +46,9 @@ namespace Microsoft.UI.Xaml.Controls
 		public const string ButtonCollapsedStateName = "ButtonCollapsed";
 	}
 
-	public partial class TextBox : Control, IFrameworkTemplatePoolAware, ITextSelectionGripperHost
+	public partial class TextBox : Control, IFrameworkTemplatePoolAware
 	{
-		/// <summary>
-		/// This is a workaround for the template pooling issue where we change IsChecked when the template is recycled.
-		/// This prevents incorrect event raising but is not a "real" solution. Pooling could still cause issues.
-		/// This workaround can be removed if pooling is removed. See https://github.com/unoplatform/uno/issues/12189
-		/// </summary>
-		private bool _suppressTextChanged;
-		private bool _wasTemplateRecycled;
-
-#pragma warning disable CS0067, CS0649
-		private IFrameworkElement _placeHolder;
-		private ContentControl _contentElement;
-		private WeakReference<Button> _deleteButton;
-
-		private Action _selectionHighlightColorChanged;
-		private Action _foregroundBrushChanged;
-		private IDisposable _selectionHighlightBrushChangedSubscription;
-		private IDisposable _foregroundBrushChangedSubscription;
-#pragma warning restore CS0067, CS0649
-
-		private ContentPresenter _header;
 		protected private bool _isButtonEnabled = true;
-		protected private bool CanShowButton => !_core.IsEmpty && FocusState != FocusState.Unfocused && !IsReadOnly && !AcceptsReturn && TextWrapping == TextWrapping.NoWrap;
 
 		public event TextChangedEventHandler TextChanged;
 		public event TypedEventHandler<TextBox, TextBoxTextChangingEventArgs> TextChanging;
@@ -111,52 +71,14 @@ namespace Microsoft.UI.Xaml.Controls
 		internal void RaisePaste(TextControlPasteEventArgs args) => Paste?.Invoke(this, args);
 #endif
 
-		/// <summary>
-		/// Set when <see cref="TextChanged"/> event is being raised, to ensure modifications by handlers don't trigger an infinite loop.
-		/// </summary>
-		private bool _isInvokingTextChanged;
-		/// <summary>
-		/// Set when <see cref="TextChanging"/> event is being raised, to ensure modifications by handlers don't trigger an infinite loop.
-		/// </summary>
-		private bool _isInvokingTextChanging;
-		/// <summary>
-		/// Set when the <see cref="Text"/> property is being modified by user input.
-		/// </summary>
-		private bool _isInputModifyingText;
-		/// <summary>
-		/// Set when the <see cref="Text"/> property is being cleared via delete button.
-		/// </summary>
-		private bool _isInputClearingText;
-		/// <summary>
-		/// Indicates how many TextChanged events are pending. This is needed for AutoSuggestBox, which needs to
-		/// respond only to the last TextChange event, not all of them.
-		/// </summary>
-		private int _textChangedPendingCount;
-		/// <summary>
-		/// True if Text has changed while the TextBox has had focus, false otherwise
-		///
-		/// This flag is checked to avoid pushing a value to a two-way binding if no edits have occurred, per UWP's behavior.
-		/// </summary>
-		private bool _hasTextChangedThisFocusSession;
 
 		public TextBox()
 		{
 			_core = new TextBoxCore(this);
 
-			this.RegisterParentChangedCallbackStrong(this, OnParentChanged);
-
 			DefaultStyleKey = typeof(TextBox);
-			SizeChanged += OnSizeChanged;
 
-#if __SKIA__
-			ActualThemeChanged += (_, _) =>
-			{
-				TextBoxView?.DisplayBlock.InvalidateInlines(false);
-				TextBoxView?.UpdateTheme();
-			};
-			_timer.Tick += TimerOnTick;
-			EnsureHistory();
-#endif
+			_core.Initialize();
 
 			InitializePartial();
 		}
@@ -171,126 +93,27 @@ namespace Microsoft.UI.Xaml.Controls
 
 			OnLoadedPartial();
 
-			// This workaround is added in OnLoaded rather than OnApplyTemplate.
-			// Apparently, sometimes (e.g, Material style), the TextBox style setters are executed after OnApplyTemplate
-			// So, the style setters would override what the workaround does.
-			// OnLoaded appears to be executed after both OnApplyTemplate and after the style setters, making sure the values set here are not modified after.
-			if (_contentElement is ScrollViewer scrollViewer)
-			{
-				// The template of TextBox contains the following:
-				/*
-					HorizontalScrollBarVisibility="{TemplateBinding ScrollViewer.HorizontalScrollBarVisibility}"
-					HorizontalScrollMode="{TemplateBinding ScrollViewer.HorizontalScrollMode}"
-					VerticalScrollBarVisibility="{TemplateBinding ScrollViewer.VerticalScrollBarVisibility}"
-					VerticalScrollMode="{TemplateBinding ScrollViewer.VerticalScrollMode}"
-				 */
-				// Historically, TemplateBinding for attached DPs wasn't supported, and TextBox worked perfectly fine.
-				// When support for TemplateBinding for attached DPs was added, TextBox broke (test: TextBox_AutoGrow_Vertically_Wrapping_Test) because of
-				// change in the values of these properties. The following code serves as a workaround to set the values to what they used to be
-				// before the support for TemplateBinding for attached DPs.
-#if !__SKIA__
-				scrollViewer.HorizontalScrollMode = ScrollMode.Enabled; // The template sets this to Auto
-				scrollViewer.VerticalScrollMode = ScrollMode.Enabled; // The template sets this to Auto
-				scrollViewer.HorizontalScrollBarVisibility = ScrollBarVisibility.Disabled; // The template sets this to Hidden
-				scrollViewer.VerticalScrollBarVisibility = ScrollBarVisibility.Auto; // The template sets this to Hidden
-#endif
-
-			}
+			_core.OnLoadedCore();
 		}
 
 		private protected override void OnUnloaded()
 		{
 			base.OnUnloaded();
 
-			OnUnloadedPartial();
+			_core.OnUnloadedCore();
 		}
 
-		partial void OnUnloadedPartial();
-
-		private void OnSizeChanged(object sender, SizeChangedEventArgs args)
-		{
-			UpdateButtonStates();
-		}
-
-		private void OnParentChanged(object instance, object key, DependencyObjectParentChangedEventArgs args) => UpdateFontPartial();
-
-		private void InitializeProperties()
-		{
-			UpdatePlaceholderVisibility();
-			UpdateButtonStates();
-			OnInputScopeChanged(InputScope);
-			OnMaxLengthChanged(MaxLength);
-			OnAcceptsReturnChanged(AcceptsReturn);
-			OnIsReadonlyChanged();
-			OnForegroundColorChanged(null, Foreground);
-			UpdateFontPartial();
-			OnHeaderChanged();
-			OnIsTextPredictionEnabledChanged(IsTextPredictionEnabled);
-			OnSelectionHighlightColorChanged(null, SelectionHighlightColor);
-			OnIsSpellCheckEnabledChanged(IsSpellCheckEnabled);
-			OnTextAlignmentChanged(TextAlignment);
-			OnTextWrappingChanged();
-			OnFocusStateChanged(FocusState.Unfocused, FocusState, initial: true);
-			OnTextCharacterCasingChanged(CharacterCasing);
-			OnInputReturnTypeChanged(TextBoxExtensions.GetInputReturnType(this), initial: true);
-			UpdateDescriptionVisibility(true);
-			var buttonRef = _deleteButton?.GetTarget();
-
-			if (buttonRef != null)
-			{
-				var thisRef = (this as IWeakReferenceProvider).WeakReference;
-				buttonRef.Command = new DelegateCommand(() => (thisRef.Target as TextBox)?.DeleteButtonClick());
-			}
-
-			InitializePropertiesPartial();
-		}
-
-		protected override void OnGotFocus(RoutedEventArgs e)
-		{
-			_forceFocusedVisualState = false;
-			StartBringIntoView(new BringIntoViewOptions
-			{
-				AnimationDesired = false
-			});
-		}
+		protected override void OnGotFocus(RoutedEventArgs e) => _core.OnGotFocusCore();
 
 		protected override void OnApplyTemplate()
 		{
 			base.OnApplyTemplate();
 
-			// Ensures we don't keep a reference to a textBoxView that exists in a previous template
-			_textBoxView = null;
-
-			_placeHolder = GetTemplateChild(TextBoxConstants.PlaceHolderPartName) as IFrameworkElement;
-			_contentElement = GetTemplateChild(TextBoxConstants.ContentElementPartName) as ContentControl;
-			_header = GetTemplateChild(TextBoxConstants.HeaderContentPartName) as ContentPresenter;
-
-			if (GetTemplateChild(TextBoxConstants.DeleteButtonPartName) is Button button)
-			{
-				_deleteButton = new WeakReference<Button>(button);
-			}
-
-			if (_contentElement is { })
-			{
-				_contentElement.SetProtectedCursor(Microsoft.UI.Input.InputSystemCursor.Create(Microsoft.UI.Input.InputSystemCursorShape.IBeam));
-			}
-
-			UpdateTextBoxView();
-			InitializeProperties();
-			UpdateVisualState();
+			_core.ApplyTemplate();
 		}
-
-		partial void InitializePropertiesPartial();
 
 		internal void OnInputReturnTypeChanged(InputReturnType inputReturnType, bool initial)
-		{
-			if (inputReturnType != InputReturnType.Default || !initial)
-			{
-				SetInputReturnTypePlatform(inputReturnType);
-			}
-		}
-
-		partial void SetInputReturnTypePlatform(InputReturnType inputReturnType);
+			=> _core.OnInputReturnTypeChanged(inputReturnType, initial);
 
 		#region Text DependencyProperty
 
@@ -308,19 +131,6 @@ namespace Microsoft.UI.Xaml.Controls
 			}
 		}
 
-		private static string GetFirstLine(string value)
-		{
-			for (int i = 0; i < value.Length; i++)
-			{
-				var c = value[i];
-				if (c == '\r' || c == '\n')
-				{
-					return value.Substring(0, i);
-				}
-			}
-
-			return value;
-		}
 
 		public static DependencyProperty TextProperty { get; } =
 			DependencyProperty.Register(
@@ -331,184 +141,16 @@ namespace Microsoft.UI.Xaml.Controls
 					defaultValue: string.Empty,
 					options: FrameworkPropertyMetadataOptions.CoerceOnlyWhenChanged,
 					propertyChangedCallback: (s, e) => ((TextBox)s)?.OnTextChanged(e),
-					coerceValueCallback: (d, v, _) => ((TextBox)d)?.CoerceText(v)
+					coerceValueCallback: (d, v, _) => ((TextBox)d)?._core.CoerceText(v)
 				)
 			);
 
 		protected virtual void OnTextChanged(DependencyPropertyChangedEventArgs e)
-		{
-			_hasTextChangedThisFocusSession = true;
+			=> _core.OnTextChangedCore((string)e.OldValue, (string)e.NewValue);
 
-			RaiseTextChanging();
 
-#if !__SKIA__
-			if (!_isInputModifyingText)
-#endif
-			{
-				_textBoxView?.SetTextNative(Text);
-			}
 
-			UpdatePlaceholderVisibility();
 
-			OnTextChangedPartial();
-
-			// Notify automation peers of the text value change.
-			// WinUI CTextBox::UpdateTextProperty fires both ValueProperty changed
-			// and TextPatternOnTextChanged events inline when text changes.
-			var peer = GetOrCreateAutomationPeer();
-			if (peer is TextBoxAutomationPeer textPeer)
-			{
-				if (AutomationPeer.ListenerExistsHelper(AutomationEvents.PropertyChanged))
-				{
-					textPeer.RaiseValuePropertyChangedEvent(
-						(string)e.OldValue ?? string.Empty,
-						(string)e.NewValue ?? string.Empty);
-				}
-
-				if (AutomationPeer.ListenerExistsHelper(AutomationEvents.TextPatternOnTextChanged))
-				{
-					textPeer.RaiseAutomationEvent(AutomationEvents.TextPatternOnTextChanged);
-				}
-			}
-
-			// Update states after the text has changed, since we're
-			// using selection values to compute SV scrolling.
-			UpdateButtonStates();
-
-			var focusManager = VisualTree.GetFocusManagerForElement(this);
-			if (focusManager?.FocusedElement != this &&
-				GetBindingExpression(TextProperty) is
-				{
-					ParentBinding:
-					{
-						IsXBind: false, // NOTE: we UpdateSource in OnTextChanged only when the binding is not an x:Bind. WinUI's generated code for x:Bind contains a simple LostFocus subscription and waits for the next LostFocus even when not focused, unlike regular Bindings.
-						UpdateSourceTrigger: UpdateSourceTrigger.Default or UpdateSourceTrigger.LostFocus
-					}
-				} bindingExpression)
-			{
-				bindingExpression.UpdateSource(Text);
-			}
-
-			var isUserModifyingText = _isInputModifyingText | _isInputClearingText;
-			_textChangedPendingCount++;
-			_ = Dispatcher.RunAsync(CoreDispatcherPriority.Normal, () => RaiseTextChanged(isUserModifyingText));
-		}
-
-		partial void OnTextChangedPartial();
-
-		private void RaiseTextChanging()
-		{
-			if (!_isInvokingTextChanging)
-			{
-				try
-				{
-					_isInvokingTextChanging = true;
-					TextChanging?.Invoke(this, new TextBoxTextChangingEventArgs());
-				}
-				finally
-				{
-					_isInvokingTextChanging = false;
-				}
-			}
-		}
-
-		/// <summary>
-		/// This is called asynchronously after the UI changes in line with WinUI.
-		/// Note that no further native text box view text modification should
-		/// be performed in this method to avoid potential race conditions
-		/// (see #6289)
-		/// </summary>
-		private void RaiseTextChanged(bool isUserModifyingText)
-		{
-			_textChangedPendingCount--;
-			if (_isInvokingTextChanged)
-			{
-				return;
-			}
-
-			try
-			{
-				_isInvokingTextChanged = true;
-				if (!_suppressTextChanged) // This workaround can be removed if pooling is removed. See https://github.com/unoplatform/uno/issues/12189
-				{
-					TextChanged?.Invoke(this, new TextChangedEventArgs(this, isUserModifyingText, _textChangedPendingCount > 0));
-				}
-			}
-			finally
-			{
-				_isInvokingTextChanged = false;
-				_suppressTextChanged = false;
-			}
-		}
-
-		private void UpdatePlaceholderVisibility()
-		{
-			if (_placeHolder != null)
-			{
-				_placeHolder.Visibility = Text.IsNullOrEmpty() ? Visibility.Visible : Visibility.Collapsed;
-			}
-		}
-
-		private object CoerceText(object baseValue)
-		{
-			if (!(baseValue is string baseString))
-			{
-				return ""; //Pushing null to the binding resets the text. (Setting null to the Text property directly throws an exception.)
-			}
-
-			if (MaxLength > 0 && baseString.Length > MaxLength)
-			{
-				// Reject the new string if it's longer than the MaxLength
-#if __SKIA__
-				_pendingSelection = null;
-#endif
-				return DependencyProperty.UnsetValue;
-			}
-
-			if (!AcceptsReturn)
-			{
-				baseString = GetFirstLine(baseString);
-			}
-#if __SKIA__
-			else
-			{
-				// WinUI replaces all \n's and and \r\n's by \r. This is annoying because
-				// the _pendingSelection uses indices before this removal.
-				// On UIKit targets we use invisible overlay and replacing newlines would break the sync between
-				// the native input and the managed representation.
-				baseString = RemoveLF(baseString);
-			}
-
-			// make sure this coercion doesn't cause the pending selection to be out of range
-			if (_pendingSelection is { } selection2)
-			{
-				var start = Math.Min(selection2.start, baseString.Length);
-				var end = Math.Min(selection2.start + selection2.length, baseString.Length);
-				_pendingSelection = (start, end - start);
-			}
-#endif
-
-			var args = new TextBoxBeforeTextChangingEventArgs(baseString);
-			BeforeTextChanging?.Invoke(this, args);
-			if (args.Cancel)
-			{
-#if __SKIA__
-				// On WinUI, when a selection is canceled, the TextBox invokes a bunch of weird
-				// SelectionChanging events followed by a bunch of matching SelectionChanged.
-				// Probing for the value of SelectionStart and SelectionLength during these SelectionChanging
-				// events will give incorrect transient values and the SelectionChanged events will end up
-				// with the selection where it started (before the text change). Also, the direction of
-				// of the selection will be reset, i.e. if the selection end was "at the start", then it won't be
-				// so anymore.
-				// In Uno, we choose a simpler sequence. We just reset the selection direction (like WinUI) and
-				// we don't invoke any selection change events (since selection was in fact not changed).
-				_pendingSelection = (SelectionStart, SelectionLength);
-#endif
-				return DependencyProperty.UnsetValue;
-			}
-
-			return baseString;
-		}
 
 		#endregion
 
@@ -528,65 +170,47 @@ namespace Microsoft.UI.Xaml.Controls
 				typeof(TextBox),
 				new FrameworkPropertyMetadata(
 					defaultValue: null,
-					propertyChangedCallback: (s, e) => ((TextBox)s)?.UpdateDescriptionVisibility(false)
+					propertyChangedCallback: (s, e) => ((TextBox)s)?._core.UpdateDescriptionVisibility(false)
 				)
 			);
 
-		private void UpdateDescriptionVisibility(bool initialization)
-		{
-			if (initialization && Description == null)
-			{
-				// Avoid loading DescriptionPresenter element in template if not needed.
-				return;
-			}
-
-			var descriptionPresenter = this.FindName("DescriptionPresenter") as ContentPresenter;
-			if (descriptionPresenter != null)
-			{
-				descriptionPresenter.Visibility = Description != null ? Visibility.Visible : Visibility.Collapsed;
-			}
-		}
 		#endregion
 
 		protected override void OnFontSizeChanged(double oldValue, double newValue)
 		{
 			base.OnFontSizeChanged(oldValue, newValue);
-			UpdateFontPartial();
+			_core.UpdateFont();
 		}
 
 		protected override void OnFontFamilyChanged(FontFamily oldValue, FontFamily newValue)
 		{
 			base.OnFontFamilyChanged(oldValue, newValue);
-			UpdateFontPartial();
+			_core.UpdateFont();
 		}
 
 		protected override void OnFontStyleChanged(FontStyle oldValue, FontStyle newValue)
 		{
 			base.OnFontStyleChanged(oldValue, newValue);
-			UpdateFontPartial();
+			_core.UpdateFont();
 		}
 
 		private protected override void OnFontStretchChanged(FontStretch oldValue, FontStretch newValue)
 		{
 			base.OnFontStretchChanged(oldValue, newValue);
-			UpdateFontPartial();
+			_core.UpdateFont();
 		}
 
 		protected override void OnFontWeightChanged(FontWeight oldValue, FontWeight newValue)
 		{
 			base.OnFontWeightChanged(oldValue, newValue);
-			UpdateFontPartial();
+			_core.UpdateFont();
 		}
 
-		partial void UpdateFontPartial();
 
 		protected override void OnForegroundColorChanged(Brush oldValue, Brush newValue)
-		{
-			_foregroundBrushChangedSubscription?.Dispose();
-			_foregroundBrushChangedSubscription = Brush.SetupBrushChanged(newValue, ref _foregroundBrushChanged, () => OnForegroundColorChangedPartial(newValue));
-		}
+			=> _core.OnForegroundColorChanged(newValue);
 
-		partial void OnForegroundColorChangedPartial(Brush newValue);
+		internal string ProcessTextInput(string newText) => _core.ProcessTextInput(newText);
 
 		#region PlaceholderText DependencyProperty
 
@@ -627,18 +251,9 @@ namespace Microsoft.UI.Xaml.Controls
 				typeof(TextBox),
 				new FrameworkPropertyMetadata(
 					DefaultBrushes.SelectionHighlightColor,
-					propertyChangedCallback: (s, e) => ((TextBox)s)?.OnSelectionHighlightColorChanged((SolidColorBrush)e.OldValue, (SolidColorBrush)e.NewValue)));
+					propertyChangedCallback: (s, e) => ((TextBox)s)?._core.OnSelectionHighlightColorChanged((SolidColorBrush)e.OldValue, (SolidColorBrush)e.NewValue)));
 
-		private void OnSelectionHighlightColorChanged(SolidColorBrush oldBrush, SolidColorBrush newBrush)
-		{
-			oldBrush ??= DefaultBrushes.SelectionHighlightColor;
-			newBrush ??= DefaultBrushes.SelectionHighlightColor;
 
-			_selectionHighlightBrushChangedSubscription?.Dispose();
-			_selectionHighlightBrushChangedSubscription = Brush.SetupBrushChanged(newBrush, ref _selectionHighlightColorChanged, () => OnSelectionHighlightColorChangedPartial(newBrush));
-		}
-
-		partial void OnSelectionHighlightColorChangedPartial(SolidColorBrush brush);
 
 		#endregion
 
@@ -689,12 +304,10 @@ namespace Microsoft.UI.Xaml.Controls
 							}
 						}
 					},
-					propertyChangedCallback: (s, e) => ((TextBox)s)?.OnInputScopeChanged((InputScope)e.NewValue)
+					propertyChangedCallback: (s, e) => ((TextBox)s)?._core.OnInputScopeChanged((InputScope)e.NewValue)
 				)
 			);
 
-		private void OnInputScopeChanged(InputScope newValue) => OnInputScopeChangedPartial(newValue);
-		partial void OnInputScopeChangedPartial(InputScope newValue);
 
 		#endregion
 
@@ -713,13 +326,11 @@ namespace Microsoft.UI.Xaml.Controls
 				typeof(TextBox),
 				new FrameworkPropertyMetadata(
 					defaultValue: 0,
-					propertyChangedCallback: (s, e) => ((TextBox)s)?.OnMaxLengthChanged((int)e.NewValue)
+					propertyChangedCallback: (s, e) => ((TextBox)s)?._core.OnMaxLengthChanged((int)e.NewValue)
 				)
 			);
 
-		private void OnMaxLengthChanged(int newValue) => OnMaxLengthChangedPartial(newValue);
 
-		partial void OnMaxLengthChangedPartial(int newValue);
 
 		#endregion
 
@@ -738,27 +349,11 @@ namespace Microsoft.UI.Xaml.Controls
 				typeof(TextBox),
 				new FrameworkPropertyMetadata(
 					defaultValue: false,
-					propertyChangedCallback: (s, e) => ((TextBox)s)?.OnAcceptsReturnChanged((bool)e.NewValue)
+					propertyChangedCallback: (s, e) => ((TextBox)s)?._core.OnAcceptsReturnChanged((bool)e.NewValue)
 				)
 			);
 
-		private void OnAcceptsReturnChanged(bool newValue)
-		{
-			if (!newValue)
-			{
-				var text = Text;
-				var singleLineText = GetFirstLine(text);
-				if (text != singleLineText)
-				{
-					Text = singleLineText;
-				}
-			}
 
-			OnAcceptsReturnChangedPartial(newValue);
-			UpdateButtonStates();
-		}
-
-		partial void OnAcceptsReturnChangedPartial(bool newValue);
 
 		#endregion
 
@@ -777,16 +372,10 @@ namespace Microsoft.UI.Xaml.Controls
 				new FrameworkPropertyMetadata(
 					defaultValue: TextWrapping.NoWrap,
 					options: FrameworkPropertyMetadataOptions.AffectsMeasure,
-					propertyChangedCallback: (s, e) => ((TextBox)s)?.OnTextWrappingChanged())
+					propertyChangedCallback: (s, e) => ((TextBox)s)?._core.OnTextWrappingChanged())
 				);
 
-		private void OnTextWrappingChanged()
-		{
-			OnTextWrappingChangedPartial();
-			UpdateButtonStates();
-		}
 
-		partial void OnTextWrappingChangedPartial();
 
 		#endregion
 #if SUPPORTS_RTL
@@ -795,11 +384,10 @@ namespace Microsoft.UI.Xaml.Controls
 			base.OnPropertyChanged2(args);
 			if (args.Property == FrameworkElement.FlowDirectionProperty)
 			{
-				OnFlowDirectionChangedPartial();
+				_core.OnFlowDirectionChanged();
 			}
 		}
 
-		partial void OnFlowDirectionChangedPartial();
 #endif
 
 #if IS_UNIT_TESTS || __SKIA__ || __NETSTD_REFERENCE__
@@ -821,15 +409,10 @@ namespace Microsoft.UI.Xaml.Controls
 				typeof(TextBox),
 				new FrameworkPropertyMetadata(
 						defaultValue: CharacterCasing.Normal,
-						propertyChangedCallback: (s, e) => ((TextBox)s)?.OnTextCharacterCasingChanged((CharacterCasing)e.NewValue))
+						propertyChangedCallback: (s, e) => ((TextBox)s)?._core.OnTextCharacterCasingChanged((CharacterCasing)e.NewValue))
 				);
 
-		private void OnTextCharacterCasingChanged(CharacterCasing newValue)
-		{
-			OnTextCharacterCasingChangedPartial(newValue);
-		}
 
-		partial void OnTextCharacterCasingChangedPartial(CharacterCasing newValue);
 
 		#region IsReadOnly DependencyProperty
 
@@ -846,17 +429,11 @@ namespace Microsoft.UI.Xaml.Controls
 				typeof(TextBox),
 				new FrameworkPropertyMetadata(
 					false,
-					propertyChangedCallback: (s, e) => ((TextBox)s)?.OnIsReadonlyChanged()
+					propertyChangedCallback: (s, e) => ((TextBox)s)?._core.OnIsReadonlyChanged()
 				)
 			);
 
-		private void OnIsReadonlyChanged()
-		{
-			OnIsReadonlyChangedPartial();
-			UpdateButtonStates();
-		}
 
-		partial void OnIsReadonlyChangedPartial();
 
 		#endregion
 
@@ -876,7 +453,7 @@ namespace Microsoft.UI.Xaml.Controls
 				new FrameworkPropertyMetadata(
 					defaultValue: null,
 					options: FrameworkPropertyMetadataOptions.AffectsMeasure,
-					propertyChangedCallback: (s, e) => ((TextBox)s)?.OnHeaderChanged()
+					propertyChangedCallback: (s, e) => ((TextBox)s)?._core.OnHeaderChanged()
 				)
 			);
 
@@ -894,19 +471,10 @@ namespace Microsoft.UI.Xaml.Controls
 				new FrameworkPropertyMetadata(
 					defaultValue: null,
 					options: FrameworkPropertyMetadataOptions.ValueDoesNotInheritDataContext | FrameworkPropertyMetadataOptions.AffectsMeasure,
-					propertyChangedCallback: (s, e) => ((TextBox)s)?.OnHeaderChanged()
+					propertyChangedCallback: (s, e) => ((TextBox)s)?._core.OnHeaderChanged()
 				)
 			);
 
-		private void OnHeaderChanged()
-		{
-			var headerVisibility = (Header != null || HeaderTemplate != null) ? Visibility.Visible : Visibility.Collapsed;
-
-			if (_header != null)
-			{
-				_header.Visibility = headerVisibility;
-			}
-		}
 
 		#endregion
 
@@ -925,13 +493,11 @@ namespace Microsoft.UI.Xaml.Controls
 				typeof(TextBox),
 				new FrameworkPropertyMetadata(
 					defaultValue: true,
-					propertyChangedCallback: (s, e) => ((TextBox)s)?.OnIsSpellCheckEnabledChanged((bool)e.NewValue)
+					propertyChangedCallback: (s, e) => ((TextBox)s)?._core.OnIsSpellCheckEnabledChanged((bool)e.NewValue)
 				)
 			);
 
-		private void OnIsSpellCheckEnabledChanged(bool newValue) => OnIsSpellCheckEnabledChangedPartial(newValue);
 
-		partial void OnIsSpellCheckEnabledChangedPartial(bool newValue);
 
 		#endregion
 
@@ -952,13 +518,11 @@ namespace Microsoft.UI.Xaml.Controls
 				typeof(TextBox),
 				new FrameworkPropertyMetadata(
 					defaultValue: true,
-					propertyChangedCallback: (s, e) => ((TextBox)s)?.OnIsTextPredictionEnabledChanged((bool)e.NewValue)
+					propertyChangedCallback: (s, e) => ((TextBox)s)?._core.OnIsTextPredictionEnabledChanged((bool)e.NewValue)
 				)
 			);
 
-		private void OnIsTextPredictionEnabledChanged(bool newValue) => OnIsTextPredictionEnabledChangedPartial(newValue);
 
-		partial void OnIsTextPredictionEnabledChangedPartial(bool newValue);
 
 		#endregion
 
@@ -978,12 +542,10 @@ namespace Microsoft.UI.Xaml.Controls
 				new FrameworkPropertyMetadata(
 					TextAlignment.Left,
 					FrameworkPropertyMetadataOptions.AffectsMeasure,
-					(s, e) => ((TextBox)s)?.OnTextAlignmentChanged((TextAlignment)e.NewValue)));
+					(s, e) => ((TextBox)s)?._core.OnTextAlignmentChanged((TextAlignment)e.NewValue)));
 
 
-		private void OnTextAlignmentChanged(TextAlignment newValue) => OnTextAlignmentChangedPartial(newValue);
 
-		partial void OnTextAlignmentChangedPartial(TextAlignment newValue);
 
 		#endregion
 
@@ -1021,45 +583,11 @@ namespace Microsoft.UI.Xaml.Controls
 			base.UpdateFocusState(focusState);
 			if (oldValue != focusState)
 			{
-				OnFocusStateChanged(oldValue, focusState, initial: false);
+				_core.OnFocusStateChanged(oldValue, focusState, initial: false);
 			}
 		}
 
-		private void OnFocusStateChanged(FocusState oldValue, FocusState newValue, bool initial)
-		{
-			OnFocusStateChangedPartial(newValue, initial);
 
-			if (_forceFocusedVisualState && newValue == FocusState.Unfocused)
-			{
-				// Context flyout is taking focus - skip binding updates and keep
-				// _hasTextChangedThisFocusSession. Deferred until actual focus loss.
-				UpdateVisualState();
-				return;
-			}
-
-			if (!initial && newValue == FocusState.Unfocused && _hasTextChangedThisFocusSession)
-			{
-				if (!_wasTemplateRecycled &&
-					GetBindingExpression(TextProperty) is { ParentBinding.UpdateSourceTrigger: UpdateSourceTrigger.LostFocus or UpdateSourceTrigger.Default } bindingExpression)
-				{
-					// Manually update Source when losing focus because TextProperty's default UpdateSourceTrigger is Explicit
-					bindingExpression.UpdateSource(Text);
-				}
-
-				_wasTemplateRecycled = false;
-			}
-
-			UpdateButtonStates();
-
-			if (newValue == FocusState.Unfocused)
-			{
-				_hasTextChangedThisFocusSession = false;
-			}
-
-			UpdateVisualState();
-		}
-
-		partial void OnFocusStateChangedPartial(FocusState focusState, bool initial);
 
 		protected override void OnVisibilityChanged(Visibility oldValue, Visibility newValue)
 		{
@@ -1093,87 +621,25 @@ namespace Microsoft.UI.Xaml.Controls
 			base.OnPointerCaptureLost(e);
 			_isPointerOver = false;
 			UpdateVisualState();
-			OnPointerCaptureLostPartial(e);
+			_core.OnPointerCaptureLost(e);
 		}
 
 		protected override void OnPointerPressed(PointerRoutedEventArgs args)
 		{
 			base.OnPointerPressed(args);
 
-			bool isPointerCaptureRequired =
-				true;
-
-			if (ShouldFocusOnPointerPressed(args)) // UWP Captures if the pointer is not Touch
-			{
-				var wasFocused = FocusState != FocusState.Unfocused;
-				if (isPointerCaptureRequired)
-				{
-					if (CapturePointer(args.Pointer))
-					{
-						Focus(FocusState.Pointer);
-					}
-				}
-				else
-				{
-					Focus(FocusState.Pointer);
-				}
-
-#if __SKIA__
-				if (wasFocused)
-				{
-					// See comment in OnPointerReleased for why we do this
-					_textBoxNotificationsSingleton?.OnFocused(this);
-				}
-#endif
-			}
-
-			args.Handled = true;
-
-			OnPointerPressedPartial(args);
+			_core.OnPointerPressed(args);
 		}
 
-		partial void OnPointerPressedPartial(PointerRoutedEventArgs args);
 
-		partial void OnPointerReleasedPartial(PointerRoutedEventArgs args, bool wasFocused);
 
-		partial void OnPointerCaptureLostPartial(PointerRoutedEventArgs e);
 
 		/// <inheritdoc />
 		protected override void OnPointerReleased(PointerRoutedEventArgs args)
 		{
 			base.OnPointerReleased(args);
 
-			bool wasFocused = FocusState != FocusState.Unfocused;
-			if (!ShouldFocusOnPointerPressed(args))
-			{
-				// Ported from: TextBoxBase.cpp OnPointerReleased
-				// Don't take focus if the context flyout is open.
-#if __SKIA__
-				if (!TextControlFlyoutHelper.IsOpen(ContextFlyout))
-#endif
-				{
-					Focus(FocusState.Pointer);
-				}
-#if __SKIA__
-				if (wasFocused)
-				{
-					// We already call UpdateFocusState in TextBoxView when focus changes, but this is not enough.
-					// UpdateFocusState should be called here even if the TextBox was already focused.
-					// This is to support re-showing the keyboard when clicking on an already-focused TextBox.
-					// For example:
-					// 1. User taps on TextBox and it gained focus and soft keyboard was shown.
-					// 2. User hides the keyboard, but TextBox is still focused.
-					// 3. User taps on TextBox again. In this case, we want to call UpdateFocusState so that the soft keyboard is re-shown again.
-					//
-					// This approach feels hacky though and may not handle programmatic focus properly, i.e, when programmatic focus is requested on an already-focused TextBox. This is a niche case though.
-					_textBoxNotificationsSingleton?.OnFocused(this);
-				}
-#endif
-			}
-
-			args.Handled = true;
-
-			OnPointerReleasedPartial(args, wasFocused);
+			_core.OnPointerReleased(args);
 		}
 
 		protected override void OnTapped(TappedRoutedEventArgs e)
@@ -1185,22 +651,17 @@ namespace Microsoft.UI.Xaml.Controls
 
 		partial void OnTappedPartial();
 
-		partial void OnKeyDownPartial(KeyRoutedEventArgs args);
 
 		protected override void OnKeyDown(KeyRoutedEventArgs args)
 		{
 			base.OnKeyDown(args);
 
-			OnKeyDownPartial(args);
+			_core.OnKeyDown(args);
 		}
 
 		private protected override void OnPostKeyDown(KeyRoutedEventArgs args)
 		{
-#if __SKIA__
-			OnKeyDownSkia(args);
-#else
-			OnKeyDownNonSkia(args);
-#endif
+			_core.OnPostKeyDown(args);
 
 			var modifiers = CoreImports.Input_GetKeyboardModifiers();
 			if (!args.Handled && KeyboardAcceleratorUtility.IsKeyValidForAccelerators(args.Key, KeyboardAcceleratorUtility.MapVirtualKeyModifiersToIntegersModifiers(modifiers)))
@@ -1214,348 +675,43 @@ namespace Microsoft.UI.Xaml.Controls
 		}
 
 #if !__SKIA__
-		private void OnKeyDownNonSkia(KeyRoutedEventArgs args)
-		{
-			var selectionStart = SelectionStart;
-
-			// Note: On windows only keys that are "moving the cursor" are handled
-			//		 AND ** only KeyDown ** is handled (not KeyUp)
-			switch (args.Key)
-			{
-				case VirtualKey.Up:
-					if (AcceptsReturn)
-					{
-						args.Handled = true;
-					}
-					break;
-				case VirtualKey.Down:
-					if (selectionStart != Text.Length)
-					{
-						SelectionStart = Text.Length;
-						args.Handled = true;
-					}
-					if (AcceptsReturn)
-					{
-						args.Handled = true;
-					}
-					break;
-				case VirtualKey.Left:
-					if (selectionStart != 0)
-					{
-						args.Handled = true;
-					}
-					break;
-				case VirtualKey.Right:
-					if (selectionStart != Text.Length)
-					{
-						args.Handled = true;
-					}
-					break;
-				case VirtualKey.Home:
-				case VirtualKey.End:
-					args.Handled = true;
-					break;
-			}
-
-		}
 #endif
 
-		protected virtual void UpdateButtonStates()
-		{
-			if (this.Log().IsEnabled(LogLevel.Debug))
-			{
-				this.Log().LogDebug(nameof(UpdateButtonStates));
-			}
-
-			var changed = false;
-			// Minimum width for TextBox with DeleteButton visible is 5em.
-			if (CanShowButton && _isButtonEnabled && ActualWidth > FontSize * 5)
-			{
-				changed |= VisualStateManager.GoToState(this, TextBoxConstants.ButtonVisibleStateName, true);
-			}
-			else
-			{
-				changed |= VisualStateManager.GoToState(this, TextBoxConstants.ButtonCollapsedStateName, true);
-			}
-
-#if __SKIA__
-			DispatchUpdateScrolling();
-#endif
-		}
+		protected virtual void UpdateButtonStates() => _core.UpdateButtonStatesCore();
 
 
-#if __SKIA__
-		bool _pendingUpdateScrolling;
 
-		private void DispatchUpdateScrolling()
-		{
-			if (!_pendingUpdateScrolling)
-			{
-				_pendingUpdateScrolling = true;
 
-				// We may be pushing scrolling updates too often
-				// when pushing keystrokes programmatically.
-				DispatcherQueue.TryEnqueue(() =>
-				{
-					_pendingUpdateScrolling = false;
 
-					UpdateScrolling();
-				});
-			}
-		}
-#endif
-
-		/// <summary>
-		/// Respond to text input from user interaction.
-		/// </summary>
-		/// <param name="newText">The most recent version of the text from the input field.</param>
-		/// <returns>The value of the <see cref="Text"/> property, which may have been modified programmatically.</returns>
-		internal string ProcessTextInput(string newText)
-		{
-			var isCurrentlyModifying = _isInputModifyingText;
-
-			try
-			{
-				_isInputModifyingText = true;
-				var oldText = Text;
-				Text = newText;
-
-#if __SKIA__
-				if (_pendingSelection is { } selection && Text == oldText)
-				{
-					// OnTextChanged won't fire, so we immediately change the selection.
-					// Note how we check that Text (after assignment) == oldText and
-					// not oldText == newText. This is because CoerceText can make it so that
-					// newText != oldText but Text (after assignment) == oldText
-					SelectInternal(selection.start, selection.length);
-				}
-#endif
-			}
-			finally
-			{
-				if (!isCurrentlyModifying)
-				{
-					// The all to ProcessTextInput may be recursing, we only want to restore
-					// the state on the last one.
-					_isInputModifyingText = false;
-				}
-			}
-
-			return Text; //This may have been modified by BeforeTextChanging, TextChanging, DP callback, etc
-		}
-
-		private void DeleteButtonClick()
-		{
-			try
-			{
-				_isInputClearingText = true;
-
-				Text = string.Empty;
-				OnDeleteButtonClickPartial();
-			}
-			finally
-			{
-				_isInputClearingText = false;
-			}
-		}
-
-		partial void OnDeleteButtonClickPartial();
 
 		internal void OnSelectionChanged() => SelectionChanged?.Invoke(this, new RoutedEventArgs(this));
 
 		void IFrameworkTemplatePoolAware.OnTemplateRecycled()
 		{
-			_suppressTextChanged = true;
-			Text = string.Empty;
-			_wasTemplateRecycled = true;
+			_core.OnTemplateRecycled();
 		}
 
 		protected override AutomationPeer OnCreateAutomationPeer() => new TextBoxAutomationPeer(this);
 
 		public override string GetAccessibilityInnerText() => Text;
 
-		public void Select(int start, int length)
-		{
-			if (start < 0)
-			{
-				throw new ArgumentException($"'{start}' cannot be negative.", nameof(start));
-			}
+		public void Select(int start, int length) => _core.Select(start, length);
 
-			if (length < 0)
-			{
-				throw new ArgumentException($"'{length}' cannot be negative.", nameof(length));
-			}
+		public void SelectAll() => _core.SelectAll();
 
-			// TODO: Test and adjust (if needed) this logic for surrogate pairs.
+		public void PasteFromClipboard() => _core.PasteFromClipboard();
 
-			var textLength = Text.Length;
 
-			if (start >= textLength)
-			{
-				start = textLength;
-				length = 0;
-			}
-			else if (start + length > textLength)
-			{
-				length = textLength - start;
-			}
-
-#if __SKIA__
-			_pendingSelection = null;
-#endif
-
-			if (SelectionStart == start && SelectionLength == length)
-			{
-				return;
-			}
-
-			var textBoxSelectionChangingEventArgs = new TextBoxSelectionChangingEventArgs(start, length);
-			SelectionChanging?.Invoke(this, textBoxSelectionChangingEventArgs);
-			if (!textBoxSelectionChangingEventArgs.Cancel || textBoxSelectionChangingEventArgs.SelectionStart + textBoxSelectionChangingEventArgs.SelectionLength > Text.Length)
-			{
-				SelectPartial(start, length);
-				OnSelectionChanged();
-			}
-		}
-
-		public void SelectAll() => SelectAllPartial();
-
-		partial void SelectPartial(int start, int length);
-
-		partial void SelectAllPartial();
-
-		public void PasteFromClipboard()
-		{
-			_ = Dispatcher.RunAsync(CoreDispatcherPriority.High, async () =>
-			{
-				var content = Clipboard.GetContent();
-				string clipboardText;
-				if (content.AvailableFormats.Contains(StandardDataFormats.Text))
-				{
-					try
-					{
-						clipboardText = await content.GetTextAsync();
-						PasteFromClipboard(clipboardText);
-					}
-					catch (InvalidOperationException e)
-					{
-						if (this.Log().IsEnabled(LogLevel.Debug))
-						{
-							this.Log().Debug("TextBox.PasteFromClipboard failed during DataPackageView.GetTextAsync: " + e);
-						}
-					}
-				}
-			});
-		}
-
-		/// <summary>
-		/// Copies content from the OS clipboard into the text control.
-		/// </summary>
-		internal void PasteFromClipboard(string clipboardText)
-		{
-			if (IsReadOnly)
-			{
-				return;
-			}
-
-			var selectionStart = SelectionStart;
-			var selectionLength = SelectionLength;
-			var currentText = Text;
-			var adjustedClipboardText = clipboardText;
-
-			if (selectionLength > 0)
-			{
-				currentText = currentText.Remove(selectionStart, selectionLength);
-			}
-
-			if (MaxLength > 0)
-			{
-				var clipboardRangeToBePasted = Math.Max(0, Math.Min(clipboardText.Length, MaxLength - currentText.Length));
-				adjustedClipboardText = clipboardText[..clipboardRangeToBePasted];
-			}
-
-			currentText = currentText.Insert(selectionStart, adjustedClipboardText);
-			PasteFromClipboardPartial(adjustedClipboardText, selectionStart, currentText);
-
-#if __SKIA__
-			try
-			{
-				_clearHistoryOnTextChanged = false;
-				_suppressCurrentlyTyping = true;
-#else
-			{
-#endif
-				ProcessTextInput(currentText);
-			}
-#if __SKIA__
-			finally
-			{
-				_suppressCurrentlyTyping = false;
-				_clearHistoryOnTextChanged = true;
-				if (Text.IsNullOrEmpty())
-				{
-					// On WinUI, the caret never has thumbs if there is no text
-					CaretMode = CaretDisplayMode.ThumblessCaretShowing;
-				}
-			}
-#endif
-
-#if !IS_UNIT_TESTS
-			RaisePaste(new TextControlPasteEventArgs());
-#endif
-		}
-
-		partial void PasteFromClipboardPartial(string adjustedClipboardText, int selectionStart, string newText);
 
 		/// <summary>
 		/// Copies the selected content to the OS clipboard.
 		/// </summary>
-		public void CopySelectionToClipboard()
-		{
-			if (this is PasswordBox)
-			{
-				return;
-			}
-
-			if (SelectionLength > 0)
-			{
-				var text = SelectedText;
-				var dataPackage = new DataPackage();
-				dataPackage.SetText(text);
-				Clipboard.SetContent(dataPackage);
-			}
-		}
+		public void CopySelectionToClipboard() => _core.CopySelectionToClipboard();
 
 		/// <summary>
 		/// Moves the selected content to the OS clipboard and removes it from the text control.
 		/// </summary>
-		public void CutSelectionToClipboard()
-		{
-			if (IsReadOnly || this is PasswordBox)
-			{
-				return;
-			}
-
-			CopySelectionToClipboard();
-			CutSelectionToClipboardPartial();
-#if __SKIA__
-			try
-			{
-				_suppressCurrentlyTyping = true;
-#else
-			{
-#endif
-				Text = Text.Remove(SelectionStart, SelectionLength);
-			}
-#if __SKIA__
-			finally
-			{
-				_suppressCurrentlyTyping = false;
-			}
-#endif
-		}
-
-		partial void CutSelectionToClipboardPartial();
+		public void CutSelectionToClipboard() => _core.CutSelectionToClipboard();
 
 		internal override bool CanHaveChildren() => true;
 
@@ -1576,1925 +732,5 @@ namespace Microsoft.UI.Xaml.Controls
 		}
 
 		partial void OnIsEnabledChangedPartial(IsEnabledChangedEventArgs e);
-
-		private bool ShouldFocusOnPointerPressed(PointerRoutedEventArgs args) =>
-			// For mouse and pen, the TextBox should focus on pointer press
-			// (and then capture pointer to make sure to handle the whol down->move->up sequence).
-			// For touch we wait for the release to focus (avoid flickering in case of cancel due to scroll for instance).
-			args.Pointer.PointerDeviceType != PointerDeviceType.Touch;
-
-		private TextSelectionGripperPresenter _gripperPresenter;
-
-		// Test hook: the pair of touch-selection grippers when they are currently showing, otherwise null.
-		internal (CaretWithStemAndThumb start, CaretWithStemAndThumb end)? VisibleGrippersForTesting => _gripperPresenter?.VisibleGrippersForTesting;
-
-		private TextBoxView _textBoxView;
-		private static ITextBoxNotificationsProviderSingleton _textBoxNotificationsSingleton;
-		private SerialDisposable _clipboardChangeSubscription = new SerialDisposable();
-
-		private SelectionDetails _selection;
-		private float _caretXOffset; // this is not necessarily the visual offset of the caret, but where the caret is logically supposed to be when moving up and down with the keyboard, even if the caret is temporarily elsewhere
-		private CaretDisplayMode _caretMode = CaretDisplayMode.ThumblessCaretHidden;
-
-		private bool _inSelectInternal;
-
-		private (int start, int length)? _pendingSelection;
-
-		private bool _clearHistoryOnTextChanged = true;
-		private CompositionBrush _cachedCaretBrush;
-		private Color _cachedCaretColor;
-
-		private static readonly VirtualKeyModifiers _platformCtrlKey;
-
-		// We track what constitutes one typing "action" that can be undone/redone. The general gist is that
-		// any sequence of characters (with backspace allowed) without any navigation moves (pointer click, arrow keys, etc.)
-		// will be one "run"/"action". However, there are some arbitrary exceptions, so that is only a rule of thumb.
-		private bool _currentlyTyping;
-		private bool _suppressCurrentlyTyping;
-		private SelectionDetails _selectionWhenTypingStarted;
-		private string _textWhenTypingStarted;
-
-		private int _historyIndex;
-		private (int start, int length, bool isBackward) _lastNotifiedSelection = (-1, -1, false);
-		private readonly List<HistoryRecord> _history = new(); // the selection of an action is what was selected right before it happened. Might turn out to be unnecessary.
-
-		private readonly DispatcherTimer _timer = new() { Interval = TimeSpan.FromSeconds(0.5) };
-
-		private MenuFlyout _proofingMenu;
-
-		internal bool IsBackwardSelection => _selection.selectionEndsAtTheStart;
-
-		internal TextBoxView TextBoxView => _textBoxView;
-
-		internal ContentControl ContentElement => _contentElement;
-
-		static TextBox()
-		{
-			_platformCtrlKey = Uno.UI.Helpers.DeviceTargetHelper.PlatformCommandModifier;
-			_ = ApiExtensibility.CreateInstance(null, out _textBoxNotificationsSingleton);
-			InitializeIme();
-		}
-
-		internal CaretDisplayMode CaretMode
-		{
-			get => _caretMode;
-			private set
-			{
-				if (_caretMode != value)
-				{
-					_caretMode = value;
-					UpdateDisplaySelection();
-					if (TextBoxView?.DisplayBlock.Visual is { } visual)
-					{
-						Visual.Compositor.InvalidateRender(visual);
-					}
-					if (value is CaretDisplayMode.ThumblessCaretShowing)
-					{
-						_timer.Start(); // restart
-					}
-					else if (value is CaretDisplayMode.CaretWithThumbsBothEndsShowing
-							 or CaretDisplayMode.CaretWithThumbsOnlyEndShowing)
-					{
-						_timer.Stop();
-					}
-				}
-			}
-		}
-
-		// Settable so runtime tests can force a mobile convention on Skia Desktop, where
-		// OperatingSystem.IsAndroid()/IsIOS() are both false and the OS-derived default is Desktop.
-		internal TouchTextSelectionConvention TouchSelectionConvention { get; set; } = GetDefaultTouchTextSelectionConvention();
-
-		private static TouchTextSelectionConvention GetDefaultTouchTextSelectionConvention()
-			=> OperatingSystem.IsAndroid() ? TouchTextSelectionConvention.Android
-				: Uno.UI.Helpers.DeviceTargetHelper.IsUIKit() ? TouchTextSelectionConvention.iOS
-				: TouchTextSelectionConvention.Desktop;
-
-		public static DependencyProperty CanUndoProperty { get; } = DependencyProperty.Register(
-			nameof(CanUndo),
-			typeof(bool),
-			typeof(TextBox),
-			new FrameworkPropertyMetadata(defaultValue: false)
-			{
-				PropMethodCall = GetCanUndo,
-			});
-
-		public bool CanUndo => (bool)GetValue(CanUndoProperty);
-
-#nullable enable
-		private static object? GetCanUndo(DependencyObject instance, bool isGet, object? valueToSet)
-		{
-			if (!isGet)
-			{
-				throw new InvalidOperationException($"{nameof(CanUndoProperty)} is read-only.");
-			}
-
-			return Uno.UI.Helpers.Boxes.Box(((TextBox)instance)._historyIndex > 0);
-		}
-#nullable restore
-
-		public static DependencyProperty CanRedoProperty { get; } = DependencyProperty.Register(
-			nameof(CanRedo),
-			typeof(bool),
-			typeof(TextBox),
-			new FrameworkPropertyMetadata(defaultValue: false)
-			{
-				PropMethodCall = GetCanRedo,
-			});
-
-		public bool CanRedo => (bool)GetValue(CanRedoProperty);
-
-#nullable enable
-		private static object? GetCanRedo(DependencyObject instance, bool isGet, object? valueToSet)
-		{
-			if (!isGet)
-			{
-				throw new InvalidOperationException($"{nameof(CanRedoProperty)} is read-only.");
-			}
-
-			var textBox = (TextBox)instance;
-			return Uno.UI.Helpers.Boxes.Box(textBox._historyIndex < textBox._history.Count - 1);
-		}
-#nullable restore
-
-		[GeneratedDependencyProperty(DefaultValue = false)]
-		public static DependencyProperty CanPasteClipboardContentProperty { get; } = CreateCanPasteClipboardContentProperty();
-
-		public bool CanPasteClipboardContent
-		{
-			get => GetCanPasteClipboardContentValue();
-			private set => SetCanPasteClipboardContentValue(value);
-		}
-
-		public static DependencyProperty SelectionFlyoutProperty { get; } =
-			DependencyProperty.Register(
-				nameof(SelectionFlyout), typeof(FlyoutBase), typeof(TextBox),
-				new FrameworkPropertyMetadata(default(FlyoutBase), FrameworkPropertyMetadataOptions.ValueDoesNotInheritDataContext));
-
-		public FlyoutBase SelectionFlyout
-		{
-			get => (FlyoutBase)GetValue(SelectionFlyoutProperty);
-			set => SetValue(SelectionFlyoutProperty, value);
-		}
-
-		public static DependencyProperty ProofingMenuFlyoutProperty { get; } =
-			DependencyProperty.Register(
-				nameof(ProofingMenuFlyout), typeof(FlyoutBase), typeof(TextBox),
-				new FrameworkPropertyMetadata(default(FlyoutBase), FrameworkPropertyMetadataOptions.ValueDoesNotInheritDataContext));
-
-		// Ported from: TextBoxBase.cpp GetProofingMenuFlyoutNoRef (lines 5507-5520)
-		public FlyoutBase ProofingMenuFlyout
-		{
-			get
-			{
-				EnsureProofingMenu();
-				TextControlFlyoutHelper.AddProofingFlyout(_proofingMenu, this);
-				if (IsSpellCheckEnabled && (FocusState != FocusState.Unfocused || _forceFocusedVisualState))
-				{
-					UpdateProofingMenu();
-				}
-				SetValue(ProofingMenuFlyoutProperty, _proofingMenu);
-				return _proofingMenu;
-			}
-		}
-
-		private void UpdateCanPasteClipboardContent()
-		{
-			if (IsReadOnly)
-			{
-				CanPasteClipboardContent = false;
-				return;
-			}
-
-			try
-			{
-				var content = Clipboard.GetContent();
-				CanPasteClipboardContent = content?.Contains(StandardDataFormats.Text) ?? false;
-			}
-			catch
-			{
-				CanPasteClipboardContent = false;
-			}
-		}
-
-		// Ported from: TextBoxBase.cpp EnsureProofingMenu (line 5470)
-		private void EnsureProofingMenu()
-		{
-			_proofingMenu ??= new MenuFlyout();
-		}
-
-		// Ported from: TextBoxBase.cpp UpdateProofingMenu (line 5536)
-		private void UpdateProofingMenu()
-		{
-			_proofingMenu.Items.Clear();
-
-			if (TextBoxView?.DisplayBlock?.ParsedText is not UnicodeText unicodeText)
-			{
-				return;
-			}
-
-			// Find correction at cursor position
-			var correction = unicodeText.GetCorrectionAtIndex(SelectionStart);
-			if (correction is null)
-			{
-				return;
-			}
-
-			var (correctionStart, correctionEnd) = correction.Value;
-
-			// Get suggestions (matches WinUI AddSpellingSuggestions, max 3 suggestions)
-			var suggestionsResult = unicodeText.GetSpellCheckSuggestions(correctionStart, correctionEnd);
-			if (suggestionsResult is not { } result || result.suggestions.Count == 0)
-			{
-				return;
-			}
-
-			var (replaceStart, replaceEnd, suggestions) = result;
-			var maxSuggestions = Math.Min(suggestions.Count, 3); // WinUI uses maxSuggestionCount = 3
-
-			for (var i = 0; i < maxSuggestions; i++)
-			{
-				var suggestion = suggestions[i];
-				var item = new MenuFlyoutItem { Text = suggestion };
-				item.Click += (_, _) => ReplaceWithSuggestion(replaceStart, replaceEnd, suggestion);
-				_proofingMenu.Items.Add(item);
-			}
-		}
-
-		// Ported from: TextBoxBase.cpp pattern for applying text replacement (similar to PasteFromClipboardPartial)
-		private void ReplaceWithSuggestion(int replaceStart, int replaceEnd, string suggestion)
-		{
-			if (IsReadOnly)
-			{
-				return;
-			}
-
-			TrySetCurrentlyTyping(false);
-
-			var oldText = Text;
-			var newText = oldText.Remove(replaceStart, replaceEnd - replaceStart).Insert(replaceStart, suggestion);
-
-			CommitAction(new ReplaceAction(oldText, newText, replaceStart + suggestion.Length));
-
-			_clearHistoryOnTextChanged = false;
-			_pendingSelection = (replaceStart + suggestion.Length, 0);
-			ProcessTextInput(newText);
-			_clearHistoryOnTextChanged = true;
-		}
-
-		private void OnClipboardContentChanged(object sender, object e)
-		{
-			if (DispatcherQueue.HasThreadAccess)
-			{
-				UpdateCanPasteClipboardContent();
-			}
-			else
-			{
-				DispatcherQueue.TryEnqueue(() => UpdateCanPasteClipboardContent());
-			}
-		}
-
-		private void TrySetCurrentlyTyping(bool newValue)
-		{
-			if (newValue == _currentlyTyping || _suppressCurrentlyTyping)
-			{
-				return;
-			}
-
-			if (newValue)
-			{
-				_textWhenTypingStarted = Text;
-				_selectionWhenTypingStarted = (
-					_selection.start,
-					_selection.length,
-					_selection.selectionEndsAtTheStart);
-			}
-			else
-			{
-				_historyIndex++;
-				_history.RemoveAllAt(_historyIndex);
-				_history.Add(new HistoryRecord(
-					new ReplaceAction(_textWhenTypingStarted, Text, _selection.start),
-					_selectionWhenTypingStarted.start,
-					_selectionWhenTypingStarted.length,
-					_selectionWhenTypingStarted.selectionEndsAtTheStart));
-			}
-
-			_currentlyTyping = newValue;
-		}
-
-		partial void OnUnloadedPartial()
-		{
-			_forceFocusedVisualState = false;
-			_timer.Stop();
-			_gripperPresenter?.Hide();
-			CaretMode = CaretDisplayMode.ThumblessCaretHidden;
-			_clipboardChangeSubscription.Disposable = null;
-		}
-
-		partial void OnIsReadonlyChangedPartial() => UpdateCanPasteClipboardContent();
-
-		partial void OnForegroundColorChangedPartial(Brush newValue)
-		{
-			TextBoxView?.OnForegroundChanged(newValue);
-			// Update the caret brush to match the new foreground color.
-			UpdateDisplaySelection();
-		}
-
-		partial void OnSelectionHighlightColorChangedPartial(SolidColorBrush brush) => TextBoxView?.OnSelectionHighlightColorChanged(brush);
-
-		partial void UpdateFontPartial()
-		{
-			TextBoxView?.UpdateFont();
-		}
-
-		partial void OnInputScopeChangedPartial(InputScope newValue) => TextBoxView?.UpdateProperties();
-
-		partial void OnIsSpellCheckEnabledChangedPartial(bool newValue)
-		{
-			if (TextBoxView is not null)
-			{
-				TextBoxView.DisplayBlock.IsSpellCheckEnabled = newValue;
-				TextBoxView.UpdateProperties();
-			}
-		}
-
-		partial void OnIsTextPredictionEnabledChangedPartial(bool newValue) => TextBoxView?.UpdateProperties();
-
-		partial void OnMaxLengthChangedPartial(int newValue) => TextBoxView?.UpdateMaxLength();
-
-		partial void OnFlowDirectionChangedPartial()
-		{
-			TextBoxView?.SetFlowDirection();
-		}
-
-		partial void OnTextWrappingChangedPartial()
-		{
-			TextBoxView?.SetWrapping();
-			if (_contentElement is ScrollViewer sv)
-			{
-				// This is to work around sv giving infinite width. This has the unfortunate problem of resetting
-				// locally-set values and/or changes in the template.
-				sv.HorizontalScrollBarVisibility = TextWrapping == TextWrapping.NoWrap ? ScrollBarVisibility.Hidden : ScrollBarVisibility.Disabled;
-			}
-		}
-
-		partial void OnTextAlignmentChangedPartial(TextAlignment newValue)
-		{
-			TextBoxView?.SetTextAlignment();
-		}
-
-		partial void SetInputReturnTypePlatform(InputReturnType inputReturnType)
-		{
-			TextBoxView?.UpdateProperties();
-		}
-
-		private void UpdateTextBoxView()
-		{
-			_textBoxView ??= new TextBoxView(this);
-			if (ContentElement != null)
-			{
-				var displayBlock = TextBoxView.DisplayBlock;
-				if (ContentElement.Content != displayBlock)
-				{
-					ContentElement.Content = displayBlock;
-
-					// The presenter owns the touch-selection gripper visuals and their drag/positioning
-					// logic (shared with selectable TextBlock). It subscribes to DisplayBlock.DrawingFinished
-					// to keep the grippers glued to the selection ends as the text is (re)drawn.
-					_gripperPresenter ??= new TextSelectionGripperPresenter(this);
-				}
-
-				TextBoxView.SetTextNative(Text);
-			}
-		}
-
-		partial void OnFocusStateChangedPartial(FocusState focusState, bool initial)
-		{
-			_clipboardChangeSubscription.Disposable = null;
-			TextBoxView?.OnFocusStateChanged(focusState);
-
-			if (focusState != FocusState.Unfocused)
-			{
-				CaretMode = CaretDisplayMode.ThumblessCaretShowing;
-				_textBoxNotificationsSingleton?.OnFocused(this);
-				StartImeSession();
-				UpdateCanPasteClipboardContent();
-				Clipboard.ContentChanged += OnClipboardContentChanged;
-				_clipboardChangeSubscription.Disposable = Disposable.Create(() => Clipboard.ContentChanged -= OnClipboardContentChanged);
-			}
-			else
-			{
-				// Ported from: TextBoxBase.cpp UpdateFocusState (lines 4996-5008)
-				// Check if focus is moving to a text control flyout before deciding to unfocus.
-				_forceFocusedVisualState = ShouldForceFocusedVisualState();
-
-				// Ported from: TextBoxBase.cpp UpdateFocusState (lines 5009-5016)
-				// Hide touch caret thumbs when context flyout is opening.
-				if (_forceFocusedVisualState && ShouldHideGrippersOnFlyoutOpening()
-					&& CaretMode is CaretDisplayMode.CaretWithThumbsOnlyEndShowing
-						or CaretDisplayMode.CaretWithThumbsBothEndsShowing)
-				{
-					CaretMode = CaretDisplayMode.ThumblessCaretShowing;
-				}
-			}
-
-			if (focusState == FocusState.Unfocused && !_forceFocusedVisualState)
-			{
-				EndImeSession();
-				TrySetCurrentlyTyping(false);
-				CaretMode = CaretDisplayMode.ThumblessCaretHidden;
-				if (SelectionFlyout?.IsOpen == true)
-				{
-					SelectionFlyout.Hide();
-				}
-				if (!initial)
-				{
-					_textBoxNotificationsSingleton?.OnUnfocused(this);
-				}
-				_timer.Stop();
-			}
-			UpdateDisplaySelection();
-		}
-
-		// Ported from: TextBoxBase.cpp ShouldForceFocusedVisualState (lines 5286-5290)
-		private bool ShouldForceFocusedVisualState()
-		{
-			return TextControlFlyoutHelper.IsGettingFocus(SelectionFlyout, this)
-				|| TextControlFlyoutHelper.IsGettingFocus(ContextFlyout, this);
-		}
-
-		// Ported from: TextBoxBase.cpp ShouldHideGrippersOnFlyoutOpening (lines 5291-5294)
-		private bool ShouldHideGrippersOnFlyoutOpening()
-		{
-			return TextControlFlyoutHelper.IsGettingFocus(ContextFlyout, this);
-		}
-
-		// Ported from: TextBoxBase.cpp DismissAllFlyouts (lines 5566-5574)
-		internal void DismissAllFlyouts()
-		{
-			TextControlFlyoutHelper.CloseIfOpen(_proofingMenu);
-			TextControlFlyoutHelper.CloseIfOpen(SelectionFlyout);
-			TextControlFlyoutHelper.CloseIfOpen(ContextFlyout);
-		}
-
-		// Ported from: TextBoxBase.cpp ForceFocusLoss (lines 5551-5560)
-		internal void ForceFocusLoss()
-		{
-			_forceFocusedVisualState = false;
-			// Trigger the unfocus behavior that was deferred
-			OnFocusStateChangedPartial(FocusState.Unfocused, initial: false);
-			UpdateVisualState();
-		}
-
-#if false // Removing temporarily. We'll need to add it back.
-		// TODO: Discuss this public API.
-		public static void FinishAutofillContext(bool shouldSave)
-		{
-			_textBoxNotificationsSingleton?.FinishAutofillContext(shouldSave);
-		}
-#endif
-
-		partial void SelectPartial(int start, int length)
-		{
-			TrySetCurrentlyTyping(false);
-
-			if (!_inSelectInternal)
-			{
-				// SelectInternal sets _selectionEndsAtTheStart and _caretXOffset on its own
-				_selection.selectionEndsAtTheStart = false;
-				_caretXOffset = (float)(TextBoxView?.DisplayBlock.ParsedText.GetRectForIndex(start + length).Left ?? 0);
-			}
-
-			var selection = (start, length, _selection.selectionEndsAtTheStart);
-			var selectionChanged = selection != _selection;
-			_selection = selection;
-
-			// Even when using Skia TextBox, we may need to call Select,
-			// which will update the native input in case of Wasm Skia for example.
-			TextBoxView?.Select(start, length);
-
-			if (length == 0 && CaretMode == CaretDisplayMode.CaretWithThumbsBothEndsShowing)
-			{
-				// It doesn't make sense to have 2 caret ends when there's no selection.
-				CaretMode = CaretDisplayMode.CaretWithThumbsOnlyEndShowing;
-			}
-			else if (CaretMode is CaretDisplayMode.ThumblessCaretHidden)
-			{
-				CaretMode = CaretDisplayMode.ThumblessCaretShowing;
-			}
-			else if (CaretMode is CaretDisplayMode.ThumblessCaretShowing)
-			{
-				_timer.Start(); // restart
-			}
-
-			if (selectionChanged)
-			{
-				UpdateScrolling();
-			}
-			UpdateDisplaySelection();
-		}
-
-		partial void SelectAllPartial() => Select(0, Text.Length);
-
-		public int SelectionStart
-		{
-			get => _selection.start;
-			set => Select(start: value, length: SelectionLength);
-		}
-
-		public int SelectionLength
-		{
-			get => _selection.length;
-			set => Select(SelectionStart, value);
-		}
-
-		private void UpdateDisplaySelection()
-		{
-			if (TextBoxView?.DisplayBlock is { } displayBlock)
-			{
-				displayBlock.Selection = new TextBlock.Range(SelectionStart, SelectionStart + SelectionLength);
-				var isFocused = FocusState != FocusState.Unfocused || _forceFocusedVisualState;
-				displayBlock.RenderSelection = isFocused;
-				if (CaretMode is CaretDisplayMode.ThumblessCaretShowing &&
-					SelectionLength == 0 &&
-					isFocused &&
-					!IsReadOnly &&
-					!FeatureConfiguration.TextBox.HideCaret)
-				{
-					// WinUI uses DestInvert composite mode with a white brush so the caret
-					// is always visible regardless of theme. Uno doesn't support DestInvert,
-					// so we use the TextBox's own Foreground (element-theme-aware via
-					// ThemeResource) with fully opaque alpha for maximum caret visibility.
-					var caretBrush = GetOpaqueCaretBrush();
-					displayBlock.RenderCaret = (IsBackwardSelection ? SelectionStart : SelectionStart + SelectionLength, caretBrush);
-				}
-				else
-				{
-					displayBlock.RenderCaret = null;
-				}
-				((IBlock)TextBoxView.DisplayBlock).Invalidate(false);
-
-				var currentSelection = (SelectionStart, SelectionLength, IsBackwardSelection);
-				if (currentSelection != _lastNotifiedSelection)
-				{
-					_lastNotifiedSelection = currentSelection;
-					_textBoxNotificationsSingleton?.NotifySelectionChanged(this);
-				}
-			}
-		}
-
-		/// <summary>
-		/// Gets a fully opaque composition brush derived from the TextBox's Foreground for caret rendering.
-		/// WinUI uses DestInvert composite mode (white brush + invert), but Uno doesn't support that.
-		/// Instead, we use the element-theme-aware Foreground with full opacity for maximum visibility.
-		/// </summary>
-		private CompositionBrush GetOpaqueCaretBrush()
-		{
-			var compositor = Compositor.GetSharedCompositor();
-			if (Foreground is SolidColorBrush scb)
-			{
-				var color = scb.Color;
-				if (color.A < 255)
-				{
-					// Force fully opaque for caret visibility
-					color = Color.FromArgb(255, color.R, color.G, color.B);
-				}
-
-				if (_cachedCaretBrush is not null && _cachedCaretColor == color)
-				{
-					return _cachedCaretBrush;
-				}
-
-				_cachedCaretColor = color;
-				_cachedCaretBrush = compositor.CreateColorBrush(color);
-				return _cachedCaretBrush;
-			}
-
-			_cachedCaretBrush = null;
-			_cachedCaretColor = default;
-			return DefaultBrushes.TextForegroundBrush.GetOrCreateCompositionBrush(compositor);
-		}
-
-		private void UpdateScrolling() => UpdateScrolling(true);
-
-		/// <summary>
-		/// Gets the position where a context menu should be shown for keyboard invocation.
-		/// Returns the position at the selection/caret location.
-		/// </summary>
-		/// <returns>The position in local coordinates, or null if position cannot be determined.</returns>
-		internal Point? GetContextMenuShowPosition()
-		{
-			if (TextBoxView?.DisplayBlock?.ParsedText == null)
-			{
-				return null;
-			}
-
-			// Determine the character index to use for positioning
-			// Use selection end for LTR (selection start for RTL to match WinUI behavior)
-			var index = FlowDirection == FlowDirection.RightToLeft && SelectionLength > 0
-				? SelectionStart
-				: SelectionStart + SelectionLength;
-
-			// Get the rect for the character at the position
-			var rect = TextBoxView.DisplayBlock.ParsedText.GetRectForIndex(index);
-
-			// Transform from DisplayBlock coordinates to TextBox coordinates
-			var transform = TextBoxView.DisplayBlock.TransformToVisual(this);
-			var point = transform.TransformPoint(new Point(rect.Left, rect.Bottom));
-
-			return point;
-		}
-
-		#region SelectionFlyout Support
-
-		// Ported from: microsoft-ui-xaml2/src/dxaml/xcp/core/native/text/Controls/TextBoxBase.h (lines 225-226)
-		// m_lastInputDeviceType, m_lastPointerPositionForFlyout, m_isSelectionFlyoutUpdateQueued
-		private PointerDeviceType _lastInputDeviceType;
-		private Point _lastPointerPosition;
-		private bool _isSelectionFlyoutUpdateQueued;
-		// Uno addition (not part of the WinUI port): lets the flyout re-open over a collapsed caret when the single
-		// insertion handle is tapped. See QueueGripperSelectionFlyout / OnGripperTapped.
-		private bool _allowEmptySelectionFlyout;
-
-		// Ported from: microsoft-ui-xaml2/src/dxaml/xcp/core/native/text/Controls/TextBoxBase.cpp (lines 5292-5302)
-		private bool HasSelectionFlyout() => SelectionFlyout is not null;
-
-		// Ported from: microsoft-ui-xaml2/src/dxaml/xcp/core/native/text/Controls/TextBoxBase.cpp (lines 5349-5377)
-		// QueueUpdateSelectionFlyoutVisibility - queues async visibility update
-		private void QueueUpdateSelectionFlyoutVisibility(PointerDeviceType deviceType, Point position, bool allowEmptySelection = false)
-		{
-			_lastInputDeviceType = deviceType;
-			_lastPointerPosition = position;
-			_allowEmptySelectionFlyout = allowEmptySelection;
-
-			// Line 5358-5360: Prevent duplicate queued updates
-			if (!_isSelectionFlyoutUpdateQueued)
-			{
-				_isSelectionFlyoutUpdateQueued = true;
-				DispatcherQueue.TryEnqueue(() => UpdateSelectionFlyoutVisibility());
-			}
-		}
-
-		// Ported from: microsoft-ui-xaml2/src/dxaml/xcp/core/native/text/Controls/TextBoxBase.cpp (lines 5379-5452)
-		// UpdateSelectionFlyoutVisibility - shows/hides SelectionFlyout based on device type and selection
-		private void UpdateSelectionFlyoutVisibility()
-		{
-			// Line 5381: Reset the queued flag
-			_isSelectionFlyoutUpdateQueued = false;
-
-			// Line 5383: Only shows SelectionFlyout if ContextFlyout is NOT open
-			if (!HasSelectionFlyout() || TextControlFlyoutHelper.IsOpen(ContextFlyout))
-			{
-				return;
-			}
-
-			var selectionLength = SelectionLength;
-			var showMode = FlyoutShowMode.Transient;
-			var shouldShow = false;
-
-			switch (_lastInputDeviceType)
-			{
-				case PointerDeviceType.Mouse:
-					// Line 5389-5401: Mouse input - only for RichEditBox, NOT TextBox
-					shouldShow = false;
-					break;
-
-				case PointerDeviceType.Pen:
-				case PointerDeviceType.Touch:
-					// Line 5403-5411: Pen/Touch - show if selection exists.
-					// Uno addition (beyond the WinUI port): also show over a collapsed caret when the single insertion
-					// handle was tapped, matching the native iOS/Android Paste/Select-all popup (gated to those
-					// conventions by the caller).
-					if (selectionLength > 0 || _allowEmptySelectionFlyout)
-					{
-						shouldShow = true;
-						showMode = FlyoutShowMode.Transient;
-					}
-					break;
-
-				default:
-					// Line 5413-5415: Keyboard/other - never show SelectionFlyout
-					shouldShow = false;
-					break;
-			}
-
-			if (shouldShow)
-			{
-				// Line 5424-5441: Get selection bounds and adjust flyout position
-				var position = _lastPointerPosition;
-
-				if (TextBoxView?.DisplayBlock?.ParsedText is { } parsedText)
-				{
-					// Get selection bounding rect in DisplayBlock coordinates
-					var startRect = parsedText.GetRectForIndex(SelectionStart);
-					var endRect = parsedText.GetRectForIndex(SelectionStart + SelectionLength);
-
-					// Compute union of start and end rects for full selection bounds
-					var selectionTop = Math.Min(startRect.Top, endRect.Top);
-
-					// Transform from DisplayBlock coordinates to TextBox coordinates
-					var transform = TextBoxView.DisplayBlock.TransformToVisual(this);
-					var transformedTop = transform.TransformPoint(new Point(0, selectionTop));
-
-					// Set Y position to top of selection for consistent placement
-					// (WinUI: "We want the point to appear at a consistent location regardless of
-					// whether it overlaps the selection, so we'll set its vertical position to be
-					// the top of the exclusion rectangle.")
-					position = new Point(position.X, transformedTop.Y);
-				}
-
-				if (SelectionFlyout is { } selectionFlyout)
-				{
-					TextControlFlyoutHelper.ShowAt(selectionFlyout, this, position, showMode);
-				}
-			}
-			else
-			{
-				// Line 5444-5447: Close SelectionFlyout if it's open and we shouldn't show it
-				if (SelectionFlyout?.IsOpen == true)
-				{
-					SelectionFlyout.Hide();
-				}
-			}
-
-			// Line 5450: Reset input device type after processing
-			_lastInputDeviceType = default;
-			_allowEmptySelectionFlyout = default;
-		}
-
-		#endregion
-
-		/// <summary>
-		/// Narrows the BringIntoView target rect to the caret position so that parent
-		/// ScrollViewers scroll to the cursor rather than the full TextBox bounds.
-		/// </summary>
-		/// <remarks>
-		/// When a TextBox has its internal scrolling disabled and relies on an outer
-		/// ScrollViewer, the default BringIntoView uses the full element bounds.
-		/// For a tall TextBox this causes the outer ScrollViewer to scroll to the
-		/// element edges rather than to the cursor position.
-		/// </remarks>
-		protected override void OnBringIntoViewRequested(BringIntoViewRequestedEventArgs e)
-		{
-			base.OnBringIntoViewRequested(e);
-
-			if ((e.TargetElement is null || e.TargetElement == this)
-				&& FocusState != FocusState.Unfocused
-				&& _contentElement is ScrollViewer { VerticalScrollMode: ScrollMode.Disabled }
-				&& TextBoxView?.DisplayBlock is { } displayBlock
-				&& displayBlock.ParsedText is { } parsedText)
-			{
-				var selectionEnd = _selection.selectionEndsAtTheStart
-					? _selection.start
-					: _selection.start + _selection.length;
-
-				var caretRect = parsedText.GetRectForIndex(selectionEnd);
-				caretRect = caretRect with { Width = Math.Max(caretRect.Width, TextBlock.CaretThickness) };
-
-				// Transform from TextBlock coordinate space to TextBox coordinate space
-				var transform = displayBlock.TransformToVisual(this);
-				var caretRectInTextBox = transform.TransformBounds(caretRect);
-
-				e.TargetRect = caretRectInTextBox;
-			}
-		}
-
-		/// <summary>
-		/// Scrolls the <see cref="_contentElement"/> so that the caret is inside the visible viewport
-		/// </summary>
-		/// <remarks>
-		/// By default, only the selection end moves, while the selection start stays fixed. This is not the
-		/// case when dragging the caret thumb, in which case both ends can move. This case requires an
-		/// explicit call to this method with <see cref="putSelectionEndInVisibleViewport"/> = false.
-		/// </remarks>>
-		private void UpdateScrolling(bool putSelectionEndInVisibleViewport)
-		{
-			if (_contentElement is ScrollViewer sv)
-			{
-				var horizontalOffset = sv.HorizontalOffset;
-				var verticalOffset = sv.VerticalOffset;
-
-				var (selectionStart, selectionEnd) = _selection.selectionEndsAtTheStart ? (_selection.start + _selection.length, _selection.start) : (_selection.start, _selection.start + _selection.length);
-				var index = putSelectionEndInVisibleViewport ? selectionEnd : selectionStart;
-
-				var caretRect = TextBoxView.DisplayBlock.ParsedText.GetRectForIndex(index) with { Width = TextBlock.CaretThickness };
-
-				// Because the caret is only a single-pixel wide, and because screens can't draw in fractions of a pixel,
-				// we need to add Math.Ceiling to ensure that the caret is (fully) included in the visible viewport. This
-				// Math.Ceiling sometimes horizontal overscrolling, but it's more acceptable than sometimes not showing the caret.
-				var newHorizontalOffset = horizontalOffset.AtMost(caretRect.Left).AtLeast(Math.Ceiling(caretRect.Right - sv.ViewportWidth + TextBlock.CaretThickness));
-
-				var newVerticalOffset = verticalOffset.AtMost(caretRect.Top).AtLeast(caretRect.Bottom - sv.ViewportHeight);
-
-				sv.ChangeView(newHorizontalOffset, newVerticalOffset, null);
-			}
-
-			// When the TextBox's internal scrolling is disabled, an outer ScrollViewer
-			// handles scrolling. Notify it so the caret stays visible after text changes.
-			if (FocusState != FocusState.Unfocused
-				&& _contentElement is ScrollViewer { VerticalScrollMode: ScrollMode.Disabled })
-			{
-				StartBringIntoView(new BringIntoViewOptions { AnimationDesired = false });
-			}
-		}
-
-		partial void OnKeyDownPartial(KeyRoutedEventArgs args)
-		{
-			// This is a minimal copy of OnkeyDownSkia that just sets args.Handled without doing any work.
-			// This is to match WinUI behavior where Handled is set for certain keys before public
-			// subscribers get the event, but before any actual text processing is done.
-			var (selectionStart, selectionLength) = _selection.selectionEndsAtTheStart ? (_selection.start + _selection.length, -_selection.length) : (_selection.start, _selection.length);
-			var text = Text;
-			var shift = args.KeyboardModifiers.HasFlag(VirtualKeyModifiers.Shift);
-			var ctrl = args.KeyboardModifiers.HasFlag(_platformCtrlKey);
-			switch (args.Key)
-			{
-				case VirtualKey.Escape:
-					if (HasPointerCapture)
-					{
-						args.Handled = true;
-					}
-					return;
-				case VirtualKey.Z when ctrl:
-				case VirtualKey.Y when ctrl:
-				case VirtualKey.Delete when !IsReadOnly:
-				case VirtualKey.A when ctrl:
-					if (!HasPointerCapture)
-					{
-						args.Handled = true;
-					}
-					return;
-				case VirtualKey.Up:
-					// on Apple platforms start of document is `Command` and `Up`
-					if (ctrl && DeviceTargetHelper.UsesAppleKeyboardLayout)
-					{
-						KeyDownHome(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
-					}
-					else
-					{
-						KeyDownUpArrow(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
-					}
-					break;
-				case VirtualKey.Down:
-					// on Apple platforms end of document is `Command` and `Down`
-					if (ctrl && DeviceTargetHelper.UsesAppleKeyboardLayout)
-					{
-						KeyDownEnd(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
-					}
-					else
-					{
-						KeyDownDownArrow(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
-					}
-					break;
-				case VirtualKey.Left when !TextBoxView.DisplayBlock.ParsedText.IsBaseDirectionRightToLeft:
-				case VirtualKey.Right when TextBoxView.DisplayBlock.ParsedText.IsBaseDirectionRightToLeft:
-					KeyDownLeftArrow(args, text, shift, ctrl, ref selectionStart, ref selectionLength);
-					break;
-				case VirtualKey.Left when TextBoxView.DisplayBlock.ParsedText.IsBaseDirectionRightToLeft:
-				case VirtualKey.Right when !TextBoxView.DisplayBlock.ParsedText.IsBaseDirectionRightToLeft:
-					KeyDownRightArrow(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
-					break;
-				case VirtualKey.Home:
-					KeyDownHome(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
-					break;
-				case VirtualKey.End:
-					KeyDownEnd(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
-					break;
-				// TODO: PageUp/Down
-				case VirtualKey.Back when !IsReadOnly:
-					KeyDownBack(args, ref text, ctrl, shift, ref selectionStart, ref selectionLength);
-					break;
-			}
-		}
-
-		private void OnKeyDownSkia(KeyRoutedEventArgs args)
-		{
-			if (SelectionFlyout?.IsOpen == true)
-			{
-				SelectionFlyout.Hide();
-			}
-
-			if (_selection.length != 0 &&
-				args.Key is not (VirtualKey.Up or VirtualKey.Down or VirtualKey.Left or VirtualKey.Right))
-			{
-				// On WinUI, pressing anything except arrow keys will immediately make the caret thumbless.
-				// Even shift + arrow keys will make the caret thumbless (because it's a shift _then_ an arrow key).
-				CaretMode = CaretDisplayMode.ThumblessCaretShowing;
-			}
-
-			// Note: On windows ** only KeyDown ** is handled (not KeyUp)
-
-			// move to possibly-negative selection length format
-			var (selectionStart, selectionLength) = _selection.selectionEndsAtTheStart ? (_selection.start + _selection.length, -_selection.length) : (_selection.start, _selection.length);
-
-			var text = Text;
-			var shift = args.KeyboardModifiers.HasFlag(VirtualKeyModifiers.Shift);
-			var ctrl = args.KeyboardModifiers.HasFlag(_platformCtrlKey);
-			// Text commands: always return from this switch, never break
-			switch (args.Key)
-			{
-				case VirtualKey.Z when ctrl:
-					if (!HasPointerCapture)
-					{
-						args.Handled = true;
-						Undo();
-					}
-					return;
-				case VirtualKey.Y when ctrl:
-					if (!HasPointerCapture)
-					{
-						args.Handled = true;
-						Redo();
-					}
-					return;
-				case VirtualKey.X when ctrl:
-					CutSelectionToClipboard();
-					return;
-				case VirtualKey.V when ctrl:
-				case VirtualKey.Insert when shift:
-					PasteFromClipboard(); // async so doesn't actually do anything right now
-					return;
-				case VirtualKey.C when ctrl:
-				case VirtualKey.Insert when ctrl:
-					CopySelectionToClipboard();
-					return;
-				case VirtualKey.Escape:
-					if (HasPointerCapture)
-					{
-						args.Handled = true;
-						ReleasePointerCaptures();
-					}
-					return;
-				case VirtualKey.LeftShift:
-				case VirtualKey.RightShift:
-				case VirtualKey.Shift:
-				case VirtualKey.Control:
-				case VirtualKey.LeftControl:
-				case VirtualKey.RightControl:
-				case VirtualKey.Menu:
-				case VirtualKey.LeftMenu:
-				case VirtualKey.RightMenu:
-				case VirtualKey.LeftWindows:
-				case VirtualKey.RightWindows:
-					// No-op when pressing these key specifically.
-					return;
-			}
-
-			// text input/movement
-			switch (args.Key)
-			{
-				case VirtualKey.Up:
-					// on Apple platforms start of document is `Command` and `Up`
-					if (ctrl && DeviceTargetHelper.UsesAppleKeyboardLayout)
-					{
-						KeyDownHome(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
-					}
-					else
-					{
-						KeyDownUpArrow(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
-					}
-					break;
-				case VirtualKey.Down:
-					// on Apple platforms end of document is `Command` and `Down`
-					if (ctrl && DeviceTargetHelper.UsesAppleKeyboardLayout)
-					{
-						KeyDownEnd(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
-					}
-					else
-					{
-						KeyDownDownArrow(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
-					}
-					break;
-				case VirtualKey.Left when !TextBoxView.DisplayBlock.ParsedText.IsBaseDirectionRightToLeft:
-				case VirtualKey.Right when TextBoxView.DisplayBlock.ParsedText.IsBaseDirectionRightToLeft:
-					KeyDownLeftArrow(args, text, shift, ctrl, ref selectionStart, ref selectionLength);
-					break;
-				case VirtualKey.Left when TextBoxView.DisplayBlock.ParsedText.IsBaseDirectionRightToLeft:
-				case VirtualKey.Right when !TextBoxView.DisplayBlock.ParsedText.IsBaseDirectionRightToLeft:
-					KeyDownRightArrow(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
-					break;
-				case VirtualKey.Home:
-					KeyDownHome(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
-					break;
-				case VirtualKey.End:
-					KeyDownEnd(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
-					break;
-				// TODO: PageUp/Down
-				case VirtualKey.Back when !IsReadOnly:
-					KeyDownBack(args, ref text, ctrl, shift, ref selectionStart, ref selectionLength);
-					break;
-				case VirtualKey.Delete when !IsReadOnly:
-					KeyDownDelete(args, ref text, ctrl, shift, ref selectionStart, ref selectionLength);
-					break;
-				case VirtualKey.A when ctrl:
-					if (!HasPointerCapture)
-					{
-						args.Handled = true;
-						TrySetCurrentlyTyping(false);
-						selectionStart = 0;
-						selectionLength = text.Length;
-					}
-					break;
-				default:
-					// During IME composition, skip normal character insertion.
-					// The IME extension handles text updates via OnImeCompositionUpdated → ProcessTextInput.
-					// On platforms where the platform applies text directly (e.g., Android),
-					// key events are independent of composition events and should not be swallowed.
-					if (ShouldSwallowKeyDuringComposition)
-					{
-						return;
-					}
-					var isEnterKey = args.UnicodeKey is '\r' or '\n' || args.Key == VirtualKey.Enter;
-					// Don't insert characters when shortcut modifier keys are held down.
-					// On non-Apple platforms, AltGr (Ctrl+Alt together) produces valid characters (like @, €),
-					// so we must not suppress those. On Apple, Option (Menu/Alt) produces special characters.
-					var altHeld = args.KeyboardModifiers.HasFlag(VirtualKeyModifiers.Menu);
-					var ctrlHeld = args.KeyboardModifiers.HasFlag(VirtualKeyModifiers.Control);
-					var isAltGr = !DeviceTargetHelper.UsesAppleKeyboardLayout && ctrlHeld && altHeld;
-					var hasShortcutModifier = !isAltGr && (
-						ctrlHeld ||
-						args.KeyboardModifiers.HasFlag(VirtualKeyModifiers.Windows) ||
-						(!DeviceTargetHelper.UsesAppleKeyboardLayout && altHeld));
-					if (!IsReadOnly && !HasPointerCapture && !hasShortcutModifier && args.UnicodeKey is { } key && (!isEnterKey || AcceptsReturn))
-					{
-						TrySetCurrentlyTyping(true);
-						var start = Math.Min(selectionStart, selectionStart + selectionLength);
-						var end = Math.Max(selectionStart, selectionStart + selectionLength);
-
-						if (key is '\n')
-						{
-							// TextBox autoconverts to \r, like WinUI
-							key = '\r';
-						}
-
-						text = text[..start] + key + text[end..];
-						selectionStart = start + 1;
-						selectionLength = 0;
-						break;
-					}
-					else
-					{
-						return;
-					}
-			}
-
-			selectionStart = Math.Max(0, Math.Min(text.Length, selectionStart));
-			selectionLength = Math.Max(-selectionStart, Math.Min(text.Length - selectionStart, selectionLength));
-
-			var caretXOffset = _caretXOffset;
-
-			_suppressCurrentlyTyping = true;
-			_clearHistoryOnTextChanged = false;
-			if (!HasPointerCapture)
-			{
-				_pendingSelection = (selectionStart, selectionLength);
-			}
-
-			ProcessTextInput(text);
-			_clearHistoryOnTextChanged = true;
-			_suppressCurrentlyTyping = false;
-
-			// don't change the caret offset when moving up and down
-			if (args.Key is VirtualKey.Up or VirtualKey.Down)
-			{
-				// this condition is accurate in the case of hitting Down on the last line
-				// or up on the first line. On WinUI, the caret offset won't change.
-				_caretXOffset = caretXOffset;
-			}
-		}
-
-		internal void SetPendingSelection(int selectionStart, int selectionLength)
-			=> _pendingSelection = (selectionStart, selectionLength);
-
-		private void KeyDownBack(KeyRoutedEventArgs args, ref string text, bool ctrl, bool shift, ref int selectionStart, ref int selectionLength)
-		{
-			// on Apple platforms it is `option` + `delete` (same location as backspace on PC keyboards) that removes the previous word
-			if (DeviceTargetHelper.UsesAppleKeyboardLayout)
-			{
-				ctrl = args.KeyboardModifiers.HasFlag(VirtualKeyModifiers.Menu);
-			}
-
-			if (HasPointerCapture)
-			{
-				return;
-			}
-			if (selectionLength != 0)
-			{
-				TrySetCurrentlyTyping(false);
-				TrySetCurrentlyTyping(true);
-
-				var start = Math.Min(selectionStart, selectionStart + selectionLength);
-				var end = Math.Max(selectionStart, selectionStart + selectionLength);
-				text = text[..start] + text[end..];
-				selectionLength = 0;
-				selectionStart = start;
-			}
-			else if (selectionStart != 0)
-			{
-				if (ctrl)
-				{
-					// ctrl always ends the previous typing run
-					TrySetCurrentlyTyping(false);
-				}
-				else
-				{
-					// idempotent call to make sure we're starting a new typing run if we're not in one already
-					TrySetCurrentlyTyping(true);
-				}
-
-				var oldText = text;
-				var index = ctrl ? TextBoxView.DisplayBlock.ParsedText.GetWordAt(selectionStart, false).start : selectionStart - 1;
-				text = text[..index] + text[selectionStart..];
-				selectionStart = index;
-
-				if (ctrl)
-				{
-					// typing after ctrl starts a new run, and not a part of the ctrl-backspace run
-					CommitAction(new ReplaceAction(oldText, text, selectionStart));
-				}
-			}
-		}
-
-		private void KeyDownUpArrow(KeyRoutedEventArgs args, string text, bool ctrl, bool shift, ref int selectionStart, ref int selectionLength)
-		{
-			// TODO ctrl+up
-			if (HasPointerCapture)
-			{
-				return;
-			}
-			if (Text.Length != 0)
-			{
-				TrySetCurrentlyTyping(false);
-			}
-
-			var start = selectionStart;
-			var end = selectionStart + selectionLength;
-			var newEnd = GetUpDownResult(text, selectionStart, selectionLength, shift, up: true);
-			if (shift)
-			{
-				selectionLength = newEnd - selectionStart;
-			}
-			else
-			{
-				selectionStart = newEnd;
-				selectionLength = 0;
-			}
-
-			args.Handled = selectionStart != start || selectionLength != end - start;
-		}
-
-		private void KeyDownDownArrow(KeyRoutedEventArgs args, string text, bool ctrl, bool shift, ref int selectionStart, ref int selectionLength)
-		{
-			// TODO ctrl+down
-			if (HasPointerCapture)
-			{
-				return;
-			}
-			if (Text.Length != 0)
-			{
-				TrySetCurrentlyTyping(false);
-			}
-
-			var start = selectionStart;
-			var end = selectionStart + selectionLength;
-			var newEnd = GetUpDownResult(text, selectionStart, selectionLength, shift, up: false);
-			if (shift)
-			{
-				selectionLength = newEnd - selectionStart;
-			}
-			else
-			{
-				selectionStart = newEnd;
-				selectionLength = 0;
-			}
-
-			args.Handled = selectionStart != start || selectionLength != end - start;
-		}
-
-		private void KeyDownLeftArrow(KeyRoutedEventArgs args, string text, bool shift, bool ctrl, ref int selectionStart, ref int selectionLength)
-		{
-			// on Apple platforms it is:
-			// * `command` + `left` that moves to the start of the line
-			// * `option` + `left` that moves to the previous word
-			if (DeviceTargetHelper.UsesAppleKeyboardLayout)
-			{
-				if (ctrl)
-				{
-					KeyDownHome(args, text, ctrl: false, shift, ref selectionStart, ref selectionLength);
-					return;
-				}
-				ctrl = args.KeyboardModifiers.HasFlag(VirtualKeyModifiers.Menu);
-			}
-
-			if (HasPointerCapture)
-			{
-				return;
-			}
-			if (Text.Length != 0)
-			{
-				TrySetCurrentlyTyping(false);
-			}
-
-			if (!shift && selectionStart == 0 && selectionLength == 0 || shift && selectionStart + selectionLength == 0)
-			{
-				return;
-			}
-
-			args.Handled = true;
-
-			if (shift)
-			{
-				var end = selectionStart + selectionLength;
-				if (ctrl)
-				{
-					end = TextBoxView.DisplayBlock.ParsedText.GetWordAt(end, false).start;
-				}
-				else
-				{
-					end--;
-				}
-
-				selectionLength = end - selectionStart;
-			}
-			else
-			{
-				if (selectionLength == 0)
-				{
-					selectionStart = ctrl ? TextBoxView.DisplayBlock.ParsedText.GetWordAt(selectionStart, false).start : selectionStart - 1;
-				}
-				else
-				{
-					selectionStart = Math.Min(selectionStart, selectionStart + selectionLength);
-				}
-				selectionLength = 0;
-			}
-		}
-
-		private void KeyDownRightArrow(KeyRoutedEventArgs args, string text, bool ctrl, bool shift, ref int selectionStart, ref int selectionLength)
-		{
-			// on Apple platforms it is:
-			// * `command` + `right` that moves to the end of the line
-			// * `option` + `right` that moves to the next word
-			// * `shift` + `option` + `right` that select the next word
-			if (DeviceTargetHelper.UsesAppleKeyboardLayout)
-			{
-				if (ctrl)
-				{
-					KeyDownEnd(args, text, ctrl: false, shift, ref selectionStart, ref selectionLength);
-					return;
-				}
-				ctrl = args.KeyboardModifiers.HasFlag(VirtualKeyModifiers.Menu);
-			}
-
-			if (HasPointerCapture)
-			{
-				return;
-			}
-			if (Text.Length != 0)
-			{
-				TrySetCurrentlyTyping(false);
-			}
-
-			var moveOutRight = !shift && selectionStart == text.Length && selectionLength == 0 || shift && selectionStart + selectionLength == Text.Length;
-			if (!moveOutRight)
-			{
-				args.Handled = true;
-
-				if (shift)
-				{
-					var end = selectionStart + selectionLength;
-					if (ctrl)
-					{
-						var chunk = TextBoxView.DisplayBlock.ParsedText.GetWordAt(end, true);
-						end = chunk.start + chunk.length;
-					}
-					else
-					{
-						end++;
-					}
-
-					selectionLength = end - selectionStart;
-				}
-				else
-				{
-					if (selectionLength == 0)
-					{
-						if (ctrl)
-						{
-							var chunk = TextBoxView.DisplayBlock.ParsedText.GetWordAt(selectionStart, true);
-							selectionStart = chunk.start + chunk.length;
-						}
-						else
-						{
-							selectionStart += 1;
-						}
-					}
-					else
-					{
-						selectionStart = Math.Max(selectionStart, selectionStart + selectionLength);
-					}
-					selectionLength = 0;
-				}
-			}
-		}
-
-		private void KeyDownHome(KeyRoutedEventArgs args, string text, bool ctrl, bool shift, ref int selectionStart, ref int selectionLength)
-		{
-			if (HasPointerCapture)
-			{
-				return;
-			}
-			if (Text.Length != 0)
-			{
-				TrySetCurrentlyTyping(false);
-			}
-
-			var start = selectionStart;
-			var end = selectionStart + selectionLength;
-			if (shift)
-			{
-				selectionLength = ctrl ? -selectionStart : TextBoxView.DisplayBlock.ParsedText.GetLineAt(selectionStart + selectionLength).start - selectionStart;
-			}
-			else
-			{
-				selectionStart = ctrl ? 0 : TextBoxView.DisplayBlock.ParsedText.GetLineAt(selectionStart + selectionLength).start;
-				selectionLength = 0;
-			}
-			args.Handled = selectionStart != start || selectionLength != end - start;
-		}
-
-		private void KeyDownEnd(KeyRoutedEventArgs args, string text, bool ctrl, bool shift, ref int selectionStart, ref int selectionLength)
-		{
-			if (HasPointerCapture)
-			{
-				return;
-			}
-			if (Text.Length != 0)
-			{
-				TrySetCurrentlyTyping(false);
-			}
-
-			var start = selectionStart;
-			var end = selectionStart + selectionLength;
-			if (shift)
-			{
-				if (ctrl)
-				{
-					selectionLength = text.Length - selectionStart;
-				}
-				else
-				{
-					var line = TextBoxView.DisplayBlock.ParsedText.GetLineAt(selectionStart + selectionLength);
-					selectionLength = line.start + line.length - selectionStart;
-				}
-			}
-			else
-			{
-				if (ctrl)
-				{
-					selectionStart = text.Length;
-				}
-				else
-				{
-					var line = TextBoxView.DisplayBlock.ParsedText.GetLineAt(selectionStart + selectionLength);
-					selectionStart = line.start + line.length;
-					if (line.length > 0 && selectionStart < text.Length && text[selectionStart - 1] == '\r')
-					{
-						// a newline is part of the line just before it, but End shouldn't go past the newline
-						selectionStart--;
-					}
-				}
-				selectionLength = 0;
-			}
-			args.Handled = selectionStart != start || selectionLength != end - start;
-		}
-
-		private void KeyDownDelete(KeyRoutedEventArgs args, ref string text, bool ctrl, bool shift, ref int selectionStart, ref int selectionLength)
-		{
-			// on Apple platforms it is `option` + `delete>` that removes the next word
-			if (DeviceTargetHelper.UsesAppleKeyboardLayout)
-			{
-				ctrl = args.KeyboardModifiers.HasFlag(VirtualKeyModifiers.Menu);
-			}
-
-			if (HasPointerCapture)
-			{
-				return;
-			}
-			TrySetCurrentlyTyping(false);
-			args.Handled = true;
-			var oldText = text;
-			if (selectionLength != 0)
-			{
-				var start = Math.Min(selectionStart, selectionStart + selectionLength);
-				var end = Math.Max(selectionStart, selectionStart + selectionLength);
-				text = text[..start] + text[end..];
-				CommitAction(new DeleteAction(oldText, text, selectionStart, selectionLength));
-				selectionLength = 0;
-				selectionStart = start;
-			}
-			else if (selectionStart != text.Length)
-			{
-				if (shift)
-				{
-					// On WinUI, shift-delete doesn't do anything if nothing is selected for some reason
-					// We still end the previous typing run
-					return;
-				}
-				int index;
-				if (ctrl)
-				{
-					var chunk = TextBoxView.DisplayBlock.ParsedText.GetWordAt(selectionStart, true);
-					index = chunk.start + chunk.length;
-				}
-				else
-				{
-					index = selectionStart + 1;
-				}
-				text = text[..selectionStart] + text[index..];
-				// On WinUI, when ctrl-delete is Undone, the deleted text actually gets selected even though initially, nothing was selected
-				CommitAction(new DeleteAction(oldText, text, selectionStart, ctrl ? index - selectionStart : 0));
-			}
-		}
-
-		/// <summary>
-		/// Takes a possibly-negative selection length, indicating a selection that goes backwards.
-		/// This makes the calculations a lot more natural.
-		/// </summary>
-		internal void SelectInternal(int selectionStart, int selectionLength)
-		{
-			_inSelectInternal = true;
-			_selection.selectionEndsAtTheStart = selectionLength < 0;
-			if (DisplayBlockInlines is { }) // this check is important because on start up, the Inlines haven't been created yet.
-			{
-				_caretXOffset = selectionLength >= 0 ?
-					(float)TextBoxView.DisplayBlock.ParsedText.GetRectForIndex(selectionStart + selectionLength).Left :
-					(float)TextBoxView.DisplayBlock.ParsedText.GetRectForIndex(selectionStart + selectionLength).Right;
-			}
-			Select(Math.Min(selectionStart, selectionStart + selectionLength), Math.Abs(selectionLength));
-			_inSelectInternal = false;
-		}
-
-		private void TimerOnTick(object sender, object e)
-		{
-			if (IsLoaded && IsFocused)
-			{
-				if (CaretMode == CaretDisplayMode.ThumblessCaretHidden)
-				{
-					CaretMode = CaretDisplayMode.ThumblessCaretShowing;
-				}
-				else if (CaretMode == CaretDisplayMode.ThumblessCaretShowing)
-				{
-					CaretMode = CaretDisplayMode.ThumblessCaretHidden;
-				}
-			}
-		}
-
-		/// <summary>
-		/// There are 2 concepts of a "line", there's a line that ends at end-of-text, \r, \n, etc.
-		/// and then there's an actual rendered line that may end due to wrapping and not a line break.
-		/// This method cares about the second kind of lines.
-		/// </summary>
-		private int GetUpDownResult(string text, int selectionStart, int selectionLength, bool shift, bool up)
-		{
-			if (text.Length == 0)
-			{
-				return 0;
-			}
-
-			var (startLineStart, startLineLength, startLineFirst, startLineLast, startLineIndex) = TextBoxView.DisplayBlock.ParsedText.GetLineAt(selectionStart);
-			var (endLineStart, endLineLength, endLineFirst, endLineLast, endLineIndex) = TextBoxView.DisplayBlock.ParsedText.GetLineAt(selectionStart + selectionLength);
-
-			if (up && shift && endLineFirst)
-			{
-				return 0; // first line, goes to the beginning
-			}
-			else if (!up && shift && endLineLast)
-			{
-				return text.Length; // last line, goes to the end
-			}
-			else if (!up && !shift && (startLineLast || endLineLast))
-			{
-				return text.Length; // last line, goes to the end
-			}
-
-			int newLineIndex;
-			if (up)
-			{
-				if (selectionLength < 0 || shift)
-				{
-					newLineIndex = !endLineFirst ? endLineIndex - 1 : endLineIndex;
-				}
-				else
-				{
-					newLineIndex = !startLineFirst ? startLineIndex - 1 : startLineIndex;
-				}
-			}
-			else
-			{
-				if (selectionLength > 0 || shift)
-				{
-					newLineIndex = !endLineLast ? endLineIndex + 1 : endLineIndex;
-				}
-				else
-				{
-					newLineIndex = !startLineLast ? startLineIndex + 1 : startLineIndex;
-				}
-			}
-
-			var rect = TextBoxView.DisplayBlock.ParsedText.GetRectForIndex(selectionStart + selectionLength);
-			var x = _caretXOffset;
-			var y = (newLineIndex + 0.5) * rect.Height; // 0.5 is to get the center of the line, rect.Height is line height
-			var index = Math.Max(0, TextBoxView.DisplayBlock.ParsedText.GetIndexAt(new Point(x, y), true, true));
-			var (newLineStart, newLineLength, newLineFirst, newLineLast, _) = TextBoxView.DisplayBlock.ParsedText.GetLineAt(index);
-			if (text.Length > index - 1
-				&& newLineLength > 1 // this check is for cases where the line has nothing but \r (i.e. is empty)
-				&& index - 1 >= 0
-				&& index == newLineStart + newLineLength
-				&& (text[index - 1] == '\r' || text[index - 1] == ' '))
-			{
-				// if we're past \r or space, we will actually be at the beginning of the next line, so we take a step back
-				index--;
-			}
-
-			return index;
-		}
-
-		private InlineCollection DisplayBlockInlines => TextBoxView?.DisplayBlock.Inlines;
-
-		/// <summary>
-		/// There are 2 concepts of a "line", there's a line that ends at end-of-text, \r, \n, etc.
-		/// and then there's an actual rendered line that may end due to wrapping and not a line break.
-		/// StartOfLine and EndOfLine care about the first kind of lines.
-		/// </summary>
-		private int StartOfLine(int i)
-		{
-			var text = Text;
-
-			i--;
-			for (; i >= 0; i--)
-			{
-				var c = text[i];
-				if (c == '\r')
-				{
-					break;
-				}
-			}
-
-			return i + 1;
-		}
-
-		private int EndOfLine(int i)
-		{
-			var index = Text.IndexOf('\r', i);
-			return index == -1 ? Text.Length - 1 : index;
-		}
-
-		partial void OnTextChangedPartial()
-		{
-			CancelCompositionOnExternalChange();
-
-			// Ported from: TextBoxBase.cpp TxNotify EN_CHANGE (line 3113)
-			// Close the selection flyout when text changes.
-			TextControlFlyoutHelper.CloseIfOpen(SelectionFlyout);
-
-			if (_pendingSelection is { } selection)
-			{
-				SelectInternal(selection.start, selection.length);
-			}
-			else
-			{
-				SelectInternal(0, 0);
-			}
-
-			if (_clearHistoryOnTextChanged)
-			{
-				ClearUndoRedoHistory();
-			}
-
-			_textBoxNotificationsSingleton?.NotifyValueChanged(this);
-		}
-
-		private string RemoveLF(string baseString)
-		{
-
-			var builder = new StringBuilder();
-			for (int i = 0; i < baseString.Length; i++)
-			{
-				var c = baseString[i];
-				if (c == '\n')
-				{
-					builder.Append('\r');
-				}
-				else if (c == '\r' && i + 1 < baseString.Length && baseString[i + 1] == '\n')
-				{
-					if (_pendingSelection is { } selection)
-					{
-						var (start, end) = (selection.start, selection.start + selection.length);
-						if (start > i)
-						{
-							start--;
-						}
-						if (end > i)
-						{
-							end--;
-						}
-						_pendingSelection = (start, end - start);
-					}
-
-					builder.Append('\r');
-					i++;
-				}
-				else
-				{
-					builder.Append(c);
-				}
-			}
-
-			baseString = builder.ToString();
-			return baseString;
-		}
-
-		partial void PasteFromClipboardPartial(string adjustedClipboardText, int selectionStart, string newText)
-		{
-			if (_currentlyTyping)
-			{
-				TrySetCurrentlyTyping(false);
-			}
-			else
-			{
-				// we only commit an action if we were not typing, because if we were typing and we now set CurrentlyTyping = false,
-				// we will already get a new action from the setter, so we don't need to commit another one here.
-				CommitAction(new ReplaceAction(Text, newText, selectionStart));
-			}
-
-			_pendingSelection = (selectionStart + adjustedClipboardText.Length, 0);
-		}
-
-		partial void CutSelectionToClipboardPartial()
-		{
-			if (_currentlyTyping)
-			{
-				TrySetCurrentlyTyping(false);
-			}
-			else
-			{
-				// we only commit an action if we were not typing, because if we were typing and we now set CurrentlyTyping = false,
-				// we will already get a new action from the setter, so we don't need to commit another one here.
-				CommitAction(new ReplaceAction(Text, Text.Remove(SelectionStart, SelectionLength), SelectionStart + SelectionLength));
-			}
-			_pendingSelection = (_selection.start, 0);
-		}
-
-		private void EnsureHistory()
-		{
-			if (_history.Count == 0)
-			{
-				_history.Add(new HistoryRecord(SentinelAction.Instance, _selection.start, _selection.length, _selection.selectionEndsAtTheStart));
-			}
-			_historyIndex = Math.Max(0, Math.Min(_history.Count - 1, _historyIndex));
-		}
-
-		public void ClearUndoRedoHistory()
-		{
-			TrySetCurrentlyTyping(false);
-			_history.Clear();
-			EnsureHistory();
-		}
-
-		/// <summary>
-		/// Adds a new Action at the present point in history and deletes the old "future"
-		/// </summary>
-		private void CommitAction(TextBoxAction action)
-		{
-			_historyIndex++;
-			_history.RemoveAllAt(_historyIndex);
-			_history.Add(new HistoryRecord(action, _selection.start, _selection.length, _selection.selectionEndsAtTheStart));
-		}
-
-		public void Undo()
-		{
-			TrySetCurrentlyTyping(false);
-			if (_historyIndex == 0 || HasPointerCapture)
-			{
-				return;
-			}
-
-			var currentAction = _history[_historyIndex];
-			_historyIndex--;
-
-			_clearHistoryOnTextChanged = false;
-			switch (currentAction.Action)
-			{
-				case ReplaceAction r:
-					// remember that we use the possibly-negative format in _pendingSelection
-					_pendingSelection = currentAction.SelectionEndsAtTheStart ?
-						(currentAction.SelectionStart + currentAction.SelectionLength, -currentAction.SelectionLength) :
-						(currentAction.SelectionStart, currentAction.SelectionLength);
-					ProcessTextInput(r.OldText);
-					break;
-				case DeleteAction d:
-					_pendingSelection = (d.UndoSelectionStart, d.UndoSelectionLength);
-					ProcessTextInput(d.OldText);
-					break;
-				case SentinelAction:
-					break;
-				default:
-					global::System.Diagnostics.CI.Assert(false, "TextBoxActions are not exhaustively switch-matched.");
-					break;
-			}
-			_clearHistoryOnTextChanged = true;
-		}
-
-		public void Redo()
-		{
-			if (_historyIndex == _history.Count - 1 || HasPointerCapture)
-			{
-				return;
-			}
-
-			TrySetCurrentlyTyping(false);
-
-			_historyIndex++;
-			var currentAction = _history[_historyIndex];
-
-			_clearHistoryOnTextChanged = false;
-			switch (currentAction.Action)
-			{
-				case ReplaceAction r:
-					_pendingSelection = (r.caretIndexAfterReplacement, 0); // we always have an empty selection here.
-					ProcessTextInput(r.NewText);
-					break;
-				case DeleteAction d:
-					_pendingSelection = (Math.Min(d.UndoSelectionStart, d.UndoSelectionStart + d.UndoSelectionLength), 0);
-					ProcessTextInput(d.NewText);
-					break;
-				case SentinelAction:
-					break;
-				default:
-					global::System.Diagnostics.CI.Assert(false, "TextBoxActions are not exhaustively switch-matched.");
-					break;
-			}
-			_clearHistoryOnTextChanged = true;
-		}
-
-		internal override bool IsDelegatingFocusToTemplateChild()
-			=> OperatingSystem.IsBrowser();
-
-		internal enum CaretDisplayMode
-		{
-			ThumblessCaretHidden,
-			ThumblessCaretShowing,
-			CaretWithThumbsOnlyEndShowing,
-			CaretWithThumbsBothEndsShowing
-		}
-
-		// Which platform's native touch text-selection conventions the Skia TextBox follows.
-		// Defaults to the running OS; runtime tests override it to exercise the mobile behavior on
-		// Skia Desktop, where OperatingSystem.IsAndroid()/IsIOS() are both false.
-		internal enum TouchTextSelectionConvention
-		{
-			// Desktop convention (Windows, macOS and Linux): the OS-derived default for every non-mobile target.
-			Desktop,
-			Android,
-			iOS
-		}
-
-		private record struct HistoryRecord(TextBoxAction Action, int SelectionStart, int SelectionLength, bool SelectionEndsAtTheStart);
-
-		private abstract record TextBoxAction;
-
-		/// <summary>
-		/// Instead of remembered what was removed and what was added in place, we just remember the initial and final states
-		/// as well as where the caret will be if we Redo. This is used by typing, paste, etc.
-		/// </summary>
-		private record ReplaceAction(string OldText, string NewText, int caretIndexAfterReplacement) : TextBoxAction;
-
-		/// <summary>
-		/// Unlike other forms of text modification, Delete doesn't follow the simple undo sequence of *unapply modification* -> *select what was selected when the action happened*
-		/// So we need to specifically need to remember what selection to go to when we Undo depending on how we got here (e.g. ctrl vs no ctrl)
-		/// Selection uses the possibly-negative format
-		/// </summary>
-		private record DeleteAction(string OldText, string NewText, int UndoSelectionStart, int UndoSelectionLength) : TextBoxAction;
-
-		/// <summary>
-		/// Probably unnecessary, but we pad the bottom of the history as it makes index manipulation easier (the invariant we
-		/// get is that history is never empty)
-		/// </summary>
-		private record SentinelAction : TextBoxAction
-		{
-			private SentinelAction() { }
-			public static SentinelAction Instance { get; } = new SentinelAction();
-		}
-
-		private sealed class TextBoxCommand(Action action) : ICommand
-		{
-			public bool CanExecute(object parameter) => true;
-
-			public void Execute(object parameter) => action();
-
-#pragma warning disable 67 // An event was declared but never used in the class in which it was declared.
-			public event EventHandler CanExecuteChanged;
-#pragma warning restore 67 // An event was declared but never used in the class in which it was declared.
-		}
-
-		#region ITextSelectionGripperHost
-		// The touch-selection grippers are driven by the shared TextSelectionGripperPresenter. The presenter
-		// owns the visuals/drag mechanics; these members expose just enough of TextBox's selection model and
-		// text surface for it to do its job.
-
-		TextBlock ITextSelectionGripperHost.GripperTextSurface => TextBoxView.DisplayBlock;
-
-		Rect ITextSelectionGripperHost.GripperClipBounds => this.GetAbsoluteBoundsRect();
-
-		GripperMode ITextSelectionGripperHost.GripperMode => _caretMode switch
-		{
-			CaretDisplayMode.CaretWithThumbsOnlyEndShowing => GripperMode.EndOnly,
-			CaretDisplayMode.CaretWithThumbsBothEndsShowing => GripperMode.Both,
-			_ => GripperMode.Hidden,
-		};
-
-		int ITextSelectionGripperHost.SelectionLowerIndex => SelectionStart;
-
-		int ITextSelectionGripperHost.SelectionUpperIndex => SelectionStart + SelectionLength;
-
-		void ITextSelectionGripperHost.SetGripperSelection(int start, int end) => SelectInternal(start, end - start);
-
-		void ITextSelectionGripperHost.MoveGripperCaret(int index) => Select(index, 0);
-
-		void ITextSelectionGripperHost.ScrollForGripper(bool isEndGripper) => UpdateScrolling(isEndGripper);
-
-		void ITextSelectionGripperHost.OnGripperPressed() => DismissSelectionFlyoutForPointerPress();
-
-		// Dismiss the selection flyout (the floating copy/paste bar) when a touch interaction begins, so it
-		// doesn't linger next to a context menu opened on the same gesture.
-		internal void DismissSelectionFlyoutForPointerPress() => TextControlFlyoutHelper.CloseIfOpen(SelectionFlyout);
-
-		void ITextSelectionGripperHost.RequestGripperContextMenu(PointerRoutedEventArgs args)
-		{
-			// Ported from TextBoxBase.cpp OnGripperHeld: a held gripper opens the context flyout.
-			var contextArgs = new ContextRequestedEventArgs();
-			contextArgs.SetGlobalPoint(args.GetCurrentPoint(null).Position);
-			OnContextRequested(this, contextArgs);
-		}
-
-		void ITextSelectionGripperHost.QueueGripperSelectionFlyout(PointerRoutedEventArgs args, bool allowEmptySelection)
-			=> QueueUpdateSelectionFlyoutVisibility(
-				args.Pointer.PointerDeviceType,
-				args.GetCurrentPoint(this).Position,
-				// The single insertion handle only exists under the mobile (iOS/Android) conventions; tapping it
-				// re-opens the flyout even over a collapsed caret. Desktop never gets the empty-selection flyout.
-				allowEmptySelection: allowEmptySelection && TouchSelectionConvention != TouchTextSelectionConvention.Desktop);
-
-		void ITextSelectionGripperHost.OnGripperTapped(PointerPoint press, int anchorIndex)
-		{
-			// The insertion handle (EndOnly gripper) sits over the character it points at, so the second tap of
-			// a double-tap lands on the handle instead of the text. Fold it into the same multi-tap counter as
-			// OnPointerReleasedPartial (press-to-press) so the double-tap still selects the word.
-			var repeatedPresses = _lastPointerDown.point is { } previous && IsTouchMultiTap(previous, press)
-				? _lastPointerDown.repeatedPresses + 1
-				: 0;
-			_lastPointerDown = (press, repeatedPresses);
-
-			// Act on the character the handle points at, not the finger's position on the thumb (which hangs below
-			// the caret line and would jump the caret to the end of the text).
-			if (TouchSelectionConvention != TouchTextSelectionConvention.Desktop && repeatedPresses >= 1)
-			{
-				TouchSelectWordAt(anchorIndex);
-			}
-			else if (_caretMode != CaretDisplayMode.CaretWithThumbsBothEndsShowing)
-			{
-				TouchTapAt(anchorIndex);
-			}
-			// Both thumbs showing: the tap grabbed a selection edge, so keep the selection (native iOS/Android)
-			// instead of collapsing it to a caret. The presenter re-shows the selection flyout on release.
-		}
-		#endregion
-
-		/// <summary>
-		/// Fires the ContextMenuOpening event synchronously and returns whether it was handled.
-		/// </summary>
-		/// <remarks>
-		/// Ported from CTextBoxBase::FireContextMenuOpeningEventSynchronously (TextBoxBase.cpp:5304)
-		/// and TextControlHelper::OnContextMenuOpeningHandler (TextControlHelper.h:10).
-		///
-		/// WinUI does m_pView->TransformToRoot(point) then divides by rasterization scale
-		/// to convert physical pixels to DIPs. In Uno/Skia, TransformToVisual(null) already
-		/// yields DIP coordinates, so no rasterization scale division is needed.
-		/// </remarks>
-		internal bool FireContextMenuOpeningEventSynchronously(Point point)
-		{
-			// WinUI: m_pView->TransformToRoot + pointerPosition /= zoomScale
-			var rootPoint = TransformToVisual(null).TransformPoint(point);
-
-			// WinUI: TextControlHelper::OnContextMenuOpeningHandler
-			// — creates ContextMenuEventArgs, sets CursorLeft/CursorTop, Handled=false
-			// — raises event synchronously
-			// — reads back Handled
-			var args = new ContextMenuEventArgs(rootPoint.X, rootPoint.Y);
-			ContextMenuOpening?.Invoke(this, args);
-			return args.Handled;
-		}
 	}
 }
