+++ src/Uno.UI/UI/Xaml/Controls/TextBoxCore/TextBoxCore.skia.cs
@@ -1,0 +1,1881 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Windows.Input;
+using Windows.Foundation;
+using Windows.System;
+using Windows.UI;
+using Microsoft.UI.Composition;
+using Microsoft.UI.Input;
+using Microsoft.UI.Xaml.Controls.Primitives;
+using Microsoft.UI.Xaml.Documents;
+using Microsoft.UI.Xaml.Internal;
+using Microsoft.UI.Xaml.Input;
+using Microsoft.UI.Xaml.Media;
+using Microsoft.UI.Xaml.Shapes;
+using SkiaSharp;
+using Uno.Extensions;
+using Uno.Foundation.Extensibility;
+using Uno.UI;
+using Uno.UI.Dispatching;
+using Uno.UI.Helpers;
+using Uno.UI.Xaml;
+using Uno.UI.Xaml.Controls.Extensions;
+using Uno.UI.Xaml.Media;
+using System.Runtime.InteropServices.JavaScript;
+using Microsoft.UI.Xaml.Documents.TextFormatting;
+using Uno.UI.Xaml.Controls;
+using Uno.UI.Xaml.Core;
+using Uno.Foundation;
+using Windows.ApplicationModel.DataTransfer;
+using DispatcherQueuePriority = Microsoft.UI.Dispatching.DispatcherQueuePriority;
+using Microsoft.UI.Xaml.Media.Media3D;
+using System.Numerics;
+using Uno.Disposables;
+
+namespace Microsoft.UI.Xaml.Controls;
+
+using SelectionDetails = (int start, int length, bool selectionEndsAtTheStart);
+
+internal sealed partial class TextBoxCore : ITextSelectionGripperHost
+{
+	private TextSelectionGripperPresenter _gripperPresenter;
+
+	// Test hook: the pair of touch-selection grippers when they are currently showing, otherwise null.
+	internal (CaretWithStemAndThumb start, CaretWithStemAndThumb end)? VisibleGrippersForTesting => _gripperPresenter?.VisibleGrippersForTesting;
+
+	private TextBoxView _textBoxView;
+	private static ITextBoxNotificationsProviderSingleton _textBoxNotificationsSingleton;
+	private SerialDisposable _clipboardChangeSubscription = new SerialDisposable();
+
+	private SelectionDetails _selection;
+	private float _caretXOffset; // this is not necessarily the visual offset of the caret, but where the caret is logically supposed to be when moving up and down with the keyboard, even if the caret is temporarily elsewhere
+	private CaretDisplayMode _caretMode = CaretDisplayMode.ThumblessCaretHidden;
+
+	private bool _inSelectInternal;
+
+	private (int start, int length)? _pendingSelection;
+
+	private bool _clearHistoryOnTextChanged = true;
+	private CompositionBrush _cachedCaretBrush;
+	private Color _cachedCaretColor;
+
+	private static readonly VirtualKeyModifiers _platformCtrlKey;
+
+	// We track what constitutes one typing "action" that can be undone/redone. The general gist is that
+	// any sequence of characters (with backspace allowed) without any navigation moves (pointer click, arrow keys, etc.)
+	// will be one "run"/"action". However, there are some arbitrary exceptions, so that is only a rule of thumb.
+	private bool _currentlyTyping;
+	private bool _suppressCurrentlyTyping;
+	private SelectionDetails _selectionWhenTypingStarted;
+	private string _textWhenTypingStarted;
+
+	private int _historyIndex;
+	private (int start, int length, bool isBackward) _lastNotifiedSelection = (-1, -1, false);
+	private readonly List<HistoryRecord> _history = new(); // the selection of an action is what was selected right before it happened. Might turn out to be unnecessary.
+
+	private readonly DispatcherTimer _timer = new() { Interval = TimeSpan.FromSeconds(0.5) };
+
+	private MenuFlyout _proofingMenu;
+
+	internal bool IsBackwardSelection => _selection.selectionEndsAtTheStart;
+
+	internal TextBoxView TextBoxView => _textBoxView;
+
+	internal ContentControl ContentElement => _contentElement;
+
+	static TextBoxCore()
+	{
+		_platformCtrlKey = Uno.UI.Helpers.DeviceTargetHelper.PlatformCommandModifier;
+		_ = ApiExtensibility.CreateInstance(null, out _textBoxNotificationsSingleton);
+		InitializeIme();
+	}
+
+	internal CaretDisplayMode CaretMode
+	{
+		get => _caretMode;
+		private set
+		{
+			if (_caretMode != value)
+			{
+				_caretMode = value;
+				UpdateDisplaySelection();
+				if (TextBoxView?.DisplayBlock.Visual is { } visual)
+				{
+					Owner.Visual.Compositor.InvalidateRender(visual);
+				}
+				if (value is CaretDisplayMode.ThumblessCaretShowing)
+				{
+					_timer.Start(); // restart
+				}
+				else if (value is CaretDisplayMode.CaretWithThumbsBothEndsShowing
+						 or CaretDisplayMode.CaretWithThumbsOnlyEndShowing)
+				{
+					_timer.Stop();
+				}
+			}
+		}
+	}
+
+	// Settable so runtime tests can force a mobile convention on Skia Desktop, where
+	// OperatingSystem.IsAndroid()/IsIOS() are both false and the OS-derived default is Desktop.
+	internal TouchTextSelectionConvention TouchSelectionConvention { get; set; } = GetDefaultTouchTextSelectionConvention();
+
+	private static TouchTextSelectionConvention GetDefaultTouchTextSelectionConvention()
+		=> OperatingSystem.IsAndroid() ? TouchTextSelectionConvention.Android
+			: Uno.UI.Helpers.DeviceTargetHelper.IsUIKit() ? TouchTextSelectionConvention.iOS
+			: TouchTextSelectionConvention.Desktop;
+
+	// The undo/redo depth backing the controls' CanUndo/CanRedo dependency properties. The properties
+	// themselves stay on the control, because their PropMethodCall callbacks need GetValue/SetValue.
+	internal bool CanUndoInternal => _historyIndex > 0;
+
+	internal bool CanRedoInternal => _historyIndex < _history.Count - 1;
+
+	// Ported from: TextBoxBase.cpp GetProofingMenuFlyoutNoRef (lines 5507-5520)
+	internal MenuFlyout EnsureAndUpdateProofingMenu()
+	{
+		EnsureProofingMenu();
+		TextControlFlyoutHelper.AddProofingFlyout(_proofingMenu, _host.Owner);
+		if (IsSpellCheckEnabled && (FocusState != FocusState.Unfocused || _forceFocusedVisualState))
+		{
+			UpdateProofingMenu();
+		}
+		return _proofingMenu;
+	}
+
+	private void UpdateCanPasteClipboardContent()
+	{
+		if (IsReadOnly)
+		{
+			CanPasteClipboardContent = false;
+			return;
+		}
+
+		try
+		{
+			var content = Clipboard.GetContent();
+			CanPasteClipboardContent = content?.Contains(StandardDataFormats.Text) ?? false;
+		}
+		catch
+		{
+			CanPasteClipboardContent = false;
+		}
+	}
+
+	// Ported from: TextBoxBase.cpp EnsureProofingMenu (line 5470)
+	private void EnsureProofingMenu()
+	{
+		_proofingMenu ??= new MenuFlyout();
+	}
+
+	// Ported from: TextBoxBase.cpp UpdateProofingMenu (line 5536)
+	private void UpdateProofingMenu()
+	{
+		_proofingMenu.Items.Clear();
+
+		if (TextBoxView?.DisplayBlock?.ParsedText is not UnicodeText unicodeText)
+		{
+			return;
+		}
+
+		// Find correction at cursor position
+		var correction = unicodeText.GetCorrectionAtIndex(SelectionStart);
+		if (correction is null)
+		{
+			return;
+		}
+
+		var (correctionStart, correctionEnd) = correction.Value;
+
+		// Get suggestions (matches WinUI AddSpellingSuggestions, max 3 suggestions)
+		var suggestionsResult = unicodeText.GetSpellCheckSuggestions(correctionStart, correctionEnd);
+		if (suggestionsResult is not { } result || result.suggestions.Count == 0)
+		{
+			return;
+		}
+
+		var (replaceStart, replaceEnd, suggestions) = result;
+		var maxSuggestions = Math.Min(suggestions.Count, 3); // WinUI uses maxSuggestionCount = 3
+
+		for (var i = 0; i < maxSuggestions; i++)
+		{
+			var suggestion = suggestions[i];
+			var item = new MenuFlyoutItem { Text = suggestion };
+			item.Click += (_, _) => ReplaceWithSuggestion(replaceStart, replaceEnd, suggestion);
+			_proofingMenu.Items.Add(item);
+		}
+	}
+
+	// Ported from: TextBoxBase.cpp pattern for applying text replacement (similar to PasteFromClipboardPartial)
+	private void ReplaceWithSuggestion(int replaceStart, int replaceEnd, string suggestion)
+	{
+		if (IsReadOnly)
+		{
+			return;
+		}
+
+		TrySetCurrentlyTyping(false);
+
+		var oldText = Text;
+		var newText = oldText.Remove(replaceStart, replaceEnd - replaceStart).Insert(replaceStart, suggestion);
+
+		CommitAction(new ReplaceAction(oldText, newText, replaceStart + suggestion.Length));
+
+		_clearHistoryOnTextChanged = false;
+		_pendingSelection = (replaceStart + suggestion.Length, 0);
+		ProcessTextInput(newText);
+		_clearHistoryOnTextChanged = true;
+	}
+
+	private void OnClipboardContentChanged(object sender, object e)
+	{
+		if (DispatcherQueue.HasThreadAccess)
+		{
+			UpdateCanPasteClipboardContent();
+		}
+		else
+		{
+			DispatcherQueue.TryEnqueue(() => UpdateCanPasteClipboardContent());
+		}
+	}
+
+	private void TrySetCurrentlyTyping(bool newValue)
+	{
+		if (newValue == _currentlyTyping || _suppressCurrentlyTyping)
+		{
+			return;
+		}
+
+		if (newValue)
+		{
+			_textWhenTypingStarted = Text;
+			_selectionWhenTypingStarted = (
+				_selection.start,
+				_selection.length,
+				_selection.selectionEndsAtTheStart);
+		}
+		else
+		{
+			_historyIndex++;
+			_history.RemoveAllAt(_historyIndex);
+			_history.Add(new HistoryRecord(
+				new ReplaceAction(_textWhenTypingStarted, Text, _selection.start),
+				_selectionWhenTypingStarted.start,
+				_selectionWhenTypingStarted.length,
+				_selectionWhenTypingStarted.selectionEndsAtTheStart));
+		}
+
+		_currentlyTyping = newValue;
+	}
+
+	partial void OnUnloadedPartial()
+	{
+		_forceFocusedVisualState = false;
+		_timer.Stop();
+		_gripperPresenter?.Hide();
+		CaretMode = CaretDisplayMode.ThumblessCaretHidden;
+		_clipboardChangeSubscription.Disposable = null;
+	}
+
+	partial void OnIsReadonlyChangedPartial() => UpdateCanPasteClipboardContent();
+
+	partial void OnForegroundColorChangedPartial(Brush newValue)
+	{
+		TextBoxView?.OnForegroundChanged(newValue);
+		// Update the caret brush to match the new foreground color.
+		UpdateDisplaySelection();
+	}
+
+	partial void OnSelectionHighlightColorChangedPartial(SolidColorBrush brush) => TextBoxView?.OnSelectionHighlightColorChanged(brush);
+
+	partial void UpdateFontPartial()
+	{
+		TextBoxView?.UpdateFont();
+	}
+
+	partial void OnInputScopeChangedPartial(InputScope newValue) => TextBoxView?.UpdateProperties();
+
+	partial void OnIsSpellCheckEnabledChangedPartial(bool newValue)
+	{
+		if (TextBoxView is not null)
+		{
+			TextBoxView.DisplayBlock.IsSpellCheckEnabled = newValue;
+			TextBoxView.UpdateProperties();
+		}
+	}
+
+	partial void OnIsTextPredictionEnabledChangedPartial(bool newValue) => TextBoxView?.UpdateProperties();
+
+	partial void OnMaxLengthChangedPartial(int newValue) => TextBoxView?.UpdateMaxLength();
+
+	partial void OnFlowDirectionChangedPartial()
+	{
+		TextBoxView?.SetFlowDirection();
+	}
+
+	partial void OnTextWrappingChangedPartial()
+	{
+		TextBoxView?.SetWrapping();
+		if (_contentElement is ScrollViewer sv)
+		{
+			// This is to work around sv giving infinite width. This has the unfortunate problem of resetting
+			// locally-set values and/or changes in the template.
+			sv.HorizontalScrollBarVisibility = TextWrapping == TextWrapping.NoWrap ? ScrollBarVisibility.Hidden : ScrollBarVisibility.Disabled;
+		}
+	}
+
+	partial void OnTextAlignmentChangedPartial(TextAlignment newValue)
+	{
+		TextBoxView?.SetTextAlignment();
+	}
+
+	partial void SetInputReturnTypePlatform(InputReturnType inputReturnType)
+	{
+		TextBoxView?.UpdateProperties();
+	}
+
+	private void UpdateTextBoxView()
+	{
+		_textBoxView ??= new TextBoxView(this);
+		if (ContentElement != null)
+		{
+			var displayBlock = TextBoxView.DisplayBlock;
+			if (ContentElement.Content != displayBlock)
+			{
+				ContentElement.Content = displayBlock;
+
+				// The presenter owns the touch-selection gripper visuals and their drag/positioning
+				// logic (shared with selectable TextBlock). It subscribes to DisplayBlock.DrawingFinished
+				// to keep the grippers glued to the selection ends as the text is (re)drawn.
+				_gripperPresenter ??= new TextSelectionGripperPresenter(this);
+			}
+
+			TextBoxView.SetTextNative(Text);
+		}
+	}
+
+	partial void OnFocusStateChangedPartial(FocusState focusState, bool initial)
+	{
+		_clipboardChangeSubscription.Disposable = null;
+		TextBoxView?.OnFocusStateChanged(focusState);
+
+		if (focusState != FocusState.Unfocused)
+		{
+			CaretMode = CaretDisplayMode.ThumblessCaretShowing;
+			_textBoxNotificationsSingleton?.OnFocused(this);
+			StartImeSession();
+			UpdateCanPasteClipboardContent();
+			Clipboard.ContentChanged += OnClipboardContentChanged;
+			_clipboardChangeSubscription.Disposable = Disposable.Create(() => Clipboard.ContentChanged -= OnClipboardContentChanged);
+		}
+		else
+		{
+			// Ported from: TextBoxBase.cpp UpdateFocusState (lines 4996-5008)
+			// Check if focus is moving to a text control flyout before deciding to unfocus.
+			_forceFocusedVisualState = ShouldForceFocusedVisualState();
+
+			// Ported from: TextBoxBase.cpp UpdateFocusState (lines 5009-5016)
+			// Hide touch caret thumbs when context flyout is opening.
+			if (_forceFocusedVisualState && ShouldHideGrippersOnFlyoutOpening()
+				&& CaretMode is CaretDisplayMode.CaretWithThumbsOnlyEndShowing
+					or CaretDisplayMode.CaretWithThumbsBothEndsShowing)
+			{
+				CaretMode = CaretDisplayMode.ThumblessCaretShowing;
+			}
+		}
+
+		if (focusState == FocusState.Unfocused && !_forceFocusedVisualState)
+		{
+			EndImeSession();
+			TrySetCurrentlyTyping(false);
+			CaretMode = CaretDisplayMode.ThumblessCaretHidden;
+			if (SelectionFlyout?.IsOpen == true)
+			{
+				SelectionFlyout.Hide();
+			}
+			if (!initial)
+			{
+				_textBoxNotificationsSingleton?.OnUnfocused(this);
+			}
+			_timer.Stop();
+		}
+		UpdateDisplaySelection();
+	}
+
+	// Ported from: TextBoxBase.cpp ShouldForceFocusedVisualState (lines 5286-5290)
+	private bool ShouldForceFocusedVisualState()
+	{
+		return TextControlFlyoutHelper.IsGettingFocus(SelectionFlyout, Owner)
+			|| TextControlFlyoutHelper.IsGettingFocus(ContextFlyout, Owner);
+	}
+
+	// Ported from: TextBoxBase.cpp ShouldHideGrippersOnFlyoutOpening (lines 5291-5294)
+	private bool ShouldHideGrippersOnFlyoutOpening()
+	{
+		return TextControlFlyoutHelper.IsGettingFocus(ContextFlyout, Owner);
+	}
+
+	// Ported from: TextBoxBase.cpp DismissAllFlyouts (lines 5566-5574)
+	internal void DismissAllFlyouts()
+	{
+		TextControlFlyoutHelper.CloseIfOpen(_proofingMenu);
+		TextControlFlyoutHelper.CloseIfOpen(SelectionFlyout);
+		TextControlFlyoutHelper.CloseIfOpen(ContextFlyout);
+	}
+
+	// Ported from: TextBoxBase.cpp ForceFocusLoss (lines 5551-5560)
+	internal void ForceFocusLoss()
+	{
+		_forceFocusedVisualState = false;
+		// Trigger the unfocus behavior that was deferred
+		OnFocusStateChangedPartial(FocusState.Unfocused, initial: false);
+		UpdateVisualState();
+	}
+
+#if false // Removing temporarily. We'll need to add it back.
+	// TODO: Discuss this public API.
+	public static void FinishAutofillContext(bool shouldSave)
+	{
+		_textBoxNotificationsSingleton?.FinishAutofillContext(shouldSave);
+	}
+#endif
+
+	partial void SelectPartial(int start, int length)
+	{
+		TrySetCurrentlyTyping(false);
+
+		if (!_inSelectInternal)
+		{
+			// SelectInternal sets _selectionEndsAtTheStart and _caretXOffset on its own
+			_selection.selectionEndsAtTheStart = false;
+			_caretXOffset = (float)(TextBoxView?.DisplayBlock.ParsedText.GetRectForIndex(start + length).Left ?? 0);
+		}
+
+		var selection = (start, length, _selection.selectionEndsAtTheStart);
+		var selectionChanged = selection != _selection;
+		_selection = selection;
+
+		// Even when using Skia TextBox, we may need to call Select,
+		// which will update the native input in case of Wasm Skia for example.
+		TextBoxView?.Select(start, length);
+
+		if (length == 0 && CaretMode == CaretDisplayMode.CaretWithThumbsBothEndsShowing)
+		{
+			// It doesn't make sense to have 2 caret ends when there's no selection.
+			CaretMode = CaretDisplayMode.CaretWithThumbsOnlyEndShowing;
+		}
+		else if (CaretMode is CaretDisplayMode.ThumblessCaretHidden)
+		{
+			CaretMode = CaretDisplayMode.ThumblessCaretShowing;
+		}
+		else if (CaretMode is CaretDisplayMode.ThumblessCaretShowing)
+		{
+			_timer.Start(); // restart
+		}
+
+		if (selectionChanged)
+		{
+			UpdateScrolling();
+		}
+		UpdateDisplaySelection();
+	}
+
+	partial void SelectAllPartial() => Select(0, Text.Length);
+
+	internal int SelectionStart
+	{
+		get => _selection.start;
+		set => Select(start: value, length: SelectionLength);
+	}
+
+	internal int SelectionLength
+	{
+		get => _selection.length;
+		set => Select(SelectionStart, value);
+	}
+
+	private void UpdateDisplaySelection()
+	{
+		if (TextBoxView?.DisplayBlock is { } displayBlock)
+		{
+			displayBlock.Selection = new TextBlock.Range(SelectionStart, SelectionStart + SelectionLength);
+			var isFocused = FocusState != FocusState.Unfocused || _forceFocusedVisualState;
+			displayBlock.RenderSelection = isFocused;
+			if (CaretMode is CaretDisplayMode.ThumblessCaretShowing &&
+				SelectionLength == 0 &&
+				isFocused &&
+				!IsReadOnly &&
+				!FeatureConfiguration.TextBox.HideCaret)
+			{
+				// WinUI uses DestInvert composite mode with a white brush so the caret
+				// is always visible regardless of theme. Uno doesn't support DestInvert,
+				// so we use the TextBox's own Foreground (element-theme-aware via
+				// ThemeResource) with fully opaque alpha for maximum caret visibility.
+				var caretBrush = GetOpaqueCaretBrush();
+				displayBlock.RenderCaret = (IsBackwardSelection ? SelectionStart : SelectionStart + SelectionLength, caretBrush);
+			}
+			else
+			{
+				displayBlock.RenderCaret = null;
+			}
+			((IBlock)TextBoxView.DisplayBlock).Invalidate(false);
+
+			var currentSelection = (SelectionStart, SelectionLength, IsBackwardSelection);
+			if (currentSelection != _lastNotifiedSelection)
+			{
+				_lastNotifiedSelection = currentSelection;
+				_textBoxNotificationsSingleton?.NotifySelectionChanged(this);
+			}
+		}
+	}
+
+	/// <summary>
+	/// Gets a fully opaque composition brush derived from the TextBox's Foreground for caret rendering.
+	/// WinUI uses DestInvert composite mode (white brush + invert), but Uno doesn't support that.
+	/// Instead, we use the element-theme-aware Foreground with full opacity for maximum visibility.
+	/// </summary>
+	private CompositionBrush GetOpaqueCaretBrush()
+	{
+		var compositor = Compositor.GetSharedCompositor();
+		if (Foreground is SolidColorBrush scb)
+		{
+			var color = scb.Color;
+			if (color.A < 255)
+			{
+				// Force fully opaque for caret visibility
+				color = Color.FromArgb(255, color.R, color.G, color.B);
+			}
+
+			if (_cachedCaretBrush is not null && _cachedCaretColor == color)
+			{
+				return _cachedCaretBrush;
+			}
+
+			_cachedCaretColor = color;
+			_cachedCaretBrush = compositor.CreateColorBrush(color);
+			return _cachedCaretBrush;
+		}
+
+		_cachedCaretBrush = null;
+		_cachedCaretColor = default;
+		return DefaultBrushes.TextForegroundBrush.GetOrCreateCompositionBrush(compositor);
+	}
+
+	private void UpdateScrolling() => UpdateScrolling(true);
+
+	/// <summary>
+	/// Gets the position where a context menu should be shown for keyboard invocation.
+	/// Returns the position at the selection/caret location.
+	/// </summary>
+	/// <returns>The position in local coordinates, or null if position cannot be determined.</returns>
+	internal Point? GetContextMenuShowPosition()
+	{
+		if (TextBoxView?.DisplayBlock?.ParsedText == null)
+		{
+			return null;
+		}
+
+		// Determine the character index to use for positioning
+		// Use selection end for LTR (selection start for RTL to match WinUI behavior)
+		var index = FlowDirection == FlowDirection.RightToLeft && SelectionLength > 0
+			? SelectionStart
+			: SelectionStart + SelectionLength;
+
+		// Get the rect for the character at the position
+		var rect = TextBoxView.DisplayBlock.ParsedText.GetRectForIndex(index);
+
+		// Transform from DisplayBlock coordinates to TextBox coordinates
+		var transform = TextBoxView.DisplayBlock.TransformToVisual(Owner);
+		var point = transform.TransformPoint(new Point(rect.Left, rect.Bottom));
+
+		return point;
+	}
+
+	#region SelectionFlyout Support
+
+	// Ported from: microsoft-ui-xaml2/src/dxaml/xcp/core/native/text/Controls/TextBoxBase.h (lines 225-226)
+	// m_lastInputDeviceType, m_lastPointerPositionForFlyout, m_isSelectionFlyoutUpdateQueued
+	private PointerDeviceType _lastInputDeviceType;
+	private Point _lastPointerPosition;
+	private bool _isSelectionFlyoutUpdateQueued;
+	// Uno addition (not part of the WinUI port): lets the flyout re-open over a collapsed caret when the single
+	// insertion handle is tapped. See QueueGripperSelectionFlyout / OnGripperTapped.
+	private bool _allowEmptySelectionFlyout;
+
+	// Ported from: microsoft-ui-xaml2/src/dxaml/xcp/core/native/text/Controls/TextBoxBase.cpp (lines 5292-5302)
+	private bool HasSelectionFlyout() => SelectionFlyout is not null;
+
+	// Ported from: microsoft-ui-xaml2/src/dxaml/xcp/core/native/text/Controls/TextBoxBase.cpp (lines 5349-5377)
+	// QueueUpdateSelectionFlyoutVisibility - queues async visibility update
+	private void QueueUpdateSelectionFlyoutVisibility(PointerDeviceType deviceType, Point position, bool allowEmptySelection = false)
+	{
+		_lastInputDeviceType = deviceType;
+		_lastPointerPosition = position;
+		_allowEmptySelectionFlyout = allowEmptySelection;
+
+		// Line 5358-5360: Prevent duplicate queued updates
+		if (!_isSelectionFlyoutUpdateQueued)
+		{
+			_isSelectionFlyoutUpdateQueued = true;
+			DispatcherQueue.TryEnqueue(() => UpdateSelectionFlyoutVisibility());
+		}
+	}
+
+	// Ported from: microsoft-ui-xaml2/src/dxaml/xcp/core/native/text/Controls/TextBoxBase.cpp (lines 5379-5452)
+	// UpdateSelectionFlyoutVisibility - shows/hides SelectionFlyout based on device type and selection
+	private void UpdateSelectionFlyoutVisibility()
+	{
+		// Line 5381: Reset the queued flag
+		_isSelectionFlyoutUpdateQueued = false;
+
+		// Line 5383: Only shows SelectionFlyout if ContextFlyout is NOT open
+		if (!HasSelectionFlyout() || TextControlFlyoutHelper.IsOpen(ContextFlyout))
+		{
+			return;
+		}
+
+		var selectionLength = SelectionLength;
+		var showMode = FlyoutShowMode.Transient;
+		var shouldShow = false;
+
+		switch (_lastInputDeviceType)
+		{
+			case PointerDeviceType.Mouse:
+				// Line 5389-5401: Mouse input - only for RichEditBox, NOT TextBox
+				shouldShow = false;
+				break;
+
+			case PointerDeviceType.Pen:
+			case PointerDeviceType.Touch:
+				// Line 5403-5411: Pen/Touch - show if selection exists.
+				// Uno addition (beyond the WinUI port): also show over a collapsed caret when the single insertion
+				// handle was tapped, matching the native iOS/Android Paste/Select-all popup (gated to those
+				// conventions by the caller).
+				if (selectionLength > 0 || _allowEmptySelectionFlyout)
+				{
+					shouldShow = true;
+					showMode = FlyoutShowMode.Transient;
+				}
+				break;
+
+			default:
+				// Line 5413-5415: Keyboard/other - never show SelectionFlyout
+				shouldShow = false;
+				break;
+		}
+
+		if (shouldShow)
+		{
+			// Line 5424-5441: Get selection bounds and adjust flyout position
+			var position = _lastPointerPosition;
+
+			if (TextBoxView?.DisplayBlock?.ParsedText is { } parsedText)
+			{
+				// Get selection bounding rect in DisplayBlock coordinates
+				var startRect = parsedText.GetRectForIndex(SelectionStart);
+				var endRect = parsedText.GetRectForIndex(SelectionStart + SelectionLength);
+
+				// Compute union of start and end rects for full selection bounds
+				var selectionTop = Math.Min(startRect.Top, endRect.Top);
+
+				// Transform from DisplayBlock coordinates to TextBox coordinates
+				var transform = TextBoxView.DisplayBlock.TransformToVisual(Owner);
+				var transformedTop = transform.TransformPoint(new Point(0, selectionTop));
+
+				// Set Y position to top of selection for consistent placement
+				// (WinUI: "We want the point to appear at a consistent location regardless of
+				// whether it overlaps the selection, so we'll set its vertical position to be
+				// the top of the exclusion rectangle.")
+				position = new Point(position.X, transformedTop.Y);
+			}
+
+			if (SelectionFlyout is { } selectionFlyout)
+			{
+				TextControlFlyoutHelper.ShowAt(selectionFlyout, Owner, position, showMode);
+			}
+		}
+		else
+		{
+			// Line 5444-5447: Close SelectionFlyout if it's open and we shouldn't show it
+			if (SelectionFlyout?.IsOpen == true)
+			{
+				SelectionFlyout.Hide();
+			}
+		}
+
+		// Line 5450: Reset input device type after processing
+		_lastInputDeviceType = default;
+		_allowEmptySelectionFlyout = default;
+	}
+
+	#endregion
+
+	/// <summary>
+	/// Narrows the BringIntoView target rect to the caret position so that parent
+	/// ScrollViewers scroll to the cursor rather than the full TextBox bounds.
+	/// </summary>
+	/// <remarks>
+	/// When a TextBox has its internal scrolling disabled and relies on an outer
+	/// ScrollViewer, the default BringIntoView uses the full element bounds.
+	/// For a tall TextBox this causes the outer ScrollViewer to scroll to the
+	/// element edges rather than to the cursor position.
+	/// </remarks>
+	internal void OnBringIntoViewRequested(BringIntoViewRequestedEventArgs e)
+	{
+		if ((e.TargetElement is null || e.TargetElement == Owner)
+			&& FocusState != FocusState.Unfocused
+			&& _contentElement is ScrollViewer { VerticalScrollMode: ScrollMode.Disabled }
+			&& TextBoxView?.DisplayBlock is { } displayBlock
+			&& displayBlock.ParsedText is { } parsedText)
+		{
+			var selectionEnd = _selection.selectionEndsAtTheStart
+				? _selection.start
+				: _selection.start + _selection.length;
+
+			var caretRect = parsedText.GetRectForIndex(selectionEnd);
+			caretRect = caretRect with { Width = Math.Max(caretRect.Width, TextBlock.CaretThickness) };
+
+			// Transform from TextBlock coordinate space to TextBox coordinate space
+			var transform = displayBlock.TransformToVisual(Owner);
+			var caretRectInTextBox = transform.TransformBounds(caretRect);
+
+			e.TargetRect = caretRectInTextBox;
+		}
+	}
+
+	/// <summary>
+	/// Scrolls the <see cref="_contentElement"/> so that the caret is inside the visible viewport
+	/// </summary>
+	/// <remarks>
+	/// By default, only the selection end moves, while the selection start stays fixed. This is not the
+	/// case when dragging the caret thumb, in which case both ends can move. This case requires an
+	/// explicit call to this method with <see cref="putSelectionEndInVisibleViewport"/> = false.
+	/// </remarks>>
+	private void UpdateScrolling(bool putSelectionEndInVisibleViewport)
+	{
+		if (_contentElement is ScrollViewer sv)
+		{
+			var horizontalOffset = sv.HorizontalOffset;
+			var verticalOffset = sv.VerticalOffset;
+
+			var (selectionStart, selectionEnd) = _selection.selectionEndsAtTheStart ? (_selection.start + _selection.length, _selection.start) : (_selection.start, _selection.start + _selection.length);
+			var index = putSelectionEndInVisibleViewport ? selectionEnd : selectionStart;
+
+			var caretRect = TextBoxView.DisplayBlock.ParsedText.GetRectForIndex(index) with { Width = TextBlock.CaretThickness };
+
+			// Because the caret is only a single-pixel wide, and because screens can't draw in fractions of a pixel,
+			// we need to add Math.Ceiling to ensure that the caret is (fully) included in the visible viewport. This
+			// Math.Ceiling sometimes horizontal overscrolling, but it's more acceptable than sometimes not showing the caret.
+			var newHorizontalOffset = horizontalOffset.AtMost(caretRect.Left).AtLeast(Math.Ceiling(caretRect.Right - sv.ViewportWidth + TextBlock.CaretThickness));
+
+			var newVerticalOffset = verticalOffset.AtMost(caretRect.Top).AtLeast(caretRect.Bottom - sv.ViewportHeight);
+
+			sv.ChangeView(newHorizontalOffset, newVerticalOffset, null);
+		}
+
+		// When the TextBox's internal scrolling is disabled, an outer ScrollViewer
+		// handles scrolling. Notify it so the caret stays visible after text changes.
+		if (FocusState != FocusState.Unfocused
+			&& _contentElement is ScrollViewer { VerticalScrollMode: ScrollMode.Disabled })
+		{
+			StartBringIntoView(new BringIntoViewOptions { AnimationDesired = false });
+		}
+	}
+
+	partial void OnKeyDownPartial(KeyRoutedEventArgs args)
+	{
+		// This is a minimal copy of OnkeyDownSkia that just sets args.Handled without doing any work.
+		// This is to match WinUI behavior where Handled is set for certain keys before public
+		// subscribers get the event, but before any actual text processing is done.
+		var (selectionStart, selectionLength) = _selection.selectionEndsAtTheStart ? (_selection.start + _selection.length, -_selection.length) : (_selection.start, _selection.length);
+		var text = Text;
+		var shift = args.KeyboardModifiers.HasFlag(VirtualKeyModifiers.Shift);
+		var ctrl = args.KeyboardModifiers.HasFlag(_platformCtrlKey);
+		switch (args.Key)
+		{
+			case VirtualKey.Escape:
+				if (HasPointerCapture)
+				{
+					args.Handled = true;
+				}
+				return;
+			case VirtualKey.Z when ctrl:
+			case VirtualKey.Y when ctrl:
+			case VirtualKey.Delete when !IsReadOnly:
+			case VirtualKey.A when ctrl:
+				if (!HasPointerCapture)
+				{
+					args.Handled = true;
+				}
+				return;
+			case VirtualKey.Up:
+				// on Apple platforms start of document is `Command` and `Up`
+				if (ctrl && DeviceTargetHelper.UsesAppleKeyboardLayout)
+				{
+					KeyDownHome(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
+				}
+				else
+				{
+					KeyDownUpArrow(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
+				}
+				break;
+			case VirtualKey.Down:
+				// on Apple platforms end of document is `Command` and `Down`
+				if (ctrl && DeviceTargetHelper.UsesAppleKeyboardLayout)
+				{
+					KeyDownEnd(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
+				}
+				else
+				{
+					KeyDownDownArrow(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
+				}
+				break;
+			case VirtualKey.Left when !TextBoxView.DisplayBlock.ParsedText.IsBaseDirectionRightToLeft:
+			case VirtualKey.Right when TextBoxView.DisplayBlock.ParsedText.IsBaseDirectionRightToLeft:
+				KeyDownLeftArrow(args, text, shift, ctrl, ref selectionStart, ref selectionLength);
+				break;
+			case VirtualKey.Left when TextBoxView.DisplayBlock.ParsedText.IsBaseDirectionRightToLeft:
+			case VirtualKey.Right when !TextBoxView.DisplayBlock.ParsedText.IsBaseDirectionRightToLeft:
+				KeyDownRightArrow(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
+				break;
+			case VirtualKey.Home:
+				KeyDownHome(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
+				break;
+			case VirtualKey.End:
+				KeyDownEnd(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
+				break;
+			// TODO: PageUp/Down
+			case VirtualKey.Back when !IsReadOnly:
+				KeyDownBack(args, ref text, ctrl, shift, ref selectionStart, ref selectionLength);
+				break;
+		}
+	}
+
+	private void OnKeyDownSkia(KeyRoutedEventArgs args)
+	{
+		if (SelectionFlyout?.IsOpen == true)
+		{
+			SelectionFlyout.Hide();
+		}
+
+		if (_selection.length != 0 &&
+			args.Key is not (VirtualKey.Up or VirtualKey.Down or VirtualKey.Left or VirtualKey.Right))
+		{
+			// On WinUI, pressing anything except arrow keys will immediately make the caret thumbless.
+			// Even shift + arrow keys will make the caret thumbless (because it's a shift _then_ an arrow key).
+			CaretMode = CaretDisplayMode.ThumblessCaretShowing;
+		}
+
+		// Note: On windows ** only KeyDown ** is handled (not KeyUp)
+
+		// move to possibly-negative selection length format
+		var (selectionStart, selectionLength) = _selection.selectionEndsAtTheStart ? (_selection.start + _selection.length, -_selection.length) : (_selection.start, _selection.length);
+
+		var text = Text;
+		var shift = args.KeyboardModifiers.HasFlag(VirtualKeyModifiers.Shift);
+		var ctrl = args.KeyboardModifiers.HasFlag(_platformCtrlKey);
+		// Text commands: always return from this switch, never break
+		switch (args.Key)
+		{
+			case VirtualKey.Z when ctrl:
+				if (!HasPointerCapture)
+				{
+					args.Handled = true;
+					Undo();
+				}
+				return;
+			case VirtualKey.Y when ctrl:
+				if (!HasPointerCapture)
+				{
+					args.Handled = true;
+					Redo();
+				}
+				return;
+			case VirtualKey.X when ctrl:
+				CutSelectionToClipboard();
+				return;
+			case VirtualKey.V when ctrl:
+			case VirtualKey.Insert when shift:
+				PasteFromClipboard(); // async so doesn't actually do anything right now
+				return;
+			case VirtualKey.C when ctrl:
+			case VirtualKey.Insert when ctrl:
+				CopySelectionToClipboard();
+				return;
+			case VirtualKey.Escape:
+				if (HasPointerCapture)
+				{
+					args.Handled = true;
+					ReleasePointerCaptures();
+				}
+				return;
+			case VirtualKey.LeftShift:
+			case VirtualKey.RightShift:
+			case VirtualKey.Shift:
+			case VirtualKey.Control:
+			case VirtualKey.LeftControl:
+			case VirtualKey.RightControl:
+			case VirtualKey.Menu:
+			case VirtualKey.LeftMenu:
+			case VirtualKey.RightMenu:
+			case VirtualKey.LeftWindows:
+			case VirtualKey.RightWindows:
+				// No-op when pressing these key specifically.
+				return;
+		}
+
+		// text input/movement
+		switch (args.Key)
+		{
+			case VirtualKey.Up:
+				// on Apple platforms start of document is `Command` and `Up`
+				if (ctrl && DeviceTargetHelper.UsesAppleKeyboardLayout)
+				{
+					KeyDownHome(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
+				}
+				else
+				{
+					KeyDownUpArrow(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
+				}
+				break;
+			case VirtualKey.Down:
+				// on Apple platforms end of document is `Command` and `Down`
+				if (ctrl && DeviceTargetHelper.UsesAppleKeyboardLayout)
+				{
+					KeyDownEnd(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
+				}
+				else
+				{
+					KeyDownDownArrow(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
+				}
+				break;
+			case VirtualKey.Left when !TextBoxView.DisplayBlock.ParsedText.IsBaseDirectionRightToLeft:
+			case VirtualKey.Right when TextBoxView.DisplayBlock.ParsedText.IsBaseDirectionRightToLeft:
+				KeyDownLeftArrow(args, text, shift, ctrl, ref selectionStart, ref selectionLength);
+				break;
+			case VirtualKey.Left when TextBoxView.DisplayBlock.ParsedText.IsBaseDirectionRightToLeft:
+			case VirtualKey.Right when !TextBoxView.DisplayBlock.ParsedText.IsBaseDirectionRightToLeft:
+				KeyDownRightArrow(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
+				break;
+			case VirtualKey.Home:
+				KeyDownHome(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
+				break;
+			case VirtualKey.End:
+				KeyDownEnd(args, text, ctrl, shift, ref selectionStart, ref selectionLength);
+				break;
+			// TODO: PageUp/Down
+			case VirtualKey.Back when !IsReadOnly:
+				KeyDownBack(args, ref text, ctrl, shift, ref selectionStart, ref selectionLength);
+				break;
+			case VirtualKey.Delete when !IsReadOnly:
+				KeyDownDelete(args, ref text, ctrl, shift, ref selectionStart, ref selectionLength);
+				break;
+			case VirtualKey.A when ctrl:
+				if (!HasPointerCapture)
+				{
+					args.Handled = true;
+					TrySetCurrentlyTyping(false);
+					selectionStart = 0;
+					selectionLength = text.Length;
+				}
+				break;
+			default:
+				// During IME composition, skip normal character insertion.
+				// The IME extension handles text updates via OnImeCompositionUpdated → ProcessTextInput.
+				// On platforms where the platform applies text directly (e.g., Android),
+				// key events are independent of composition events and should not be swallowed.
+				if (ShouldSwallowKeyDuringComposition)
+				{
+					return;
+				}
+				var isEnterKey = args.UnicodeKey is '\r' or '\n' || args.Key == VirtualKey.Enter;
+				// Don't insert characters when shortcut modifier keys are held down.
+				// On non-Apple platforms, AltGr (Ctrl+Alt together) produces valid characters (like @, €),
+				// so we must not suppress those. On Apple, Option (Menu/Alt) produces special characters.
+				var altHeld = args.KeyboardModifiers.HasFlag(VirtualKeyModifiers.Menu);
+				var ctrlHeld = args.KeyboardModifiers.HasFlag(VirtualKeyModifiers.Control);
+				var isAltGr = !DeviceTargetHelper.UsesAppleKeyboardLayout && ctrlHeld && altHeld;
+				var hasShortcutModifier = !isAltGr && (
+					ctrlHeld ||
+					args.KeyboardModifiers.HasFlag(VirtualKeyModifiers.Windows) ||
+					(!DeviceTargetHelper.UsesAppleKeyboardLayout && altHeld));
+				if (!IsReadOnly && !HasPointerCapture && !hasShortcutModifier && args.UnicodeKey is { } key && (!isEnterKey || AcceptsReturn))
+				{
+					TrySetCurrentlyTyping(true);
+					var start = Math.Min(selectionStart, selectionStart + selectionLength);
+					var end = Math.Max(selectionStart, selectionStart + selectionLength);
+
+					if (key is '\n')
+					{
+						// TextBox autoconverts to \r, like WinUI
+						key = '\r';
+					}
+
+					text = text[..start] + key + text[end..];
+					selectionStart = start + 1;
+					selectionLength = 0;
+					break;
+				}
+				else
+				{
+					return;
+				}
+		}
+
+		selectionStart = Math.Max(0, Math.Min(text.Length, selectionStart));
+		selectionLength = Math.Max(-selectionStart, Math.Min(text.Length - selectionStart, selectionLength));
+
+		var caretXOffset = _caretXOffset;
+
+		_suppressCurrentlyTyping = true;
+		_clearHistoryOnTextChanged = false;
+		if (!HasPointerCapture)
+		{
+			_pendingSelection = (selectionStart, selectionLength);
+		}
+
+		ProcessTextInput(text);
+		_clearHistoryOnTextChanged = true;
+		_suppressCurrentlyTyping = false;
+
+		// don't change the caret offset when moving up and down
+		if (args.Key is VirtualKey.Up or VirtualKey.Down)
+		{
+			// this condition is accurate in the case of hitting Down on the last line
+			// or up on the first line. On WinUI, the caret offset won't change.
+			_caretXOffset = caretXOffset;
+		}
+	}
+
+	internal void SetPendingSelection(int selectionStart, int selectionLength)
+		=> _pendingSelection = (selectionStart, selectionLength);
+
+	private void KeyDownBack(KeyRoutedEventArgs args, ref string text, bool ctrl, bool shift, ref int selectionStart, ref int selectionLength)
+	{
+		// on Apple platforms it is `option` + `delete` (same location as backspace on PC keyboards) that removes the previous word
+		if (DeviceTargetHelper.UsesAppleKeyboardLayout)
+		{
+			ctrl = args.KeyboardModifiers.HasFlag(VirtualKeyModifiers.Menu);
+		}
+
+		if (HasPointerCapture)
+		{
+			return;
+		}
+		if (selectionLength != 0)
+		{
+			TrySetCurrentlyTyping(false);
+			TrySetCurrentlyTyping(true);
+
+			var start = Math.Min(selectionStart, selectionStart + selectionLength);
+			var end = Math.Max(selectionStart, selectionStart + selectionLength);
+			text = text[..start] + text[end..];
+			selectionLength = 0;
+			selectionStart = start;
+		}
+		else if (selectionStart != 0)
+		{
+			if (ctrl)
+			{
+				// ctrl always ends the previous typing run
+				TrySetCurrentlyTyping(false);
+			}
+			else
+			{
+				// idempotent call to make sure we're starting a new typing run if we're not in one already
+				TrySetCurrentlyTyping(true);
+			}
+
+			var oldText = text;
+			var index = ctrl ? TextBoxView.DisplayBlock.ParsedText.GetWordAt(selectionStart, false).start : selectionStart - 1;
+			text = text[..index] + text[selectionStart..];
+			selectionStart = index;
+
+			if (ctrl)
+			{
+				// typing after ctrl starts a new run, and not a part of the ctrl-backspace run
+				CommitAction(new ReplaceAction(oldText, text, selectionStart));
+			}
+		}
+	}
+
+	private void KeyDownUpArrow(KeyRoutedEventArgs args, string text, bool ctrl, bool shift, ref int selectionStart, ref int selectionLength)
+	{
+		// TODO ctrl+up
+		if (HasPointerCapture)
+		{
+			return;
+		}
+		if (Text.Length != 0)
+		{
+			TrySetCurrentlyTyping(false);
+		}
+
+		var start = selectionStart;
+		var end = selectionStart + selectionLength;
+		var newEnd = GetUpDownResult(text, selectionStart, selectionLength, shift, up: true);
+		if (shift)
+		{
+			selectionLength = newEnd - selectionStart;
+		}
+		else
+		{
+			selectionStart = newEnd;
+			selectionLength = 0;
+		}
+
+		args.Handled = selectionStart != start || selectionLength != end - start;
+	}
+
+	private void KeyDownDownArrow(KeyRoutedEventArgs args, string text, bool ctrl, bool shift, ref int selectionStart, ref int selectionLength)
+	{
+		// TODO ctrl+down
+		if (HasPointerCapture)
+		{
+			return;
+		}
+		if (Text.Length != 0)
+		{
+			TrySetCurrentlyTyping(false);
+		}
+
+		var start = selectionStart;
+		var end = selectionStart + selectionLength;
+		var newEnd = GetUpDownResult(text, selectionStart, selectionLength, shift, up: false);
+		if (shift)
+		{
+			selectionLength = newEnd - selectionStart;
+		}
+		else
+		{
+			selectionStart = newEnd;
+			selectionLength = 0;
+		}
+
+		args.Handled = selectionStart != start || selectionLength != end - start;
+	}
+
+	private void KeyDownLeftArrow(KeyRoutedEventArgs args, string text, bool shift, bool ctrl, ref int selectionStart, ref int selectionLength)
+	{
+		// on Apple platforms it is:
+		// * `command` + `left` that moves to the start of the line
+		// * `option` + `left` that moves to the previous word
+		if (DeviceTargetHelper.UsesAppleKeyboardLayout)
+		{
+			if (ctrl)
+			{
+				KeyDownHome(args, text, ctrl: false, shift, ref selectionStart, ref selectionLength);
+				return;
+			}
+			ctrl = args.KeyboardModifiers.HasFlag(VirtualKeyModifiers.Menu);
+		}
+
+		if (HasPointerCapture)
+		{
+			return;
+		}
+		if (Text.Length != 0)
+		{
+			TrySetCurrentlyTyping(false);
+		}
+
+		if (!shift && selectionStart == 0 && selectionLength == 0 || shift && selectionStart + selectionLength == 0)
+		{
+			return;
+		}
+
+		args.Handled = true;
+
+		if (shift)
+		{
+			var end = selectionStart + selectionLength;
+			if (ctrl)
+			{
+				end = TextBoxView.DisplayBlock.ParsedText.GetWordAt(end, false).start;
+			}
+			else
+			{
+				end--;
+			}
+
+			selectionLength = end - selectionStart;
+		}
+		else
+		{
+			if (selectionLength == 0)
+			{
+				selectionStart = ctrl ? TextBoxView.DisplayBlock.ParsedText.GetWordAt(selectionStart, false).start : selectionStart - 1;
+			}
+			else
+			{
+				selectionStart = Math.Min(selectionStart, selectionStart + selectionLength);
+			}
+			selectionLength = 0;
+		}
+	}
+
+	private void KeyDownRightArrow(KeyRoutedEventArgs args, string text, bool ctrl, bool shift, ref int selectionStart, ref int selectionLength)
+	{
+		// on Apple platforms it is:
+		// * `command` + `right` that moves to the end of the line
+		// * `option` + `right` that moves to the next word
+		// * `shift` + `option` + `right` that select the next word
+		if (DeviceTargetHelper.UsesAppleKeyboardLayout)
+		{
+			if (ctrl)
+			{
+				KeyDownEnd(args, text, ctrl: false, shift, ref selectionStart, ref selectionLength);
+				return;
+			}
+			ctrl = args.KeyboardModifiers.HasFlag(VirtualKeyModifiers.Menu);
+		}
+
+		if (HasPointerCapture)
+		{
+			return;
+		}
+		if (Text.Length != 0)
+		{
+			TrySetCurrentlyTyping(false);
+		}
+
+		var moveOutRight = !shift && selectionStart == text.Length && selectionLength == 0 || shift && selectionStart + selectionLength == Text.Length;
+		if (!moveOutRight)
+		{
+			args.Handled = true;
+
+			if (shift)
+			{
+				var end = selectionStart + selectionLength;
+				if (ctrl)
+				{
+					var chunk = TextBoxView.DisplayBlock.ParsedText.GetWordAt(end, true);
+					end = chunk.start + chunk.length;
+				}
+				else
+				{
+					end++;
+				}
+
+				selectionLength = end - selectionStart;
+			}
+			else
+			{
+				if (selectionLength == 0)
+				{
+					if (ctrl)
+					{
+						var chunk = TextBoxView.DisplayBlock.ParsedText.GetWordAt(selectionStart, true);
+						selectionStart = chunk.start + chunk.length;
+					}
+					else
+					{
+						selectionStart += 1;
+					}
+				}
+				else
+				{
+					selectionStart = Math.Max(selectionStart, selectionStart + selectionLength);
+				}
+				selectionLength = 0;
+			}
+		}
+	}
+
+	private void KeyDownHome(KeyRoutedEventArgs args, string text, bool ctrl, bool shift, ref int selectionStart, ref int selectionLength)
+	{
+		if (HasPointerCapture)
+		{
+			return;
+		}
+		if (Text.Length != 0)
+		{
+			TrySetCurrentlyTyping(false);
+		}
+
+		var start = selectionStart;
+		var end = selectionStart + selectionLength;
+		if (shift)
+		{
+			selectionLength = ctrl ? -selectionStart : TextBoxView.DisplayBlock.ParsedText.GetLineAt(selectionStart + selectionLength).start - selectionStart;
+		}
+		else
+		{
+			selectionStart = ctrl ? 0 : TextBoxView.DisplayBlock.ParsedText.GetLineAt(selectionStart + selectionLength).start;
+			selectionLength = 0;
+		}
+		args.Handled = selectionStart != start || selectionLength != end - start;
+	}
+
+	private void KeyDownEnd(KeyRoutedEventArgs args, string text, bool ctrl, bool shift, ref int selectionStart, ref int selectionLength)
+	{
+		if (HasPointerCapture)
+		{
+			return;
+		}
+		if (Text.Length != 0)
+		{
+			TrySetCurrentlyTyping(false);
+		}
+
+		var start = selectionStart;
+		var end = selectionStart + selectionLength;
+		if (shift)
+		{
+			if (ctrl)
+			{
+				selectionLength = text.Length - selectionStart;
+			}
+			else
+			{
+				var line = TextBoxView.DisplayBlock.ParsedText.GetLineAt(selectionStart + selectionLength);
+				selectionLength = line.start + line.length - selectionStart;
+			}
+		}
+		else
+		{
+			if (ctrl)
+			{
+				selectionStart = text.Length;
+			}
+			else
+			{
+				var line = TextBoxView.DisplayBlock.ParsedText.GetLineAt(selectionStart + selectionLength);
+				selectionStart = line.start + line.length;
+				if (line.length > 0 && selectionStart < text.Length && text[selectionStart - 1] == '\r')
+				{
+					// a newline is part of the line just before it, but End shouldn't go past the newline
+					selectionStart--;
+				}
+			}
+			selectionLength = 0;
+		}
+		args.Handled = selectionStart != start || selectionLength != end - start;
+	}
+
+	private void KeyDownDelete(KeyRoutedEventArgs args, ref string text, bool ctrl, bool shift, ref int selectionStart, ref int selectionLength)
+	{
+		// on Apple platforms it is `option` + `delete>` that removes the next word
+		if (DeviceTargetHelper.UsesAppleKeyboardLayout)
+		{
+			ctrl = args.KeyboardModifiers.HasFlag(VirtualKeyModifiers.Menu);
+		}
+
+		if (HasPointerCapture)
+		{
+			return;
+		}
+		TrySetCurrentlyTyping(false);
+		args.Handled = true;
+		var oldText = text;
+		if (selectionLength != 0)
+		{
+			var start = Math.Min(selectionStart, selectionStart + selectionLength);
+			var end = Math.Max(selectionStart, selectionStart + selectionLength);
+			text = text[..start] + text[end..];
+			CommitAction(new DeleteAction(oldText, text, selectionStart, selectionLength));
+			selectionLength = 0;
+			selectionStart = start;
+		}
+		else if (selectionStart != text.Length)
+		{
+			if (shift)
+			{
+				// On WinUI, shift-delete doesn't do anything if nothing is selected for some reason
+				// We still end the previous typing run
+				return;
+			}
+			int index;
+			if (ctrl)
+			{
+				var chunk = TextBoxView.DisplayBlock.ParsedText.GetWordAt(selectionStart, true);
+				index = chunk.start + chunk.length;
+			}
+			else
+			{
+				index = selectionStart + 1;
+			}
+			text = text[..selectionStart] + text[index..];
+			// On WinUI, when ctrl-delete is Undone, the deleted text actually gets selected even though initially, nothing was selected
+			CommitAction(new DeleteAction(oldText, text, selectionStart, ctrl ? index - selectionStart : 0));
+		}
+	}
+
+	/// <summary>
+	/// Takes a possibly-negative selection length, indicating a selection that goes backwards.
+	/// This makes the calculations a lot more natural.
+	/// </summary>
+	internal void SelectInternal(int selectionStart, int selectionLength)
+	{
+		_inSelectInternal = true;
+		_selection.selectionEndsAtTheStart = selectionLength < 0;
+		if (DisplayBlockInlines is { }) // this check is important because on start up, the Inlines haven't been created yet.
+		{
+			_caretXOffset = selectionLength >= 0 ?
+				(float)TextBoxView.DisplayBlock.ParsedText.GetRectForIndex(selectionStart + selectionLength).Left :
+				(float)TextBoxView.DisplayBlock.ParsedText.GetRectForIndex(selectionStart + selectionLength).Right;
+		}
+		Select(Math.Min(selectionStart, selectionStart + selectionLength), Math.Abs(selectionLength));
+		_inSelectInternal = false;
+	}
+
+	private void TimerOnTick(object sender, object e)
+	{
+		if (IsLoaded && IsFocused)
+		{
+			if (CaretMode == CaretDisplayMode.ThumblessCaretHidden)
+			{
+				CaretMode = CaretDisplayMode.ThumblessCaretShowing;
+			}
+			else if (CaretMode == CaretDisplayMode.ThumblessCaretShowing)
+			{
+				CaretMode = CaretDisplayMode.ThumblessCaretHidden;
+			}
+		}
+	}
+
+	/// <summary>
+	/// There are 2 concepts of a "line", there's a line that ends at end-of-text, \r, \n, etc.
+	/// and then there's an actual rendered line that may end due to wrapping and not a line break.
+	/// This method cares about the second kind of lines.
+	/// </summary>
+	private int GetUpDownResult(string text, int selectionStart, int selectionLength, bool shift, bool up)
+	{
+		if (text.Length == 0)
+		{
+			return 0;
+		}
+
+		var (startLineStart, startLineLength, startLineFirst, startLineLast, startLineIndex) = TextBoxView.DisplayBlock.ParsedText.GetLineAt(selectionStart);
+		var (endLineStart, endLineLength, endLineFirst, endLineLast, endLineIndex) = TextBoxView.DisplayBlock.ParsedText.GetLineAt(selectionStart + selectionLength);
+
+		if (up && shift && endLineFirst)
+		{
+			return 0; // first line, goes to the beginning
+		}
+		else if (!up && shift && endLineLast)
+		{
+			return text.Length; // last line, goes to the end
+		}
+		else if (!up && !shift && (startLineLast || endLineLast))
+		{
+			return text.Length; // last line, goes to the end
+		}
+
+		int newLineIndex;
+		if (up)
+		{
+			if (selectionLength < 0 || shift)
+			{
+				newLineIndex = !endLineFirst ? endLineIndex - 1 : endLineIndex;
+			}
+			else
+			{
+				newLineIndex = !startLineFirst ? startLineIndex - 1 : startLineIndex;
+			}
+		}
+		else
+		{
+			if (selectionLength > 0 || shift)
+			{
+				newLineIndex = !endLineLast ? endLineIndex + 1 : endLineIndex;
+			}
+			else
+			{
+				newLineIndex = !startLineLast ? startLineIndex + 1 : startLineIndex;
+			}
+		}
+
+		var rect = TextBoxView.DisplayBlock.ParsedText.GetRectForIndex(selectionStart + selectionLength);
+		var x = _caretXOffset;
+		var y = (newLineIndex + 0.5) * rect.Height; // 0.5 is to get the center of the line, rect.Height is line height
+		var index = Math.Max(0, TextBoxView.DisplayBlock.ParsedText.GetIndexAt(new Point(x, y), true, true));
+		var (newLineStart, newLineLength, newLineFirst, newLineLast, _) = TextBoxView.DisplayBlock.ParsedText.GetLineAt(index);
+		if (text.Length > index - 1
+			&& newLineLength > 1 // this check is for cases where the line has nothing but \r (i.e. is empty)
+			&& index - 1 >= 0
+			&& index == newLineStart + newLineLength
+			&& (text[index - 1] == '\r' || text[index - 1] == ' '))
+		{
+			// if we're past \r or space, we will actually be at the beginning of the next line, so we take a step back
+			index--;
+		}
+
+		return index;
+	}
+
+	private InlineCollection DisplayBlockInlines => TextBoxView?.DisplayBlock.Inlines;
+
+	/// <summary>
+	/// There are 2 concepts of a "line", there's a line that ends at end-of-text, \r, \n, etc.
+	/// and then there's an actual rendered line that may end due to wrapping and not a line break.
+	/// StartOfLine and EndOfLine care about the first kind of lines.
+	/// </summary>
+	private int StartOfLine(int i)
+	{
+		var text = Text;
+
+		i--;
+		for (; i >= 0; i--)
+		{
+			var c = text[i];
+			if (c == '\r')
+			{
+				break;
+			}
+		}
+
+		return i + 1;
+	}
+
+	private int EndOfLine(int i)
+	{
+		var index = Text.IndexOf('\r', i);
+		return index == -1 ? Text.Length - 1 : index;
+	}
+
+	partial void OnTextChangedPartial()
+	{
+		CancelCompositionOnExternalChange();
+
+		// Ported from: TextBoxBase.cpp TxNotify EN_CHANGE (line 3113)
+		// Close the selection flyout when text changes.
+		TextControlFlyoutHelper.CloseIfOpen(SelectionFlyout);
+
+		if (_pendingSelection is { } selection)
+		{
+			SelectInternal(selection.start, selection.length);
+		}
+		else
+		{
+			SelectInternal(0, 0);
+		}
+
+		if (_clearHistoryOnTextChanged)
+		{
+			ClearUndoRedoHistory();
+		}
+
+		_textBoxNotificationsSingleton?.NotifyValueChanged(this);
+	}
+
+	private string RemoveLF(string baseString)
+	{
+
+		var builder = new StringBuilder();
+		for (int i = 0; i < baseString.Length; i++)
+		{
+			var c = baseString[i];
+			if (c == '\n')
+			{
+				builder.Append('\r');
+			}
+			else if (c == '\r' && i + 1 < baseString.Length && baseString[i + 1] == '\n')
+			{
+				if (_pendingSelection is { } selection)
+				{
+					var (start, end) = (selection.start, selection.start + selection.length);
+					if (start > i)
+					{
+						start--;
+					}
+					if (end > i)
+					{
+						end--;
+					}
+					_pendingSelection = (start, end - start);
+				}
+
+				builder.Append('\r');
+				i++;
+			}
+			else
+			{
+				builder.Append(c);
+			}
+		}
+
+		baseString = builder.ToString();
+		return baseString;
+	}
+
+	partial void PasteFromClipboardPartial(string adjustedClipboardText, int selectionStart, string newText)
+	{
+		if (_currentlyTyping)
+		{
+			TrySetCurrentlyTyping(false);
+		}
+		else
+		{
+			// we only commit an action if we were not typing, because if we were typing and we now set CurrentlyTyping = false,
+			// we will already get a new action from the setter, so we don't need to commit another one here.
+			CommitAction(new ReplaceAction(Text, newText, selectionStart));
+		}
+
+		_pendingSelection = (selectionStart + adjustedClipboardText.Length, 0);
+	}
+
+	partial void CutSelectionToClipboardPartial()
+	{
+		if (_currentlyTyping)
+		{
+			TrySetCurrentlyTyping(false);
+		}
+		else
+		{
+			// we only commit an action if we were not typing, because if we were typing and we now set CurrentlyTyping = false,
+			// we will already get a new action from the setter, so we don't need to commit another one here.
+			CommitAction(new ReplaceAction(Text, Text.Remove(SelectionStart, SelectionLength), SelectionStart + SelectionLength));
+		}
+		_pendingSelection = (_selection.start, 0);
+	}
+
+	private void EnsureHistory()
+	{
+		if (_history.Count == 0)
+		{
+			_history.Add(new HistoryRecord(SentinelAction.Instance, _selection.start, _selection.length, _selection.selectionEndsAtTheStart));
+		}
+		_historyIndex = Math.Max(0, Math.Min(_history.Count - 1, _historyIndex));
+	}
+
+	internal void ClearUndoRedoHistory()
+	{
+		TrySetCurrentlyTyping(false);
+		_history.Clear();
+		EnsureHistory();
+	}
+
+	/// <summary>
+	/// Adds a new Action at the present point in history and deletes the old "future"
+	/// </summary>
+	private void CommitAction(TextBoxAction action)
+	{
+		_historyIndex++;
+		_history.RemoveAllAt(_historyIndex);
+		_history.Add(new HistoryRecord(action, _selection.start, _selection.length, _selection.selectionEndsAtTheStart));
+	}
+
+	internal void Undo()
+	{
+		TrySetCurrentlyTyping(false);
+		if (_historyIndex == 0 || HasPointerCapture)
+		{
+			return;
+		}
+
+		var currentAction = _history[_historyIndex];
+		_historyIndex--;
+
+		_clearHistoryOnTextChanged = false;
+		switch (currentAction.Action)
+		{
+			case ReplaceAction r:
+				// remember that we use the possibly-negative format in _pendingSelection
+				_pendingSelection = currentAction.SelectionEndsAtTheStart ?
+					(currentAction.SelectionStart + currentAction.SelectionLength, -currentAction.SelectionLength) :
+					(currentAction.SelectionStart, currentAction.SelectionLength);
+				ProcessTextInput(r.OldText);
+				break;
+			case DeleteAction d:
+				_pendingSelection = (d.UndoSelectionStart, d.UndoSelectionLength);
+				ProcessTextInput(d.OldText);
+				break;
+			case SentinelAction:
+				break;
+			default:
+				global::System.Diagnostics.CI.Assert(false, "TextBoxActions are not exhaustively switch-matched.");
+				break;
+		}
+		_clearHistoryOnTextChanged = true;
+	}
+
+	internal void Redo()
+	{
+		if (_historyIndex == _history.Count - 1 || HasPointerCapture)
+		{
+			return;
+		}
+
+		TrySetCurrentlyTyping(false);
+
+		_historyIndex++;
+		var currentAction = _history[_historyIndex];
+
+		_clearHistoryOnTextChanged = false;
+		switch (currentAction.Action)
+		{
+			case ReplaceAction r:
+				_pendingSelection = (r.caretIndexAfterReplacement, 0); // we always have an empty selection here.
+				ProcessTextInput(r.NewText);
+				break;
+			case DeleteAction d:
+				_pendingSelection = (Math.Min(d.UndoSelectionStart, d.UndoSelectionStart + d.UndoSelectionLength), 0);
+				ProcessTextInput(d.NewText);
+				break;
+			case SentinelAction:
+				break;
+			default:
+				global::System.Diagnostics.CI.Assert(false, "TextBoxActions are not exhaustively switch-matched.");
+				break;
+		}
+		_clearHistoryOnTextChanged = true;
+	}
+
+	internal bool IsDelegatingFocusToTemplateChild()
+		=> OperatingSystem.IsBrowser();
+
+	internal enum CaretDisplayMode
+	{
+		ThumblessCaretHidden,
+		ThumblessCaretShowing,
+		CaretWithThumbsOnlyEndShowing,
+		CaretWithThumbsBothEndsShowing
+	}
+
+	// Which platform's native touch text-selection conventions the Skia TextBox follows.
+	// Defaults to the running OS; runtime tests override it to exercise the mobile behavior on
+	// Skia Desktop, where OperatingSystem.IsAndroid()/IsIOS() are both false.
+	internal enum TouchTextSelectionConvention
+	{
+		// Desktop convention (Windows, macOS and Linux): the OS-derived default for every non-mobile target.
+		Desktop,
+		Android,
+		iOS
+	}
+
+	private record struct HistoryRecord(TextBoxAction Action, int SelectionStart, int SelectionLength, bool SelectionEndsAtTheStart);
+
+	private abstract record TextBoxAction;
+
+	/// <summary>
+	/// Instead of remembered what was removed and what was added in place, we just remember the initial and final states
+	/// as well as where the caret will be if we Redo. This is used by typing, paste, etc.
+	/// </summary>
+	private record ReplaceAction(string OldText, string NewText, int caretIndexAfterReplacement) : TextBoxAction;
+
+	/// <summary>
+	/// Unlike other forms of text modification, Delete doesn't follow the simple undo sequence of *unapply modification* -> *select what was selected when the action happened*
+	/// So we need to specifically need to remember what selection to go to when we Undo depending on how we got here (e.g. ctrl vs no ctrl)
+	/// Selection uses the possibly-negative format
+	/// </summary>
+	private record DeleteAction(string OldText, string NewText, int UndoSelectionStart, int UndoSelectionLength) : TextBoxAction;
+
+	/// <summary>
+	/// Probably unnecessary, but we pad the bottom of the history as it makes index manipulation easier (the invariant we
+	/// get is that history is never empty)
+	/// </summary>
+	private record SentinelAction : TextBoxAction
+	{
+		private SentinelAction() { }
+		public static SentinelAction Instance { get; } = new SentinelAction();
+	}
+
+	private sealed class TextBoxCommand(Action action) : ICommand
+	{
+		public bool CanExecute(object parameter) => true;
+
+		public void Execute(object parameter) => action();
+
+#pragma warning disable 67 // An event was declared but never used in the class in which it was declared.
+		public event EventHandler CanExecuteChanged;
+#pragma warning restore 67 // An event was declared but never used in the class in which it was declared.
+	}
+
+	#region ITextSelectionGripperHost
+	// The touch-selection grippers are driven by the shared TextSelectionGripperPresenter. The presenter
+	// owns the visuals/drag mechanics; these members expose just enough of TextBox's selection model and
+	// text surface for it to do its job.
+
+	TextBlock ITextSelectionGripperHost.GripperTextSurface => TextBoxView.DisplayBlock;
+
+	Rect ITextSelectionGripperHost.GripperClipBounds => Owner.GetAbsoluteBoundsRect();
+
+	GripperMode ITextSelectionGripperHost.GripperMode => _caretMode switch
+	{
+		CaretDisplayMode.CaretWithThumbsOnlyEndShowing => GripperMode.EndOnly,
+		CaretDisplayMode.CaretWithThumbsBothEndsShowing => GripperMode.Both,
+		_ => GripperMode.Hidden,
+	};
+
+	int ITextSelectionGripperHost.SelectionLowerIndex => SelectionStart;
+
+	int ITextSelectionGripperHost.SelectionUpperIndex => SelectionStart + SelectionLength;
+
+	void ITextSelectionGripperHost.SetGripperSelection(int start, int end) => SelectInternal(start, end - start);
+
+	void ITextSelectionGripperHost.MoveGripperCaret(int index) => Select(index, 0);
+
+	void ITextSelectionGripperHost.ScrollForGripper(bool isEndGripper) => UpdateScrolling(isEndGripper);
+
+	void ITextSelectionGripperHost.OnGripperPressed() => DismissSelectionFlyoutForPointerPress();
+
+	// Dismiss the selection flyout (the floating copy/paste bar) when a touch interaction begins, so it
+	// doesn't linger next to a context menu opened on the same gesture.
+	internal void DismissSelectionFlyoutForPointerPress() => TextControlFlyoutHelper.CloseIfOpen(SelectionFlyout);
+
+	void ITextSelectionGripperHost.RequestGripperContextMenu(PointerRoutedEventArgs args)
+	{
+		// Ported from TextBoxBase.cpp OnGripperHeld: a held gripper opens the context flyout.
+		var contextArgs = new ContextRequestedEventArgs();
+		contextArgs.SetGlobalPoint(args.GetCurrentPoint(null).Position);
+		UIElement.OnContextRequested(_host.Owner, contextArgs);
+	}
+
+	void ITextSelectionGripperHost.QueueGripperSelectionFlyout(PointerRoutedEventArgs args, bool allowEmptySelection)
+		=> QueueUpdateSelectionFlyoutVisibility(
+			args.Pointer.PointerDeviceType,
+			args.GetCurrentPoint(Owner).Position,
+			// The single insertion handle only exists under the mobile (iOS/Android) conventions; tapping it
+			// re-opens the flyout even over a collapsed caret. Desktop never gets the empty-selection flyout.
+			allowEmptySelection: allowEmptySelection && TouchSelectionConvention != TouchTextSelectionConvention.Desktop);
+
+	void ITextSelectionGripperHost.OnGripperTapped(PointerPoint press, int anchorIndex)
+	{
+		// The insertion handle (EndOnly gripper) sits over the character it points at, so the second tap of
+		// a double-tap lands on the handle instead of the text. Fold it into the same multi-tap counter as
+		// OnPointerReleasedPartial (press-to-press) so the double-tap still selects the word.
+		var repeatedPresses = _lastPointerDown.point is { } previous && IsTouchMultiTap(previous, press)
+			? _lastPointerDown.repeatedPresses + 1
+			: 0;
+		_lastPointerDown = (press, repeatedPresses);
+
+		// Act on the character the handle points at, not the finger's position on the thumb (which hangs below
+		// the caret line and would jump the caret to the end of the text).
+		if (TouchSelectionConvention != TouchTextSelectionConvention.Desktop && repeatedPresses >= 1)
+		{
+			TouchSelectWordAt(anchorIndex);
+		}
+		else if (_caretMode != CaretDisplayMode.CaretWithThumbsBothEndsShowing)
+		{
+			TouchTapAt(anchorIndex);
+		}
+		// Both thumbs showing: the tap grabbed a selection edge, so keep the selection (native iOS/Android)
+		// instead of collapsing it to a caret. The presenter re-shows the selection flyout on release.
+	}
+	#endregion
+
+	/// <summary>
+	/// Fires the ContextMenuOpening event synchronously and returns whether it was handled.
+	/// </summary>
+	/// <remarks>
+	/// Ported from CTextBoxBase::FireContextMenuOpeningEventSynchronously (TextBoxBase.cpp:5304)
+	/// and TextControlHelper::OnContextMenuOpeningHandler (TextControlHelper.h:10).
+	///
+	/// WinUI does m_pView->TransformToRoot(point) then divides by rasterization scale
+	/// to convert physical pixels to DIPs. In Uno/Skia, TransformToVisual(null) already
+	/// yields DIP coordinates, so no rasterization scale division is needed.
+	/// </remarks>
+	internal bool FireContextMenuOpeningEventSynchronously(Point point)
+	{
+		// WinUI: m_pView->TransformToRoot + pointerPosition /= zoomScale
+		var rootPoint = TransformToVisual(null).TransformPoint(point);
+
+		// WinUI: TextControlHelper::OnContextMenuOpeningHandler
+		// — creates ContextMenuEventArgs, sets CursorLeft/CursorTop, Handled=false
+		// — raises event synchronously
+		// — reads back Handled
+		return _host.RaiseContextMenuOpening(rootPoint.X, rootPoint.Y);
+	}
+}
